Network.AWS.Redshift

Description

Amazon Redshift is a fast, fully managed, petabyte-scale data warehouse service that makes it simple and cost-effective to efficiently analyze all your data using your existing business intelligence tools. You can start small for just $0.25 per hour with no commitments or upfront costs and scale to a petabyte or more for $1,000 per terabyte per year, less than a tenth of most other data warehousing solutions.

Documentation

module Network.AWS.Redshift.AuthorizeClusterSecurityGroupIngress

module Network.AWS.Redshift.AuthorizeSnapshotAccess

module Network.AWS.Redshift.CopyClusterSnapshot

module Network.AWS.Redshift.CreateCluster

module Network.AWS.Redshift.CreateClusterParameterGroup

module Network.AWS.Redshift.CreateClusterSecurityGroup

module Network.AWS.Redshift.CreateClusterSnapshot

module Network.AWS.Redshift.CreateClusterSubnetGroup

module Network.AWS.Redshift.CreateEventSubscription

module Network.AWS.Redshift.CreateHsmClientCertificate

module Network.AWS.Redshift.CreateHsmConfiguration

module Network.AWS.Redshift.CreateTags

module Network.AWS.Redshift.DeleteCluster

module Network.AWS.Redshift.DeleteClusterParameterGroup

module Network.AWS.Redshift.DeleteClusterSecurityGroup

module Network.AWS.Redshift.DeleteClusterSnapshot

module Network.AWS.Redshift.DeleteClusterSubnetGroup

module Network.AWS.Redshift.DeleteEventSubscription

module Network.AWS.Redshift.DeleteHsmClientCertificate

module Network.AWS.Redshift.DeleteHsmConfiguration

module Network.AWS.Redshift.DeleteTags

module Network.AWS.Redshift.DescribeClusterParameterGroups

module Network.AWS.Redshift.DescribeClusterParameters

module Network.AWS.Redshift.DescribeClusterSecurityGroups

module Network.AWS.Redshift.DescribeClusterSnapshots

module Network.AWS.Redshift.DescribeClusterSubnetGroups

module Network.AWS.Redshift.DescribeClusterVersions

module Network.AWS.Redshift.DescribeClusters

module Network.AWS.Redshift.DescribeDefaultClusterParameters

module Network.AWS.Redshift.DescribeEventCategories

module Network.AWS.Redshift.DescribeEventSubscriptions

module Network.AWS.Redshift.DescribeEvents

module Network.AWS.Redshift.DescribeHsmClientCertificates

module Network.AWS.Redshift.DescribeHsmConfigurations

module Network.AWS.Redshift.DescribeLoggingStatus

module Network.AWS.Redshift.DescribeOrderableClusterOptions

module Network.AWS.Redshift.DescribeReservedNodeOfferings

module Network.AWS.Redshift.DescribeReservedNodes

module Network.AWS.Redshift.DescribeResize

module Network.AWS.Redshift.DescribeTags

module Network.AWS.Redshift.DisableLogging

module Network.AWS.Redshift.DisableSnapshotCopy

module Network.AWS.Redshift.EnableLogging

module Network.AWS.Redshift.EnableSnapshotCopy

module Network.AWS.Redshift.ModifyCluster

module Network.AWS.Redshift.ModifyClusterParameterGroup

module Network.AWS.Redshift.ModifyClusterSubnetGroup

module Network.AWS.Redshift.ModifyEventSubscription

module Network.AWS.Redshift.ModifySnapshotCopyRetentionPeriod

module Network.AWS.Redshift.PurchaseReservedNodeOffering

module Network.AWS.Redshift.RebootCluster

module Network.AWS.Redshift.ResetClusterParameterGroup

module Network.AWS.Redshift.RestoreFromClusterSnapshot

module Network.AWS.Redshift.RevokeClusterSecurityGroupIngress

module Network.AWS.Redshift.RevokeSnapshotAccess

module Network.AWS.Redshift.RotateEncryptionKey

module Network.AWS.Redshift.Types

module Network.AWS.Redshift.Waiters