Network.AWS.Redshift.AuthorizeClusterSecurityGroupIngress

Description

Adds an inbound (ingress) rule to an Amazon Redshift security group. Depending on whether the application accessing your cluster is running on the Internet or an EC2 instance, you can authorize inbound access to either a Classless Interdomain Routing (CIDR) IP address range or an EC2 security group. You can add as many as 20 ingress rules to an Amazon Redshift security group.

The EC2 security group must be defined in the AWS region where the cluster resides. For an overview of CIDR blocks, see the Wikipedia article on Classless Inter-Domain Routing.

You must also associate the security group with a cluster so that clients running on these IP addresses or the EC2 instance are authorized to connect to the cluster. For information about managing security groups, go to Workingwith Security Groups in the Amazon Redshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_AuthorizeClusterSecurityGroupIngress.html

Request

data AuthorizeClusterSecurityGroupIngress

Request constructor

authorizeClusterSecurityGroupIngress

Arguments

:: Text
acsgiClusterSecurityGroupName

-> AuthorizeClusterSecurityGroupIngress

AuthorizeClusterSecurityGroupIngress constructor.

The fields accessible through corresponding lenses are:

acsgiCIDRIP :: Maybe Text

acsgiClusterSecurityGroupName :: Text

acsgiEC2SecurityGroupName :: Maybe Text

acsgiEC2SecurityGroupOwnerId :: Maybe Text

Request lenses

acsgiCIDRIP :: Lens' AuthorizeClusterSecurityGroupIngress (Maybe Text)

The IP range to be added the Amazon Redshift security group.

acsgiClusterSecurityGroupName :: Lens' AuthorizeClusterSecurityGroupIngress Text

The name of the security group to which the ingress rule is added.

acsgiEC2SecurityGroupName :: Lens' AuthorizeClusterSecurityGroupIngress (Maybe Text)

The EC2 security group to be added the Amazon Redshift security group.

acsgiEC2SecurityGroupOwnerId :: Lens' AuthorizeClusterSecurityGroupIngress (Maybe Text)

The AWS account number of the owner of the security group specified by the EC2SecurityGroupName parameter. The AWS Access Key ID is not an acceptable value.

Example: '111122223333'

Response

data AuthorizeClusterSecurityGroupIngressResponse

Response constructor

authorizeClusterSecurityGroupIngressResponse :: AuthorizeClusterSecurityGroupIngressResponse

AuthorizeClusterSecurityGroupIngressResponse constructor.

The fields accessible through corresponding lenses are:

acsgirClusterSecurityGroup :: Maybe ClusterSecurityGroup

Response lenses

acsgirClusterSecurityGroup :: Lens' AuthorizeClusterSecurityGroupIngressResponse (Maybe ClusterSecurityGroup)