Network.AWS.Redshift.AuthorizeSnapshotAccess

Description

Authorizes the specified AWS customer account to restore the specified snapshot.

For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Cluster Management Guide .

Creating a Request

authorizeSnapshotAccess

Arguments

:: Text
asaSnapshotIdentifier

-> Text
asaAccountWithRestoreAccess

-> AuthorizeSnapshotAccess

Creates a value of AuthorizeSnapshotAccess with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asaSnapshotClusterIdentifier - The identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

asaSnapshotIdentifier - The identifier of the snapshot the account is authorized to restore.

asaAccountWithRestoreAccess - The identifier of the AWS customer account authorized to restore the specified snapshot. To share a snapshot with AWS support, specify amazon-redshift-support.

data AuthorizeSnapshotAccess

See: authorizeSnapshotAccess smart constructor.

Request Lenses

asaSnapshotClusterIdentifier :: Lens' AuthorizeSnapshotAccess (Maybe Text)

The identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

asaSnapshotIdentifier :: Lens' AuthorizeSnapshotAccess Text

The identifier of the snapshot the account is authorized to restore.

asaAccountWithRestoreAccess :: Lens' AuthorizeSnapshotAccess Text

The identifier of the AWS customer account authorized to restore the specified snapshot. To share a snapshot with AWS support, specify amazon-redshift-support.

Destructuring the Response

authorizeSnapshotAccessResponse

Arguments

:: Int
asarsResponseStatus

-> AuthorizeSnapshotAccessResponse

Creates a value of AuthorizeSnapshotAccessResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asarsSnapshot - Undocumented member.

asarsResponseStatus - -- | The response status code.

data AuthorizeSnapshotAccessResponse

See: authorizeSnapshotAccessResponse smart constructor.

Response Lenses

asarsSnapshot :: Lens' AuthorizeSnapshotAccessResponse (Maybe Snapshot)

Undocumented member.

asarsResponseStatus :: Lens' AuthorizeSnapshotAccessResponse Int

- | The response status code.