Network.AWS.Redshift.CreateClusterParameterGroup

Description

Creates an Amazon Redshift parameter group. Creating parameter groups is independent of creating clusters. You can associate a cluster with a parameter group when you create the cluster. You can also associate an existing cluster with a parameter group after the cluster is created by using ModifyCluster. Parameters in the parameter group define specific behavior that applies to the databases you create on the cluster. For more information about managing parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Management Guide. https:/redshift.us-east-1.amazonaws.com ?Action=CreateClusterParameterGroup &Description=description my parameter group &ParameterGroupFamily=redshift-1.0 &ParameterGroupName=parametergroup1 &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20130123us-east-1redshiftaws4_request &x-amz-date=20130123T002544Z &x-amz-signedheaders=content-type;host;x-amz-date redshift-1.0 description my parameter group parametergroup1 6d6df847-64f3-11e2-bea9-49e0ce183f07.

Request

data CreateClusterParameterGroup

Request constructor

createClusterParameterGroup

Arguments

:: Text
ccpgParameterGroupName

-> Text
ccpgParameterGroupFamily

-> Text
ccpgDescription

-> CreateClusterParameterGroup

Smart constructor for the minimum required parameters to construct a valid CreateClusterParameterGroup request.

The fields accessible through corresponding lenses are:

ParameterGroupName :: Text

ParameterGroupFamily :: Text

Description :: Text

Request lenses

ccpgParameterGroupName :: Lens' CreateClusterParameterGroup Text

The name of the cluster parameter group. Constraints: Must be 1 to 255 alphanumeric characters or hyphens First character must be a letter. Cannot end with a hyphen or contain two consecutive hyphens. Must be unique withing your AWS account. This value is stored as a lower-case string.

ccpgParameterGroupFamily :: Lens' CreateClusterParameterGroup Text

The Amazon Redshift engine version to which the cluster parameter group applies. The cluster engine version determines the set of parameters. To get a list of valid parameter group family names, you can call DescribeClusterParameterGroups. By default, Amazon Redshift returns a list of all the parameter groups that are owned by your AWS account, including the default parameter groups for each Amazon Redshift engine version. The parameter group family names associated with the default parameter groups provide you the valid values. For example, a valid family name is "redshift-1.0".

ccpgDescription :: Lens' CreateClusterParameterGroup Text

A description of the parameter group.

Response

data CreateClusterParameterGroupResponse

Response constructor

createClusterParameterGroupResponse :: CreateClusterParameterGroupResponse

Smart constructor for the minimum required parameters to construct a valid CreateClusterParameterGroupResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ClusterParameterGroup :: Maybe ClusterParameterGroup

Response lenses

ccpgrClusterParameterGroup :: Lens' CreateClusterParameterGroupResponse (Maybe ClusterParameterGroup)

Describes a parameter group.