Network.AWS.Redshift.CreateClusterSecurityGroup

Description

Creates a new Amazon Redshift security group. You use security groups to control access to non-VPC clusters.

For information about managing security groups, go to Amazon RedshiftCluster Security Groups in the Amazon Redshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_CreateClusterSecurityGroup.html

Request

data CreateClusterSecurityGroup

Request constructor

createClusterSecurityGroup

Arguments

:: Text
ccsg1ClusterSecurityGroupName

-> Text
ccsg1Description

-> CreateClusterSecurityGroup

CreateClusterSecurityGroup constructor.

The fields accessible through corresponding lenses are:

ccsg1ClusterSecurityGroupName :: Text

ccsg1Description :: Text

ccsg1Tags :: [Tag]

Request lenses

ccsg1ClusterSecurityGroupName :: Lens' CreateClusterSecurityGroup Text

The name for the security group. Amazon Redshift stores the value as a lowercase string.

Constraints:

Must contain no more than 255 alphanumeric characters or hyphens. Must not be Default. Must be unique for all security groups that are created by your AWS account. Example: examplesecuritygroup

ccsg1Description :: Lens' CreateClusterSecurityGroup Text

A description for the security group.

ccsg1Tags :: Lens' CreateClusterSecurityGroup [Tag]

A list of tag instances.

Response

data CreateClusterSecurityGroupResponse

Response constructor

createClusterSecurityGroupResponse :: CreateClusterSecurityGroupResponse

CreateClusterSecurityGroupResponse constructor.

The fields accessible through corresponding lenses are:

ccsgrClusterSecurityGroup :: Maybe ClusterSecurityGroup

Response lenses

ccsgrClusterSecurityGroup :: Lens' CreateClusterSecurityGroupResponse (Maybe ClusterSecurityGroup)