Network.AWS.Redshift.CreateClusterSnapshot

Description

Creates a manual snapshot of the specified cluster. The cluster must be in the available state.

For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Cluster Management Guide .

Creating a Request

createClusterSnapshot

Arguments

:: Text
ccsSnapshotIdentifier

-> Text
ccsClusterIdentifier

-> CreateClusterSnapshot

Creates a value of CreateClusterSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccsTags - A list of tag instances.

ccsSnapshotIdentifier - A unique identifier for the snapshot that you are requesting. This identifier must be unique for all snapshots within the AWS account. Constraints: * Cannot be null, empty, or blank * Must contain from 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-snapshot-id

ccsClusterIdentifier - The cluster identifier for which you want a snapshot.

data CreateClusterSnapshot

See: createClusterSnapshot smart constructor.

Request Lenses

ccsTags :: Lens' CreateClusterSnapshot [Tag]

A list of tag instances.

ccsSnapshotIdentifier :: Lens' CreateClusterSnapshot Text

A unique identifier for the snapshot that you are requesting. This identifier must be unique for all snapshots within the AWS account. Constraints: * Cannot be null, empty, or blank * Must contain from 1 to 255 alphanumeric characters or hyphens * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens Example: my-snapshot-id

ccsClusterIdentifier :: Lens' CreateClusterSnapshot Text

The cluster identifier for which you want a snapshot.

Destructuring the Response

createClusterSnapshotResponse

Arguments

:: Int
crersResponseStatus

-> CreateClusterSnapshotResponse

Creates a value of CreateClusterSnapshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crersSnapshot - Undocumented member.

crersResponseStatus - -- | The response status code.

data CreateClusterSnapshotResponse

See: createClusterSnapshotResponse smart constructor.

Response Lenses

crersSnapshot :: Lens' CreateClusterSnapshotResponse (Maybe Snapshot)

Undocumented member.

crersResponseStatus :: Lens' CreateClusterSnapshotResponse Int

- | The response status code.