Network.AWS.Redshift.CreateClusterSubnetGroup

Description

Creates a new Amazon Redshift subnet group. You must provide a list of one or more subnets in your existing Amazon Virtual Private Cloud (Amazon VPC) when creating Amazon Redshift subnet group.

For information about subnet groups, go to Amazon Redshift Cluster SubnetGroups in the Amazon Redshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_CreateClusterSubnetGroup.html

Request

data CreateClusterSubnetGroup

Request constructor

createClusterSubnetGroup

Arguments

:: Text
ccsgClusterSubnetGroupName

-> Text
ccsgDescription

-> CreateClusterSubnetGroup

CreateClusterSubnetGroup constructor.

The fields accessible through corresponding lenses are:

ccsgClusterSubnetGroupName :: Text

ccsgDescription :: Text

ccsgSubnetIds :: [Text]

ccsgTags :: [Tag]

Request lenses

ccsgClusterSubnetGroupName :: Lens' CreateClusterSubnetGroup Text

The name for the subnet group. Amazon Redshift stores the value as a lowercase string.

Constraints:

Must contain no more than 255 alphanumeric characters or hyphens. Must not be Default. Must be unique for all subnet groups that are created by your AWS account. Example: examplesubnetgroup

ccsgDescription :: Lens' CreateClusterSubnetGroup Text

A description for the subnet group.

ccsgSubnetIds :: Lens' CreateClusterSubnetGroup [Text]

An array of VPC subnet IDs. A maximum of 20 subnets can be modified in a single request.

ccsgTags :: Lens' CreateClusterSubnetGroup [Tag]

A list of tag instances.

Response

data CreateClusterSubnetGroupResponse

Response constructor

createClusterSubnetGroupResponse :: CreateClusterSubnetGroupResponse

CreateClusterSubnetGroupResponse constructor.

The fields accessible through corresponding lenses are:

ccsgrClusterSubnetGroup :: Maybe ClusterSubnetGroup

Response lenses

ccsgrClusterSubnetGroup :: Lens' CreateClusterSubnetGroupResponse (Maybe ClusterSubnetGroup)