Network.AWS.Redshift.CreateHSMConfiguration

Description

Creates an HSM configuration that contains the information required by an Amazon Redshift cluster to store and use database encryption keys in a Hardware Security Module (HSM). After creating the HSM configuration, you can specify it as a parameter when creating a cluster. The cluster will then store its encryption keys in the HSM.

In addition to creating an HSM configuration, you must also create an HSM client certificate. For more information, go to Hardware Security Modules in the Amazon Redshift Cluster Management Guide.

Creating a Request

createHSMConfiguration

Arguments

:: Text
chcHSMConfigurationIdentifier

-> Text
chcDescription

-> Text
chcHSMIPAddress

-> Text
chcHSMPartitionName

-> Text
chcHSMPartitionPassword

-> Text
chcHSMServerPublicCertificate

-> CreateHSMConfiguration

Creates a value of CreateHSMConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chcTags - A list of tag instances.

chcHSMConfigurationIdentifier - The identifier to be assigned to the new Amazon Redshift HSM configuration.

chcDescription - A text description of the HSM configuration to be created.

chcHSMIPAddress - The IP address that the Amazon Redshift cluster must use to access the HSM.

chcHSMPartitionName - The name of the partition in the HSM where the Amazon Redshift clusters will store their database encryption keys.

chcHSMPartitionPassword - The password required to access the HSM partition.

chcHSMServerPublicCertificate - The HSMs public certificate file. When using Cloud HSM, the file name is server.pem.

data CreateHSMConfiguration

See: createHSMConfiguration smart constructor.

Request Lenses

chcTags :: Lens' CreateHSMConfiguration [Tag]

A list of tag instances.

chcHSMConfigurationIdentifier :: Lens' CreateHSMConfiguration Text

The identifier to be assigned to the new Amazon Redshift HSM configuration.

chcDescription :: Lens' CreateHSMConfiguration Text

A text description of the HSM configuration to be created.

chcHSMIPAddress :: Lens' CreateHSMConfiguration Text

The IP address that the Amazon Redshift cluster must use to access the HSM.

chcHSMPartitionName :: Lens' CreateHSMConfiguration Text

The name of the partition in the HSM where the Amazon Redshift clusters will store their database encryption keys.

chcHSMPartitionPassword :: Lens' CreateHSMConfiguration Text

The password required to access the HSM partition.

chcHSMServerPublicCertificate :: Lens' CreateHSMConfiguration Text

The HSMs public certificate file. When using Cloud HSM, the file name is server.pem.

Destructuring the Response

createHSMConfigurationResponse

Arguments

:: Int
chcrsResponseStatus

-> CreateHSMConfigurationResponse

Creates a value of CreateHSMConfigurationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chcrsHSMConfiguration - Undocumented member.

chcrsResponseStatus - -- | The response status code.

data CreateHSMConfigurationResponse

See: createHSMConfigurationResponse smart constructor.

Response Lenses

chcrsHSMConfiguration :: Lens' CreateHSMConfigurationResponse (Maybe HSMConfiguration)

Undocumented member.

chcrsResponseStatus :: Lens' CreateHSMConfigurationResponse Int

- | The response status code.