Network.AWS.Redshift.CreateHsmClientCertificate

Description

Creates an HSM client certificate that an Amazon Redshift cluster will use to connect to the client's HSM in order to store and retrieve the keys used to encrypt the cluster databases.

The command returns a public key, which you must store in the HSM. In addition to creating the HSM certificate, you must create an Amazon Redshift HSM configuration that provides a cluster the information needed to store and use encryption keys in the HSM. For more information, go to Hardware SecurityModules in the Amazon Redshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_CreateHsmClientCertificate.html

Request

data CreateHsmClientCertificate

Request constructor

createHsmClientCertificate

Arguments

:: Text
chccHsmClientCertificateIdentifier

-> CreateHsmClientCertificate

CreateHsmClientCertificate constructor.

The fields accessible through corresponding lenses are:

chccHsmClientCertificateIdentifier :: Text

chccTags :: [Tag]

Request lenses

chccHsmClientCertificateIdentifier :: Lens' CreateHsmClientCertificate Text

The identifier to be assigned to the new HSM client certificate that the cluster will use to connect to the HSM to use the database encryption keys.

chccTags :: Lens' CreateHsmClientCertificate [Tag]

A list of tag instances.

Response

data CreateHsmClientCertificateResponse

Response constructor

createHsmClientCertificateResponse :: CreateHsmClientCertificateResponse

CreateHsmClientCertificateResponse constructor.

The fields accessible through corresponding lenses are:

chccrHsmClientCertificate :: Maybe HsmClientCertificate

Response lenses

chccrHsmClientCertificate :: Lens' CreateHsmClientCertificateResponse (Maybe HsmClientCertificate)