Network.AWS.Redshift.CreateHsmConfiguration

Description

Creates an HSM configuration that contains the information required by an Amazon Redshift cluster to store and use database encryption keys in a Hardware Security Module (HSM). After creating the HSM configuration, you can specify it as a parameter when creating a cluster. The cluster will then store its encryption keys in the HSM. In addition to creating an HSM configuration, you must also create an HSM client certificate. For more information, go to Hardware Security Modules in the Amazon Redshift Management Guide.

Request

data CreateHsmConfiguration

Request constructor

mkCreateHsmConfiguration

Arguments

:: Text
chcHsmConfigurationIdentifier

-> Text
chcDescription

-> Text
chcHsmIpAddress

-> Text
chcHsmPartitionName

-> Text
chcHsmPartitionPassword

-> Text
chcHsmServerPublicCertificate

-> CreateHsmConfiguration

Smart constructor for the minimum required parameters to construct a valid CreateHsmConfiguration request.

The fields accessible through corresponding lenses are:

HsmConfigurationIdentifier :: Text

Description :: Text

HsmIpAddress :: Text

HsmPartitionName :: Text

HsmPartitionPassword :: Text

HsmServerPublicCertificate :: Text

Request lenses

chcHsmConfigurationIdentifier :: Lens' CreateHsmConfiguration Text

The identifier to be assigned to the new Amazon Redshift HSM configuration.

chcDescription :: Lens' CreateHsmConfiguration Text

A text description of the HSM configuration to be created.

chcHsmIpAddress :: Lens' CreateHsmConfiguration Text

The IP address that the Amazon Redshift cluster must use to access the HSM.

chcHsmPartitionName :: Lens' CreateHsmConfiguration Text

The name of the partition in the HSM where the Amazon Redshift clusters will store their database encryption keys.

chcHsmPartitionPassword :: Lens' CreateHsmConfiguration Text

The password required to access the HSM partition.

chcHsmServerPublicCertificate :: Lens' CreateHsmConfiguration Text

The HSMs public certificate file. When using Cloud HSM, the file name is server.pem.

Response

data CreateHsmConfigurationResponse

Response constructor

mkCreateHsmConfigurationResponse :: CreateHsmConfigurationResponse

Smart constructor for the minimum required parameters to construct a valid CreateHsmConfigurationResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

HsmConfiguration :: Maybe HsmConfiguration

Response lenses

chcrHsmConfiguration :: Lens' CreateHsmConfigurationResponse (Maybe HsmConfiguration)

Returns information about an HSM configuration, which is an object that describes to Amazon Redshift clusters the information they require to connect to an HSM where they can store database encryption keys.