Network.AWS.Redshift.CreateSnapshotCopyGrant

Description

Creates a snapshot copy grant that permits Amazon Redshift to use a customer master key (CMK) from AWS Key Management Service (AWS KMS) to encrypt copied snapshots in a destination region.

For more information about managing snapshot copy grants, go to Amazon Redshift Database Encryption in the Amazon Redshift Cluster Management Guide .

Creating a Request

createSnapshotCopyGrant

Arguments

:: Text
cscgSnapshotCopyGrantName

-> CreateSnapshotCopyGrant

Creates a value of CreateSnapshotCopyGrant with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cscgKMSKeyId - The unique identifier of the customer master key (CMK) to which to grant Amazon Redshift permission. If no key is specified, the default key is used.

cscgTags - A list of tag instances.

cscgSnapshotCopyGrantName - The name of the snapshot copy grant. This name must be unique in the region for the AWS account. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * Alphabetic characters must be lowercase. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. * Must be unique for all clusters within an AWS account.

data CreateSnapshotCopyGrant

The result of the CreateSnapshotCopyGrant action.

See: createSnapshotCopyGrant smart constructor.

Request Lenses

cscgKMSKeyId :: Lens' CreateSnapshotCopyGrant (Maybe Text)

The unique identifier of the customer master key (CMK) to which to grant Amazon Redshift permission. If no key is specified, the default key is used.

cscgTags :: Lens' CreateSnapshotCopyGrant [Tag]

A list of tag instances.

cscgSnapshotCopyGrantName :: Lens' CreateSnapshotCopyGrant Text

The name of the snapshot copy grant. This name must be unique in the region for the AWS account. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * Alphabetic characters must be lowercase. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. * Must be unique for all clusters within an AWS account.

Destructuring the Response

createSnapshotCopyGrantResponse

Arguments

:: Int
cscgrsResponseStatus

-> CreateSnapshotCopyGrantResponse

Creates a value of CreateSnapshotCopyGrantResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cscgrsSnapshotCopyGrant - Undocumented member.

cscgrsResponseStatus - -- | The response status code.

data CreateSnapshotCopyGrantResponse

See: createSnapshotCopyGrantResponse smart constructor.

Response Lenses

cscgrsSnapshotCopyGrant :: Lens' CreateSnapshotCopyGrantResponse (Maybe SnapshotCopyGrant)

Undocumented member.

cscgrsResponseStatus :: Lens' CreateSnapshotCopyGrantResponse Int

- | The response status code.