Network.AWS.Redshift.CreateTags

Description

Adds one or more tags to a specified resource.

A resource can have up to 10 tags. If you try to create more than 10 tags for a resource, you will receive an error and the attempt will fail.

If you specify a key that already exists for the resource, the value for that key will be updated with the new value.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_CreateTags.html

Request

data CreateTags

Request constructor

createTags

Arguments

:: Text
ctResourceName

-> CreateTags

CreateTags constructor.

The fields accessible through corresponding lenses are:

ctResourceName :: Text

ctTags :: [Tag]

Request lenses

ctResourceName :: Lens' CreateTags Text

The Amazon Resource Name (ARN) to which you want to add the tag or tags. For example, 'arn:aws:redshift:us-east-1:123456789:cluster:t1'.

ctTags :: Lens' CreateTags [Tag]

One or more name/value pairs to add as tags to the specified resource. Each tag name is passed in with the parameter 'tag-key' and the corresponding value is passed in with the parameter 'tag-value'. The 'tag-key' and 'tag-value' parameters are separated by a colon (:). Separate multiple tags with a space. For example, '--tags "tag-key"="owner":"tag-value"="admin""tag-key"="environment":"tag-value"="test""tag-key"="version":"tag-value"="1.0"'.

Response

data CreateTagsResponse

Response constructor

createTagsResponse :: CreateTagsResponse

CreateTagsResponse constructor.