Network.AWS.Redshift.DeleteCluster

Description

Deletes a previously provisioned cluster. A successful response from the web service indicates that the request was received correctly. Use DescribeClusters to monitor the status of the deletion. The delete operation cannot be canceled or reverted once submitted. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift ClusterManagement Guide .

If you want to shut down the cluster and retain it for future use, set SkipFinalClusterSnapshot to false and specify a name for FinalClusterSnapshotIdentifier. You can later restore this snapshot to resume using the cluster. If a final cluster snapshot is requested, the status of the cluster will be "final-snapshot" while the snapshot is being taken, then it's "deleting" once Amazon Redshift begins deleting the cluster.

For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Cluster Management Guide .

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DeleteCluster.html

Request

data DeleteCluster

Request constructor

deleteCluster

Arguments

:: Text
dc1ClusterIdentifier

-> DeleteCluster

DeleteCluster constructor.

The fields accessible through corresponding lenses are:

dc1ClusterIdentifier :: Text

dc1FinalClusterSnapshotIdentifier :: Maybe Text

dc1SkipFinalClusterSnapshot :: Maybe Bool

Request lenses

dc1ClusterIdentifier :: Lens' DeleteCluster Text

The identifier of the cluster to be deleted.

Constraints:

Must contain lowercase characters. Must contain from 1 to 63 alphanumeric characters or hyphens. First character must be a letter. Cannot end with a hyphen or contain two consecutive hyphens.

dc1FinalClusterSnapshotIdentifier :: Lens' DeleteCluster (Maybe Text)

The identifier of the final snapshot that is to be created immediately before deleting the cluster. If this parameter is provided, SkipFinalClusterSnapshot must be false.

Constraints:

Must be 1 to 255 alphanumeric characters. First character must be a letter. Cannot end with a hyphen or contain two consecutive hyphens.

dc1SkipFinalClusterSnapshot :: Lens' DeleteCluster (Maybe Bool)

Determines whether a final snapshot of the cluster is created before Amazon Redshift deletes the cluster. If true, a final cluster snapshot is not created. If false, a final cluster snapshot is created before the cluster is deleted.

The FinalClusterSnapshotIdentifier parameter must be specified if SkipFinalClusterSnapshot is false. Default: false

Response

data DeleteClusterResponse

Response constructor

deleteClusterResponse :: DeleteClusterResponse

DeleteClusterResponse constructor.

The fields accessible through corresponding lenses are:

dcrCluster :: Maybe Cluster

Response lenses

dcrCluster :: Lens' DeleteClusterResponse (Maybe Cluster)