Network.AWS.Redshift.DeleteClusterSecurityGroup

Description

Deletes an Amazon Redshift security group. You cannot delete a security group that is associated with any clusters. You cannot delete the default security group. For information about managing security groups, go to Amazon Redshift Cluster Security Groups in the Amazon Redshift Management Guide. https:/redshift.us-east-1.amazonaws.com ?Action=DeleteClusterSecurityGroup &ClusterSecurityGroupName=securitygroup1 &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20121208us-east-1redshiftaws4_request &x-amz-date=20121208T015926Z &x-amz-signedheaders=content-type;host;x-amz-date e54e05dc-40da-11e2-955f-313c36e9e01d.

Request

data DeleteClusterSecurityGroup

Request constructor

deleteClusterSecurityGroup

Arguments

:: Text
dcsgClusterSecurityGroupName

-> DeleteClusterSecurityGroup

Smart constructor for the minimum required parameters to construct a valid DeleteClusterSecurityGroup request.

The fields accessible through corresponding lenses are:

ClusterSecurityGroupName :: Text

Request lenses

dcsgClusterSecurityGroupName :: Lens' DeleteClusterSecurityGroup Text

The name of the cluster security group to be deleted.

Response

data DeleteClusterSecurityGroupResponse

Response constructor

deleteClusterSecurityGroupResponse :: DeleteClusterSecurityGroupResponse

Smart constructor for the minimum required parameters to construct a valid DeleteClusterSecurityGroupResponse response.

This constructor is provided for convenience and testing purposes.