Network.AWS.Redshift.DeleteClusterSnapshot

Description

Deletes the specified manual snapshot. The snapshot must be in the available state, with no other users authorized to access the snapshot.

Unlike automated snapshots, manual snapshots are retained even after you delete your cluster. Amazon Redshift does not delete your manual snapshots. You must delete manual snapshot explicitly to avoid getting charged. If other accounts are authorized to access the snapshot, you must revoke all of the authorizations before you can delete the snapshot.

Creating a Request

deleteClusterSnapshot

Arguments

:: Text
dcsSnapshotIdentifier

-> DeleteClusterSnapshot

Creates a value of DeleteClusterSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcsSnapshotClusterIdentifier - The unique identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name. Constraints: Must be the name of valid cluster.

dcsSnapshotIdentifier - The unique identifier of the manual snapshot to be deleted. Constraints: Must be the name of an existing snapshot that is in the available state.

data DeleteClusterSnapshot

See: deleteClusterSnapshot smart constructor.

Request Lenses

dcsSnapshotClusterIdentifier :: Lens' DeleteClusterSnapshot (Maybe Text)

The unique identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name. Constraints: Must be the name of valid cluster.

dcsSnapshotIdentifier :: Lens' DeleteClusterSnapshot Text

The unique identifier of the manual snapshot to be deleted. Constraints: Must be the name of an existing snapshot that is in the available state.

Destructuring the Response

deleteClusterSnapshotResponse

Arguments

:: Int
dcsrsResponseStatus

-> DeleteClusterSnapshotResponse

Creates a value of DeleteClusterSnapshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcsrsSnapshot - Undocumented member.

dcsrsResponseStatus - -- | The response status code.

data DeleteClusterSnapshotResponse

See: deleteClusterSnapshotResponse smart constructor.

Response Lenses

dcsrsSnapshot :: Lens' DeleteClusterSnapshotResponse (Maybe Snapshot)

Undocumented member.

dcsrsResponseStatus :: Lens' DeleteClusterSnapshotResponse Int

- | The response status code.