Network.AWS.Redshift.DescribeClusterParameterGroups

Description

Returns a list of Amazon Redshift parameter groups, including parameter groups you created and the default parameter group. For each parameter group, the response includes the parameter group name, description, and parameter group family name. You can optionally specify a name to retrieve the description of a specific parameter group. For more information about managing parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeClusterParameterGroups.html

Request

data DescribeClusterParameterGroups

Request constructor

describeClusterParameterGroups :: DescribeClusterParameterGroups

DescribeClusterParameterGroups constructor.

The fields accessible through corresponding lenses are:

dcpgMarker :: Maybe Text

dcpgMaxRecords :: Maybe Int

dcpgParameterGroupName :: Maybe Text

Request lenses

dcpgMarker :: Lens' DescribeClusterParameterGroups (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterParameterGroups request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dcpgMaxRecords :: Lens' DescribeClusterParameterGroups (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

dcpgParameterGroupName :: Lens' DescribeClusterParameterGroups (Maybe Text)

The name of a specific parameter group for which to return details. By default, details about all parameter groups and the default parameter group are returned.

Response

data DescribeClusterParameterGroupsResponse

Response constructor

describeClusterParameterGroupsResponse :: DescribeClusterParameterGroupsResponse

DescribeClusterParameterGroupsResponse constructor.

The fields accessible through corresponding lenses are:

dcpgrMarker :: Maybe Text

dcpgrParameterGroups :: [ClusterParameterGroup]

Response lenses

dcpgrMarker :: Lens' DescribeClusterParameterGroupsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcpgrParameterGroups :: Lens' DescribeClusterParameterGroupsResponse [ClusterParameterGroup]

A list of ClusterParameterGroup instances. Each instance describes one cluster parameter group.