Network.AWS.Redshift.DescribeClusterParameters

Description

Returns a detailed list of parameters contained within the specified Amazon Redshift parameter group. For each parameter the response includes information such as parameter name, description, data type, value, whether the parameter value is modifiable, and so on.

You can specify source filter to retrieve parameters of only specific type. For example, to retrieve parameters that were modified by a user action such as from ModifyClusterParameterGroup, you can specify source equal to user.

For more information about managing parameter groups, go to Amazon RedshiftParameter Groups in the Amazon Redshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeClusterParameters.html

Request

data DescribeClusterParameters

Request constructor

describeClusterParameters

Arguments

:: Text
dcp1ParameterGroupName

-> DescribeClusterParameters

DescribeClusterParameters constructor.

The fields accessible through corresponding lenses are:

dcp1Marker :: Maybe Text

dcp1MaxRecords :: Maybe Int

dcp1ParameterGroupName :: Text

dcp1Source :: Maybe Text

Request lenses

dcp1Marker :: Lens' DescribeClusterParameters (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterParameters request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dcp1MaxRecords :: Lens' DescribeClusterParameters (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: '100'

Constraints: minimum 20, maximum 100.

dcp1ParameterGroupName :: Lens' DescribeClusterParameters Text

The name of a cluster parameter group for which to return details.

dcp1Source :: Lens' DescribeClusterParameters (Maybe Text)

The parameter types to return. Specify user to show parameters that are different form the default. Similarly, specify 'engine-default' to show parameters that are the same as the default parameter group.

Default: All parameter types returned.

Valid Values: user | 'engine-default'

Response

data DescribeClusterParametersResponse

Response constructor

describeClusterParametersResponse :: DescribeClusterParametersResponse

DescribeClusterParametersResponse constructor.

The fields accessible through corresponding lenses are:

dcprMarker :: Maybe Text

dcprParameters :: [Parameter]

Response lenses

dcprMarker :: Lens' DescribeClusterParametersResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcprParameters :: Lens' DescribeClusterParametersResponse [Parameter]

A list of Parameter instances. Each instance lists the parameters of one cluster parameter group.