Network.AWS.Redshift.DescribeClusterSubnetGroups

Description

Returns one or more cluster subnet group objects, which contain metadata about your cluster subnet groups. By default, this operation returns information about all cluster subnet groups that are defined in you AWS account. https:/redshift.us-east-1.amazonaws.com ?Action=DescribeClusterSubnetGroups &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20130130us-east-1redshiftaws4_request &x-amz-date=20130130T153938Z &x-amz-signedheaders=content-type;host;x-amz-date vpc-5d917a30 my subnet group my-subnet-group Complete Active subnet-71c5091c us-east-1a Active subnet-78de1215 us-east-1a 42024b68-6af3-11e2-a726-6368a468fa67.

Request

data DescribeClusterSubnetGroups

Request constructor

describeClusterSubnetGroups :: DescribeClusterSubnetGroups

Smart constructor for the minimum required parameters to construct a valid DescribeClusterSubnetGroups request.

The fields accessible through corresponding lenses are:

ClusterSubnetGroupName :: Maybe Text

MaxRecords :: Maybe Integer

Marker :: Maybe Text

Request lenses

dcsg3ClusterSubnetGroupName :: Lens' DescribeClusterSubnetGroups (Maybe Text)

The name of the cluster subnet group for which information is requested.

dcsg3MaxRecords :: Lens' DescribeClusterSubnetGroups (Maybe Integer)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

dcsg3Marker :: Lens' DescribeClusterSubnetGroups (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterSubnetGroups request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

Response

data DescribeClusterSubnetGroupsResponse

Contains the output from the DescribeClusterSubnetGroups action.

Response constructor

describeClusterSubnetGroupsResponse :: DescribeClusterSubnetGroupsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeClusterSubnetGroupsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

ClusterSubnetGroups :: [ClusterSubnetGroup]

Response lenses

dcsgrrMarker :: Lens' DescribeClusterSubnetGroupsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcsgrrClusterSubnetGroups :: Lens' DescribeClusterSubnetGroupsResponse [ClusterSubnetGroup]

A list of ClusterSubnetGroup instances.