Network.AWS.Redshift.DescribeClusterVersions

Description

Returns descriptions of the available Amazon Redshift cluster versions. You can call this operation even before creating any clusters to learn more about the Amazon Redshift versions. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Cluster Management Guide

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeClusterVersions.html

Request

data DescribeClusterVersions

Request constructor

describeClusterVersions :: DescribeClusterVersions

DescribeClusterVersions constructor.

The fields accessible through corresponding lenses are:

dcvClusterParameterGroupFamily :: Maybe Text

dcvClusterVersion :: Maybe Text

dcvMarker :: Maybe Text

dcvMaxRecords :: Maybe Int

Request lenses

dcvClusterParameterGroupFamily :: Lens' DescribeClusterVersions (Maybe Text)

The name of a specific cluster parameter group family to return details for.

Constraints:

Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens

dcvClusterVersion :: Lens' DescribeClusterVersions (Maybe Text)

The specific cluster version to return.

Example: '1.0'

dcvMarker :: Lens' DescribeClusterVersions (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterVersions request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dcvMaxRecords :: Lens' DescribeClusterVersions (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: '100'

Constraints: minimum 20, maximum 100.

Response

data DescribeClusterVersionsResponse

Response constructor

describeClusterVersionsResponse :: DescribeClusterVersionsResponse

DescribeClusterVersionsResponse constructor.

The fields accessible through corresponding lenses are:

dcvrClusterVersions :: [ClusterVersion]

dcvrMarker :: Maybe Text

Response lenses

dcvrClusterVersions :: Lens' DescribeClusterVersionsResponse [ClusterVersion]

A list of Version elements.

dcvrMarker :: Lens' DescribeClusterVersionsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.