Network.AWS.Redshift.DescribeClusters

Description

Returns properties of provisioned clusters including general cluster properties, cluster database properties, maintenance and backup properties, and security and access properties. This operation supports pagination. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Management Guide .

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeClusters.html

Request

data DescribeClusters

Request constructor

describeClusters :: DescribeClusters

DescribeClusters constructor.

The fields accessible through corresponding lenses are:

dcClusterIdentifier :: Maybe Text

dcMarker :: Maybe Text

dcMaxRecords :: Maybe Int

Request lenses

dcClusterIdentifier :: Lens' DescribeClusters (Maybe Text)

The unique identifier of a cluster whose properties you are requesting. This parameter is case sensitive. The default is that all clusters defined for an account are returned.

dcMarker :: Lens' DescribeClusters (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusters request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request. Constraints: You can specify either the ClusterIdentifier parameter or the Marker parameter, but not both.

dcMaxRecords :: Lens' DescribeClusters (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

Response

data DescribeClustersResponse

Response constructor

describeClustersResponse :: DescribeClustersResponse

DescribeClustersResponse constructor.

The fields accessible through corresponding lenses are:

dcrClusters :: [Cluster]

dcrMarker :: Maybe Text

Response lenses

dcrClusters :: Lens' DescribeClustersResponse [Cluster]

A list of Cluster objects, where each object describes one cluster.

dcrMarker :: Lens' DescribeClustersResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.