Network.AWS.Redshift.DescribeDefaultClusterParameters

Description

Returns a list of parameter settings for the specified parameter group family. For more information about managing parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Management Guide. https:/redshift.us-east-1.amazonaws.com ?Action=DescribeDefaultClusterParameters &ParameterGroupFamily=redshift-1.0 &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20121207us-east-1redshiftaws4_request &x-amz-date=20121207T231708Z &x-amz-signedheaders=content-type;host;x-amz-date redshift-1.0 ISO, MDY string engine-default true Sets the display format for date and time values. datestyle 0 integer engine-default true Sets the number of digits displayed for floating-point values -15-2 extra_float_digits default string engine-default true This parameter applies a user-defined label to a group of queries that are run during the same session.. query_group false boolean engine-default true require ssl for all databaseconnections true,false require_ssl $user, public string engine-default true Sets the schema search order for names that are not schema-qualified. search_path 0 integer engine-default true Aborts any statement that takes over the specified number of milliseconds. statement_timeout [{&quot;query_concurrency&quot;:5}] string engine-default true wlm json configuration wlm_json_configuration 396df00b-40c4-11e2-82cf-0b45b05c0221.

Request

data DescribeDefaultClusterParameters

Request constructor

mkDescribeDefaultClusterParameters

Arguments

:: Text
ddcpParameterGroupFamily

-> DescribeDefaultClusterParameters

Smart constructor for the minimum required parameters to construct a valid DescribeDefaultClusterParameters request.

The fields accessible through corresponding lenses are:

ParameterGroupFamily :: Text

MaxRecords :: Maybe Integer

Marker :: Maybe Text

Request lenses

ddcpParameterGroupFamily :: Lens' DescribeDefaultClusterParameters Text

The name of the cluster parameter group family.

ddcpMaxRecords :: Lens' DescribeDefaultClusterParameters (Maybe Integer)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

ddcpMarker :: Lens' DescribeDefaultClusterParameters (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDefaultClusterParameters request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

Response

data DescribeDefaultClusterParametersResponse

Response constructor

mkDescribeDefaultClusterParametersResponse

Arguments

:: DefaultClusterParameters
ddcprDefaultClusterParameters

-> DescribeDefaultClusterParametersResponse

Smart constructor for the minimum required parameters to construct a valid DescribeDefaultClusterParametersResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

DefaultClusterParameters :: DefaultClusterParameters

Response lenses

ddcprDefaultClusterParameters :: Lens' DescribeDefaultClusterParametersResponse DefaultClusterParameters

Describes the default cluster parameters for a parameter group family.