Network.AWS.Redshift.DescribeHSMConfigurations

Description

Returns information about the specified Amazon Redshift HSM configuration. If no configuration ID is specified, returns information about all the HSM configurations owned by your AWS customer account.

If you specify both tag keys and tag values in the same request, Amazon Redshift returns all HSM connections that match any combination of the specified keys and values. For example, if you have owner and environment for tag keys, and admin and test for tag values, all HSM connections that have any combination of those values are returned.

If both tag keys and values are omitted from the request, HSM connections are returned regardless of whether they have tag keys or values associated with them.

This operation returns paginated results.

Creating a Request

describeHSMConfigurations :: DescribeHSMConfigurations

Creates a value of DescribeHSMConfigurations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhsmcTagValues - A tag value or values for which you want to return all matching HSM configurations that are associated with the specified tag value or values. For example, suppose that you have HSM configurations that are tagged with values called admin and test . If you specify both of these tag values in the request, Amazon Redshift returns a response with the HSM configurations that have either or both of these tag values associated with them.

dhsmcHSMConfigurationIdentifier - The identifier of a specific Amazon Redshift HSM configuration to be described. If no identifier is specified, information is returned for all HSM configurations owned by your AWS customer account.

dhsmcTagKeys - A tag key or keys for which you want to return all matching HSM configurations that are associated with the specified key or keys. For example, suppose that you have HSM configurations that are tagged with keys called owner and environment . If you specify both of these tag keys in the request, Amazon Redshift returns a response with the HSM configurations that have either or both of these tag keys associated with them.

dhsmcMarker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeHsmConfigurations request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dhsmcMaxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

data DescribeHSMConfigurations

See: describeHSMConfigurations smart constructor.

Request Lenses

dhsmcTagValues :: Lens' DescribeHSMConfigurations [Text]

A tag value or values for which you want to return all matching HSM configurations that are associated with the specified tag value or values. For example, suppose that you have HSM configurations that are tagged with values called admin and test . If you specify both of these tag values in the request, Amazon Redshift returns a response with the HSM configurations that have either or both of these tag values associated with them.

dhsmcHSMConfigurationIdentifier :: Lens' DescribeHSMConfigurations (Maybe Text)

The identifier of a specific Amazon Redshift HSM configuration to be described. If no identifier is specified, information is returned for all HSM configurations owned by your AWS customer account.

dhsmcTagKeys :: Lens' DescribeHSMConfigurations [Text]

A tag key or keys for which you want to return all matching HSM configurations that are associated with the specified key or keys. For example, suppose that you have HSM configurations that are tagged with keys called owner and environment . If you specify both of these tag keys in the request, Amazon Redshift returns a response with the HSM configurations that have either or both of these tag keys associated with them.

dhsmcMarker :: Lens' DescribeHSMConfigurations (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeHsmConfigurations request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dhsmcMaxRecords :: Lens' DescribeHSMConfigurations (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

Destructuring the Response

describeHSMConfigurationsResponse

Arguments

:: Int
dhcrsResponseStatus

-> DescribeHSMConfigurationsResponse

Creates a value of DescribeHSMConfigurationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhcrsMarker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dhcrsHSMConfigurations - A list of HsmConfiguration objects.

dhcrsResponseStatus - -- | The response status code.

data DescribeHSMConfigurationsResponse

See: describeHSMConfigurationsResponse smart constructor.

Response Lenses

dhcrsMarker :: Lens' DescribeHSMConfigurationsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dhcrsHSMConfigurations :: Lens' DescribeHSMConfigurationsResponse [HSMConfiguration]

A list of HsmConfiguration objects.

dhcrsResponseStatus :: Lens' DescribeHSMConfigurationsResponse Int

- | The response status code.