Network.AWS.Redshift.DescribeLoggingStatus

Description

Describes whether information, such as queries and connection attempts, is being logged for the specified Amazon Redshift cluster.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeLoggingStatus.html

Request

data DescribeLoggingStatus

Request constructor

describeLoggingStatus

Arguments

:: Text
dlsClusterIdentifier

-> DescribeLoggingStatus

DescribeLoggingStatus constructor.

The fields accessible through corresponding lenses are:

dlsClusterIdentifier :: Text

Request lenses

dlsClusterIdentifier :: Lens' DescribeLoggingStatus Text

The identifier of the cluster to get the logging status from.

Example: examplecluster

Response

data DescribeLoggingStatusResponse

Response constructor

describeLoggingStatusResponse :: DescribeLoggingStatusResponse

DescribeLoggingStatusResponse constructor.

The fields accessible through corresponding lenses are:

dlsrBucketName :: Maybe Text

dlsrLastFailureMessage :: Maybe Text

dlsrLastFailureTime :: Maybe UTCTime

dlsrLastSuccessfulDeliveryTime :: Maybe UTCTime

dlsrLoggingEnabled :: Maybe Bool

dlsrS3KeyPrefix :: Maybe Text

Response lenses

dlsrBucketName :: Lens' DescribeLoggingStatusResponse (Maybe Text)

The name of the S3 bucket where the log files are stored.

dlsrLastFailureMessage :: Lens' DescribeLoggingStatusResponse (Maybe Text)

The message indicating that logs failed to be delivered.

dlsrLastFailureTime :: Lens' DescribeLoggingStatusResponse (Maybe UTCTime)

The last time when logs failed to be delivered.

dlsrLastSuccessfulDeliveryTime :: Lens' DescribeLoggingStatusResponse (Maybe UTCTime)

The last time when logs were delivered.

dlsrLoggingEnabled :: Lens' DescribeLoggingStatusResponse (Maybe Bool)

true if logging is on, false if logging is off.

dlsrS3KeyPrefix :: Lens' DescribeLoggingStatusResponse (Maybe Text)

The prefix applied to the log file names.