Network.AWS.Redshift.DescribeOrderableClusterOptions

Description

Returns a list of orderable cluster options. Before you create a new cluster you can use this operation to find what options are available, such as the EC2 Availability Zones (AZ) in the specific AWS region that you can specify, and the node types you can request. The node types differ by available storage, memory, CPU and price. With the cost involved you might want to obtain a list of cluster options in the specific region and specify values when creating a cluster. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Cluster Management Guide

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeOrderableClusterOptions.html

Request

data DescribeOrderableClusterOptions

Request constructor

describeOrderableClusterOptions :: DescribeOrderableClusterOptions

DescribeOrderableClusterOptions constructor.

The fields accessible through corresponding lenses are:

docoClusterVersion :: Maybe Text

docoMarker :: Maybe Text

docoMaxRecords :: Maybe Int

docoNodeType :: Maybe Text

Request lenses

docoClusterVersion :: Lens' DescribeOrderableClusterOptions (Maybe Text)

The version filter value. Specify this parameter to show only the available offerings matching the specified version.

Default: All versions.

Constraints: Must be one of the version returned from DescribeClusterVersions .

docoMarker :: Lens' DescribeOrderableClusterOptions (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeOrderableClusterOptions request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

docoMaxRecords :: Lens' DescribeOrderableClusterOptions (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: '100'

Constraints: minimum 20, maximum 100.

docoNodeType :: Lens' DescribeOrderableClusterOptions (Maybe Text)

The node type filter value. Specify this parameter to show only the available offerings matching the specified node type.

Response

data DescribeOrderableClusterOptionsResponse

Response constructor

describeOrderableClusterOptionsResponse :: DescribeOrderableClusterOptionsResponse

DescribeOrderableClusterOptionsResponse constructor.

The fields accessible through corresponding lenses are:

docorMarker :: Maybe Text

docorOrderableClusterOptions :: [OrderableClusterOption]

Response lenses

docorMarker :: Lens' DescribeOrderableClusterOptionsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

docorOrderableClusterOptions :: Lens' DescribeOrderableClusterOptionsResponse [OrderableClusterOption]

An OrderableClusterOption structure containing information about orderable options for the Cluster.