Network.AWS.Redshift.DescribeReservedNodes

Description

Returns the descriptions of the reserved nodes.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeReservedNodes.html

Request

data DescribeReservedNodes

Request constructor

describeReservedNodes :: DescribeReservedNodes

DescribeReservedNodes constructor.

The fields accessible through corresponding lenses are:

drnMarker :: Maybe Text

drnMaxRecords :: Maybe Int

drnReservedNodeId :: Maybe Text

Request lenses

drnMarker :: Lens' DescribeReservedNodes (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeReservedNodes request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

drnMaxRecords :: Lens' DescribeReservedNodes (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: '100'

Constraints: minimum 20, maximum 100.

drnReservedNodeId :: Lens' DescribeReservedNodes (Maybe Text)

Identifier for the node reservation.

Response

data DescribeReservedNodesResponse

Response constructor

describeReservedNodesResponse :: DescribeReservedNodesResponse

DescribeReservedNodesResponse constructor.

The fields accessible through corresponding lenses are:

drnrMarker :: Maybe Text

drnrReservedNodes :: [ReservedNode]

Response lenses

drnrMarker :: Lens' DescribeReservedNodesResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

drnrReservedNodes :: Lens' DescribeReservedNodesResponse [ReservedNode]

The list of reserved nodes.