Network.AWS.Redshift.DescribeResize

Description

Returns information about the last resize operation for the specified cluster. If no resize operation has ever been initiated for the specified cluster, a 'HTTP 404' error is returned. If a resize operation was initiated and completed, the status of the resize remains as SUCCEEDED until the next resize.

A resize operation can be requested using ModifyCluster and specifying a different number or type of nodes for the cluster.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeResize.html

Request

data DescribeResize

Request constructor

describeResize

Arguments

:: Text
drClusterIdentifier

-> DescribeResize

DescribeResize constructor.

The fields accessible through corresponding lenses are:

drClusterIdentifier :: Text

Request lenses

drClusterIdentifier :: Lens' DescribeResize Text

The unique identifier of a cluster whose resize progress you are requesting. This parameter is case-sensitive.

By default, resize operations for all clusters defined for an AWS account are returned.

Response

data DescribeResizeResponse

Response constructor

describeResizeResponse :: DescribeResizeResponse

DescribeResizeResponse constructor.

The fields accessible through corresponding lenses are:

drrAvgResizeRateInMegaBytesPerSecond :: Maybe Double

drrElapsedTimeInSeconds :: Maybe Integer

drrEstimatedTimeToCompletionInSeconds :: Maybe Integer

drrImportTablesCompleted :: [Text]

drrImportTablesInProgress :: [Text]

drrImportTablesNotStarted :: [Text]

drrProgressInMegaBytes :: Maybe Integer

drrStatus :: Maybe Text

drrTargetClusterType :: Maybe Text

drrTargetNodeType :: Maybe Text

drrTargetNumberOfNodes :: Maybe Int

drrTotalResizeDataInMegaBytes :: Maybe Integer

Response lenses

drrAvgResizeRateInMegaBytesPerSecond :: Lens' DescribeResizeResponse (Maybe Double)

The average rate of the resize operation over the last few minutes, measured in megabytes per second. After the resize operation completes, this value shows the average rate of the entire resize operation.

drrElapsedTimeInSeconds :: Lens' DescribeResizeResponse (Maybe Integer)

The amount of seconds that have elapsed since the resize operation began. After the resize operation completes, this value shows the total actual time, in seconds, for the resize operation.

drrEstimatedTimeToCompletionInSeconds :: Lens' DescribeResizeResponse (Maybe Integer)

The estimated time remaining, in seconds, until the resize operation is complete. This value is calculated based on the average resize rate and the estimated amount of data remaining to be processed. Once the resize operation is complete, this value will be 0.

drrImportTablesCompleted :: Lens' DescribeResizeResponse [Text]

The names of tables that have been completely imported .

Valid Values: List of table names.

drrImportTablesInProgress :: Lens' DescribeResizeResponse [Text]

The names of tables that are being currently imported.

Valid Values: List of table names.

drrImportTablesNotStarted :: Lens' DescribeResizeResponse [Text]

The names of tables that have not been yet imported.

Valid Values: List of table names

drrProgressInMegaBytes :: Lens' DescribeResizeResponse (Maybe Integer)

While the resize operation is in progress, this value shows the current amount of data, in megabytes, that has been processed so far. When the resize operation is complete, this value shows the total amount of data, in megabytes, on the cluster, which may be more or less than TotalResizeDataInMegaBytes (the estimated total amount of data before resize).

drrStatus :: Lens' DescribeResizeResponse (Maybe Text)

The status of the resize operation.

Valid Values: NONE | IN_PROGRESS | FAILED | SUCCEEDED

drrTargetClusterType :: Lens' DescribeResizeResponse (Maybe Text)

The cluster type after the resize operation is complete.

Valid Values: 'multi-node' | 'single-node'

drrTargetNodeType :: Lens' DescribeResizeResponse (Maybe Text)

The node type that the cluster will have after the resize operation is complete.

drrTargetNumberOfNodes :: Lens' DescribeResizeResponse (Maybe Int)

The number of nodes that the cluster will have after the resize operation is complete.

drrTotalResizeDataInMegaBytes :: Lens' DescribeResizeResponse (Maybe Integer)

The estimated total amount of data, in megabytes, on the cluster before the resize operation began.