Network.AWS.Redshift.DescribeSnapshotCopyGrants

Description

Returns a list of snapshot copy grants owned by the AWS account in the destination region.

For more information about managing snapshot copy grants, go to Amazon Redshift Database Encryption in the Amazon Redshift Cluster Management Guide .

Creating a Request

describeSnapshotCopyGrants :: DescribeSnapshotCopyGrants

Creates a value of DescribeSnapshotCopyGrants with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dscgsTagValues - A tag value or values for which you want to return all matching resources that are associated with the specified value or values. For example, suppose that you have resources tagged with values called admin and test . If you specify both of these tag values in the request, Amazon Redshift returns a response with all resources that have either or both of these tag values associated with them.

dscgsTagKeys - A tag key or keys for which you want to return all matching resources that are associated with the specified key or keys. For example, suppose that you have resources tagged with keys called owner and environment . If you specify both of these tag keys in the request, Amazon Redshift returns a response with all resources that have either or both of these tag keys associated with them.

dscgsMarker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeSnapshotCopyGrant request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request. Constraints: You can specify either the SnapshotCopyGrantName parameter or the Marker parameter, but not both.

dscgsMaxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

dscgsSnapshotCopyGrantName - The name of the snapshot copy grant.

data DescribeSnapshotCopyGrants

The result of the DescribeSnapshotCopyGrants action.

See: describeSnapshotCopyGrants smart constructor.

Request Lenses

dscgsTagValues :: Lens' DescribeSnapshotCopyGrants [Text]

A tag value or values for which you want to return all matching resources that are associated with the specified value or values. For example, suppose that you have resources tagged with values called admin and test . If you specify both of these tag values in the request, Amazon Redshift returns a response with all resources that have either or both of these tag values associated with them.

dscgsTagKeys :: Lens' DescribeSnapshotCopyGrants [Text]

A tag key or keys for which you want to return all matching resources that are associated with the specified key or keys. For example, suppose that you have resources tagged with keys called owner and environment . If you specify both of these tag keys in the request, Amazon Redshift returns a response with all resources that have either or both of these tag keys associated with them.

dscgsMarker :: Lens' DescribeSnapshotCopyGrants (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeSnapshotCopyGrant request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request. Constraints: You can specify either the SnapshotCopyGrantName parameter or the Marker parameter, but not both.

dscgsMaxRecords :: Lens' DescribeSnapshotCopyGrants (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

dscgsSnapshotCopyGrantName :: Lens' DescribeSnapshotCopyGrants (Maybe Text)

The name of the snapshot copy grant.

Destructuring the Response

describeSnapshotCopyGrantsResponse

Arguments

:: Int
dscgrsResponseStatus

-> DescribeSnapshotCopyGrantsResponse

Creates a value of DescribeSnapshotCopyGrantsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dscgrsSnapshotCopyGrants - The list of SnapshotCopyGrant objects.

dscgrsMarker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeSnapshotCopyGrant request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request. Constraints: You can specify either the SnapshotCopyGrantName parameter or the Marker parameter, but not both.

dscgrsResponseStatus - -- | The response status code.

data DescribeSnapshotCopyGrantsResponse

See: describeSnapshotCopyGrantsResponse smart constructor.

Response Lenses

dscgrsSnapshotCopyGrants :: Lens' DescribeSnapshotCopyGrantsResponse [SnapshotCopyGrant]

The list of SnapshotCopyGrant objects.

dscgrsMarker :: Lens' DescribeSnapshotCopyGrantsResponse (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeSnapshotCopyGrant request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request. Constraints: You can specify either the SnapshotCopyGrantName parameter or the Marker parameter, but not both.

dscgrsResponseStatus :: Lens' DescribeSnapshotCopyGrantsResponse Int

- | The response status code.