Network.AWS.Redshift.DescribeTableRestoreStatus

Description

Lists the status of one or more table restore requests made using the RestoreTableFromClusterSnapshot API action. If you don't specify a value for the TableRestoreRequestId parameter, then DescribeTableRestoreStatus returns the status of all table restore requests ordered by the date and time of the request in ascending order. Otherwise DescribeTableRestoreStatus returns the status of the table specified by TableRestoreRequestId .

Creating a Request

describeTableRestoreStatus :: DescribeTableRestoreStatus

Creates a value of DescribeTableRestoreStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtrssTableRestoreRequestId - The identifier of the table restore request to return status for. If you don't specify a TableRestoreRequestId value, then DescribeTableRestoreStatus returns the status of all in-progress table restore requests.

dtrssClusterIdentifier - The Amazon Redshift cluster that the table is being restored to.

dtrssMarker - An optional pagination token provided by a previous DescribeTableRestoreStatus request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by the MaxRecords parameter.

dtrssMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

data DescribeTableRestoreStatus

See: describeTableRestoreStatus smart constructor.

Request Lenses

dtrssTableRestoreRequestId :: Lens' DescribeTableRestoreStatus (Maybe Text)

The identifier of the table restore request to return status for. If you don't specify a TableRestoreRequestId value, then DescribeTableRestoreStatus returns the status of all in-progress table restore requests.

dtrssClusterIdentifier :: Lens' DescribeTableRestoreStatus (Maybe Text)

The Amazon Redshift cluster that the table is being restored to.

dtrssMarker :: Lens' DescribeTableRestoreStatus (Maybe Text)

An optional pagination token provided by a previous DescribeTableRestoreStatus request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by the MaxRecords parameter.

dtrssMaxRecords :: Lens' DescribeTableRestoreStatus (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Destructuring the Response

describeTableRestoreStatusResponse

Arguments

:: Int
dtrsrsResponseStatus

-> DescribeTableRestoreStatusResponse

Creates a value of DescribeTableRestoreStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtrsrsMarker - A pagination token that can be used in a subsequent DescribeTableRestoreStatus request.

dtrsrsTableRestoreStatusDetails - A list of status details for one or more table restore requests.

dtrsrsResponseStatus - -- | The response status code.

data DescribeTableRestoreStatusResponse

See: describeTableRestoreStatusResponse smart constructor.

Response Lenses

dtrsrsMarker :: Lens' DescribeTableRestoreStatusResponse (Maybe Text)

A pagination token that can be used in a subsequent DescribeTableRestoreStatus request.

dtrsrsTableRestoreStatusDetails :: Lens' DescribeTableRestoreStatusResponse [TableRestoreStatus]

A list of status details for one or more table restore requests.

dtrsrsResponseStatus :: Lens' DescribeTableRestoreStatusResponse Int

- | The response status code.