Network.AWS.Redshift.DescribeTags

Description

Returns a list of tags. You can return tags from a specific resource by specifying an ARN, or you can return all tags for a given type of resource, such as clusters, snapshots, and so on.

The following are limitations for DescribeTags: You cannot specify an ARN and a resource-type value together in the same request. You cannot use the MaxRecords and Marker parameters together with the ARN parameter. The MaxRecords parameter can be a range from 10 to 50 results to return in a request.

If you specify both tag keys and tag values in the same request, Amazon Redshift returns all resources that match any combination of the specified keys and values. For example, if you have owner and environment for tag keys, and admin and test for tag values, all resources that have any combination of those values are returned.

If both tag keys and values are omitted from the request, resources are returned regardless of whether they have tag keys or values associated with them.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeTags.html

Request

data DescribeTags

Request constructor

describeTags :: DescribeTags

DescribeTags constructor.

The fields accessible through corresponding lenses are:

dtMarker :: Maybe Text

dtMaxRecords :: Maybe Int

dtResourceName :: Maybe Text

dtResourceType :: Maybe Text

dtTagKeys :: [Text]

dtTagValues :: [Text]

Request lenses

dtMarker :: Lens' DescribeTags (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the marker parameter and retrying the command. If the marker field is empty, all response records have been retrieved for the request.

dtMaxRecords :: Lens' DescribeTags (Maybe Int)

The maximum number or response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

dtResourceName :: Lens' DescribeTags (Maybe Text)

The Amazon Resource Name (ARN) for which you want to describe the tag or tags. For example, 'arn:aws:redshift:us-east-1:123456789:cluster:t1'.

dtResourceType :: Lens' DescribeTags (Maybe Text)

The type of resource with which you want to view tags. Valid resource types are: Cluster CIDR/IP EC2 security group Snapshot Cluster security group Subnet group HSM connection HSM certificate Parameter group

For more information about Amazon Redshift resource types and constructing ARNs, go to Constructing an Amazon Redshift Amazon Resource Name (ARN) in the Amazon Redshift Cluster Management Guide.

dtTagKeys :: Lens' DescribeTags [Text]

A tag key or keys for which you want to return all matching resources that are associated with the specified key or keys. For example, suppose that you have resources tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with all resources that have either or both of these tag keys associated with them.

dtTagValues :: Lens' DescribeTags [Text]

A tag value or values for which you want to return all matching resources that are associated with the specified value or values. For example, suppose that you have resources tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with all resources that have either or both of these tag values associated with them.

Response

data DescribeTagsResponse

Response constructor

describeTagsResponse :: DescribeTagsResponse

DescribeTagsResponse constructor.

The fields accessible through corresponding lenses are:

dtrMarker :: Maybe Text

dtrTaggedResources :: [TaggedResource]

Response lenses

dtrMarker :: Lens' DescribeTagsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dtrTaggedResources :: Lens' DescribeTagsResponse [TaggedResource]

A list of tags with their associated resources.