Network.AWS.Redshift.DisableSnapshotCopy

Description

Disables the automatic copying of snapshots from one region to another region for a specified cluster.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DisableSnapshotCopy.html

Request

data DisableSnapshotCopy

Request constructor

disableSnapshotCopy

Arguments

:: Text
dscClusterIdentifier

-> DisableSnapshotCopy

DisableSnapshotCopy constructor.

The fields accessible through corresponding lenses are:

dscClusterIdentifier :: Text

Request lenses

dscClusterIdentifier :: Lens' DisableSnapshotCopy Text

The unique identifier of the source cluster that you want to disable copying of snapshots to a destination region.

Constraints: Must be the valid name of an existing cluster that has cross-region snapshot copy enabled.

Response

data DisableSnapshotCopyResponse

Response constructor

disableSnapshotCopyResponse :: DisableSnapshotCopyResponse

DisableSnapshotCopyResponse constructor.

The fields accessible through corresponding lenses are:

dscrCluster :: Maybe Cluster

Response lenses

dscrCluster :: Lens' DisableSnapshotCopyResponse (Maybe Cluster)