Network.AWS.Redshift.EnableLogging

Description

Starts logging information, such as queries and connection attempts, for the specified Amazon Redshift cluster.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_EnableLogging.html

Request

data EnableLogging

Request constructor

enableLogging

Arguments

:: Text
elClusterIdentifier

-> Text
elBucketName

-> EnableLogging

EnableLogging constructor.

The fields accessible through corresponding lenses are:

elBucketName :: Text

elClusterIdentifier :: Text

elS3KeyPrefix :: Maybe Text

Request lenses

elBucketName :: Lens' EnableLogging Text

The name of an existing S3 bucket where the log files are to be stored.

Constraints:

Must be in the same region as the cluster The cluster must have read bucket and put object permissions

elClusterIdentifier :: Lens' EnableLogging Text

The identifier of the cluster on which logging is to be started.

Example: examplecluster

elS3KeyPrefix :: Lens' EnableLogging (Maybe Text)

The prefix applied to the log file names.

Constraints:

Cannot exceed 512 characters Cannot contain spaces( ), double quotes ("), single quotes ('), a backslash (), or control characters. The hexadecimal codes for invalid characters are: x00 to x20 x22 x27 x5c x7f or larger

Response

data EnableLoggingResponse

Response constructor

enableLoggingResponse :: EnableLoggingResponse

EnableLoggingResponse constructor.

The fields accessible through corresponding lenses are:

elrBucketName :: Maybe Text

elrLastFailureMessage :: Maybe Text

elrLastFailureTime :: Maybe UTCTime

elrLastSuccessfulDeliveryTime :: Maybe UTCTime

elrLoggingEnabled :: Maybe Bool

elrS3KeyPrefix :: Maybe Text

Response lenses

elrBucketName :: Lens' EnableLoggingResponse (Maybe Text)

The name of the S3 bucket where the log files are stored.

elrLastFailureMessage :: Lens' EnableLoggingResponse (Maybe Text)

The message indicating that logs failed to be delivered.

elrLastFailureTime :: Lens' EnableLoggingResponse (Maybe UTCTime)

The last time when logs failed to be delivered.

elrLastSuccessfulDeliveryTime :: Lens' EnableLoggingResponse (Maybe UTCTime)

The last time when logs were delivered.

elrLoggingEnabled :: Lens' EnableLoggingResponse (Maybe Bool)

true if logging is on, false if logging is off.

elrS3KeyPrefix :: Lens' EnableLoggingResponse (Maybe Text)

The prefix applied to the log file names.