Network.AWS.Redshift.GetClusterCredentials

Description

Returns a database user name and temporary password with temporary authorization to log in to an Amazon Redshift database. The action returns the database user name prefixed with IAM: if AutoCreate is False or IAMA: if AutoCreate is True . You can optionally specify one or more database user groups that the user will join at log in. By default, the temporary credentials expire in 900 seconds. You can optionally specify a duration between 900 seconds (15 minutes) and 3600 seconds (60 minutes). For more information, see Generating IAM Database User Credentials in the Amazon Redshift Cluster Management Guide.

The IAM user or role that executes GetClusterCredentials must have an IAM policy attached that allows the redshift:GetClusterCredentials action with access to the dbuser resource on the cluster. The user name specified for dbuser in the IAM policy and the user name specified for the DbUser parameter must match.

If the DbGroups parameter is specified, the IAM policy must allow the redshift:JoinGroup action with access to the listed dbgroups .

In addition, if the AutoCreate parameter is set to True , then the policy must include the redshift:CreateClusterUser privilege.

If the DbName parameter is specified, the IAM policy must allow access to the resource dbname for the specified database name.

Creating a Request

getClusterCredentials

Arguments

:: Text
gccDBUser

-> Text
gccClusterIdentifier

-> GetClusterCredentials

Creates a value of GetClusterCredentials with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gccDBGroups - A list of the names of existing database groups that DbUser will join for the current session. If not specified, the new user is added only to PUBLIC.

gccDurationSeconds - The number of seconds until the returned temporary password expires. Constraint: minimum 900, maximum 3600. Default: 900

gccAutoCreate - Create a database user with the name specified for DbUser if one does not exist.

gccDBName - The name of a database that DbUser is authorized to log on to. If DbName is not specified, DbUser can log in to any existing database. Constraints: * Must be 1 to 64 alphanumeric characters or hyphens * Must contain only lowercase letters. * Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

gccDBUser - The name of a database user. If a user name matching DbUser exists in the database, the temporary user credentials have the same permissions as the existing user. If DbUser doesn't exist in the database and Autocreate is True , a new user is created using the value for DbUser with PUBLIC permissions. If a database user matching the value for DbUser doesn't exist and Autocreate is False , then the command succeeds but the connection attempt will fail because the user doesn't exist in the database. For more information, see CREATE USER in the Amazon Redshift Database Developer Guide. Constraints: * Must be 1 to 128 alphanumeric characters or hyphens * Must contain only lowercase letters. * First character must be a letter. * Must not contain a colon ( : ) or slash ( ). * Cannot be a reserved word. A list of reserved words can be found in <http:docs.aws.amazon.comredshiftlatestdg/r_pg_keywords.html Reserved Words> in the Amazon Redshift Database Developer Guide.

gccClusterIdentifier - The unique identifier of the cluster that contains the database for which your are requesting credentials. This parameter is case sensitive.

data GetClusterCredentials

The request parameters to get cluster credentials.

See: getClusterCredentials smart constructor.

Request Lenses

gccDBGroups :: Lens' GetClusterCredentials [Text]

A list of the names of existing database groups that DbUser will join for the current session. If not specified, the new user is added only to PUBLIC.

gccDurationSeconds :: Lens' GetClusterCredentials (Maybe Int)

The number of seconds until the returned temporary password expires. Constraint: minimum 900, maximum 3600. Default: 900

gccAutoCreate :: Lens' GetClusterCredentials (Maybe Bool)

Create a database user with the name specified for DbUser if one does not exist.

gccDBName :: Lens' GetClusterCredentials (Maybe Text)

The name of a database that DbUser is authorized to log on to. If DbName is not specified, DbUser can log in to any existing database. Constraints: * Must be 1 to 64 alphanumeric characters or hyphens * Must contain only lowercase letters. * Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

gccDBUser :: Lens' GetClusterCredentials Text

The name of a database user. If a user name matching DbUser exists in the database, the temporary user credentials have the same permissions as the existing user. If DbUser doesn't exist in the database and Autocreate is True , a new user is created using the value for DbUser with PUBLIC permissions. If a database user matching the value for DbUser doesn't exist and Autocreate is False , then the command succeeds but the connection attempt will fail because the user doesn't exist in the database. For more information, see CREATE USER in the Amazon Redshift Database Developer Guide. Constraints: * Must be 1 to 128 alphanumeric characters or hyphens * Must contain only lowercase letters. * First character must be a letter. * Must not contain a colon ( : ) or slash ( ). * Cannot be a reserved word. A list of reserved words can be found in <http:docs.aws.amazon.comredshiftlatestdg/r_pg_keywords.html Reserved Words> in the Amazon Redshift Database Developer Guide.

gccClusterIdentifier :: Lens' GetClusterCredentials Text

The unique identifier of the cluster that contains the database for which your are requesting credentials. This parameter is case sensitive.

Destructuring the Response

getClusterCredentialsResponse

Arguments

:: Int
gccrsResponseStatus

-> GetClusterCredentialsResponse

Creates a value of GetClusterCredentialsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gccrsDBUser - A database user name that is authorized to log on to the database DbName using the password DbPassword . If the DbGroups parameter is specifed, DbUser is added to the listed groups for the current session. The user name is prefixed with IAM: for an existing user name or IAMA: if the user was auto-created.

gccrsExpiration - The date and time DbPassword expires.

gccrsDBPassword - A temporary password that authorizes the user name returned by DbUser to log on to the database DbName .

gccrsResponseStatus - -- | The response status code.

data GetClusterCredentialsResponse

Temporary credentials with authorization to log in to an Amazon Redshift database.

See: getClusterCredentialsResponse smart constructor.

Response Lenses

gccrsDBUser :: Lens' GetClusterCredentialsResponse (Maybe Text)

A database user name that is authorized to log on to the database DbName using the password DbPassword . If the DbGroups parameter is specifed, DbUser is added to the listed groups for the current session. The user name is prefixed with IAM: for an existing user name or IAMA: if the user was auto-created.

gccrsExpiration :: Lens' GetClusterCredentialsResponse (Maybe UTCTime)

The date and time DbPassword expires.

gccrsDBPassword :: Lens' GetClusterCredentialsResponse (Maybe Text)

A temporary password that authorizes the user name returned by DbUser to log on to the database DbName .

gccrsResponseStatus :: Lens' GetClusterCredentialsResponse Int

- | The response status code.