Network.AWS.Redshift.ModifyClusterIAMRoles

Description

Modifies the list of AWS Identity and Access Management (IAM) roles that can be used by the cluster to access other AWS services.

A cluster can have up to 10 IAM roles associated at any time.

Creating a Request

modifyClusterIAMRoles

Arguments

:: Text
mcirClusterIdentifier

-> ModifyClusterIAMRoles

Creates a value of ModifyClusterIAMRoles with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mcirRemoveIAMRoles - Zero or more IAM roles in ARN format to disassociate from the cluster. You can disassociate up to 10 IAM roles from a single cluster in a single request.

mcirAddIAMRoles - Zero or more IAM roles to associate with the cluster. The roles must be in their Amazon Resource Name (ARN) format. You can associate up to 10 IAM roles with a single cluster in a single request.

mcirClusterIdentifier - The unique identifier of the cluster for which you want to associate or disassociate IAM roles.

data ModifyClusterIAMRoles

See: modifyClusterIAMRoles smart constructor.

Request Lenses

mcirRemoveIAMRoles :: Lens' ModifyClusterIAMRoles [Text]

Zero or more IAM roles in ARN format to disassociate from the cluster. You can disassociate up to 10 IAM roles from a single cluster in a single request.

mcirAddIAMRoles :: Lens' ModifyClusterIAMRoles [Text]

Zero or more IAM roles to associate with the cluster. The roles must be in their Amazon Resource Name (ARN) format. You can associate up to 10 IAM roles with a single cluster in a single request.

mcirClusterIdentifier :: Lens' ModifyClusterIAMRoles Text

The unique identifier of the cluster for which you want to associate or disassociate IAM roles.

Destructuring the Response

modifyClusterIAMRolesResponse

Arguments

:: Int
mcirrsResponseStatus

-> ModifyClusterIAMRolesResponse

Creates a value of ModifyClusterIAMRolesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mcirrsCluster - Undocumented member.

mcirrsResponseStatus - -- | The response status code.

data ModifyClusterIAMRolesResponse

See: modifyClusterIAMRolesResponse smart constructor.

Response Lenses

mcirrsCluster :: Lens' ModifyClusterIAMRolesResponse (Maybe Cluster)

Undocumented member.

mcirrsResponseStatus :: Lens' ModifyClusterIAMRolesResponse Int

- | The response status code.