Network.AWS.Redshift.ModifyClusterSubnetGroup

Description

Modifies a cluster subnet group to include the specified list of VPC subnets. The operation replaces the existing list of subnets with the new list of subnets.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_ModifyClusterSubnetGroup.html

Request

data ModifyClusterSubnetGroup

Request constructor

modifyClusterSubnetGroup

Arguments

:: Text
mcsgClusterSubnetGroupName

-> ModifyClusterSubnetGroup

ModifyClusterSubnetGroup constructor.

The fields accessible through corresponding lenses are:

mcsgClusterSubnetGroupName :: Text

mcsgDescription :: Maybe Text

mcsgSubnetIds :: [Text]

Request lenses

mcsgClusterSubnetGroupName :: Lens' ModifyClusterSubnetGroup Text

The name of the subnet group to be modified.

mcsgDescription :: Lens' ModifyClusterSubnetGroup (Maybe Text)

A text description of the subnet group to be modified.

mcsgSubnetIds :: Lens' ModifyClusterSubnetGroup [Text]

An array of VPC subnet IDs. A maximum of 20 subnets can be modified in a single request.

Response

data ModifyClusterSubnetGroupResponse

Response constructor

modifyClusterSubnetGroupResponse :: ModifyClusterSubnetGroupResponse

ModifyClusterSubnetGroupResponse constructor.

The fields accessible through corresponding lenses are:

mcsgrClusterSubnetGroup :: Maybe ClusterSubnetGroup

Response lenses

mcsgrClusterSubnetGroup :: Lens' ModifyClusterSubnetGroupResponse (Maybe ClusterSubnetGroup)