Network.AWS.Redshift.ModifySnapshotCopyRetentionPeriod

Description

Modifies the number of days to retain automated snapshots in the destination region after they are copied from the source region.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_ModifySnapshotCopyRetentionPeriod.html

Request

data ModifySnapshotCopyRetentionPeriod

Request constructor

modifySnapshotCopyRetentionPeriod

Arguments

:: Text
mscrpClusterIdentifier

-> Int
mscrpRetentionPeriod

-> ModifySnapshotCopyRetentionPeriod

ModifySnapshotCopyRetentionPeriod constructor.

The fields accessible through corresponding lenses are:

mscrpClusterIdentifier :: Text

mscrpRetentionPeriod :: Int

Request lenses

mscrpClusterIdentifier :: Lens' ModifySnapshotCopyRetentionPeriod Text

The unique identifier of the cluster for which you want to change the retention period for automated snapshots that are copied to a destination region.

Constraints: Must be the valid name of an existing cluster that has cross-region snapshot copy enabled.

mscrpRetentionPeriod :: Lens' ModifySnapshotCopyRetentionPeriod Int

The number of days to retain automated snapshots in the destination region after they are copied from the source region.

If you decrease the retention period for automated snapshots that are copied to a destination region, Amazon Redshift will delete any existing automated snapshots that were copied to the destination region and that fall outside of the new retention period.

Constraints: Must be at least 1 and no more than 35.

Response

data ModifySnapshotCopyRetentionPeriodResponse

Response constructor

modifySnapshotCopyRetentionPeriodResponse :: ModifySnapshotCopyRetentionPeriodResponse

ModifySnapshotCopyRetentionPeriodResponse constructor.

The fields accessible through corresponding lenses are:

mscrprCluster :: Maybe Cluster

Response lenses

mscrprCluster :: Lens' ModifySnapshotCopyRetentionPeriodResponse (Maybe Cluster)