Network.AWS.Redshift.PurchaseReservedNodeOffering

Description

Allows you to purchase reserved nodes. Amazon Redshift offers a predefined set of reserved node offerings. You can purchase one of the offerings. You can call the DescribeReservedNodeOfferings API to obtain the available reserved node offerings. You can call this API by providing a specific reserved node offering and the number of nodes you want to reserve.

For more information about managing parameter groups, go to PurchasingReserved Nodes in the Amazon Redshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_PurchaseReservedNodeOffering.html

Request

data PurchaseReservedNodeOffering

Request constructor

purchaseReservedNodeOffering

Arguments

:: Text
prnoReservedNodeOfferingId

-> PurchaseReservedNodeOffering

PurchaseReservedNodeOffering constructor.

The fields accessible through corresponding lenses are:

prnoNodeCount :: Maybe Int

prnoReservedNodeOfferingId :: Text

Request lenses

prnoNodeCount :: Lens' PurchaseReservedNodeOffering (Maybe Int)

The number of reserved nodes you want to purchase.

Default: '1'

prnoReservedNodeOfferingId :: Lens' PurchaseReservedNodeOffering Text

The unique identifier of the reserved node offering you want to purchase.

Response

data PurchaseReservedNodeOfferingResponse

Response constructor

purchaseReservedNodeOfferingResponse :: PurchaseReservedNodeOfferingResponse

PurchaseReservedNodeOfferingResponse constructor.

The fields accessible through corresponding lenses are:

prnorReservedNode :: Maybe ReservedNode

Response lenses

prnorReservedNode :: Lens' PurchaseReservedNodeOfferingResponse (Maybe ReservedNode)