Network.AWS.Redshift.RebootCluster

Description

Reboots a cluster. This action is taken as soon as possible. It results in a momentary outage to the cluster, during which the cluster status is set to rebooting. A cluster event is created when the reboot is completed. Any pending cluster modifications (see ModifyCluster) are applied at this reboot. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Cluster Management Guide

http://docs.aws.amazon.com/redshift/latest/APIReference/API_RebootCluster.html

Request

data RebootCluster

Request constructor

rebootCluster

Arguments

:: Text
rcClusterIdentifier

-> RebootCluster

RebootCluster constructor.

The fields accessible through corresponding lenses are:

rcClusterIdentifier :: Text

Request lenses

rcClusterIdentifier :: Lens' RebootCluster Text

The cluster identifier.

Response

data RebootClusterResponse

Response constructor

rebootClusterResponse :: RebootClusterResponse

RebootClusterResponse constructor.

The fields accessible through corresponding lenses are:

rcrCluster :: Maybe Cluster

Response lenses

rcrCluster :: Lens' RebootClusterResponse (Maybe Cluster)