Network.AWS.Redshift.ResetClusterParameterGroup

Description

Sets one or more parameters of the specified parameter group to their default values and sets the source values of the parameters to "engine-default". To reset the entire parameter group specify the ResetAllParameters parameter. For parameter changes to take effect you must reboot any associated clusters.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_ResetClusterParameterGroup.html

Request

data ResetClusterParameterGroup

Request constructor

resetClusterParameterGroup

Arguments

:: Text
rcpgParameterGroupName

-> ResetClusterParameterGroup

ResetClusterParameterGroup constructor.

The fields accessible through corresponding lenses are:

rcpgParameterGroupName :: Text

rcpgParameters :: [Parameter]

rcpgResetAllParameters :: Maybe Bool

Request lenses

rcpgParameterGroupName :: Lens' ResetClusterParameterGroup Text

The name of the cluster parameter group to be reset.

rcpgParameters :: Lens' ResetClusterParameterGroup [Parameter]

An array of names of parameters to be reset. If ResetAllParameters option is not used, then at least one parameter name must be supplied.

Constraints: A maximum of 20 parameters can be reset in a single request.

rcpgResetAllParameters :: Lens' ResetClusterParameterGroup (Maybe Bool)

If true, all parameters in the specified parameter group will be reset to their default values.

Default: true

Response

data ResetClusterParameterGroupResponse

Response constructor

resetClusterParameterGroupResponse :: ResetClusterParameterGroupResponse

ResetClusterParameterGroupResponse constructor.

The fields accessible through corresponding lenses are:

rcpgrParameterGroupName :: Maybe Text

rcpgrParameterGroupStatus :: Maybe Text

Response lenses

rcpgrParameterGroupName :: Lens' ResetClusterParameterGroupResponse (Maybe Text)

The name of the cluster parameter group.

rcpgrParameterGroupStatus :: Lens' ResetClusterParameterGroupResponse (Maybe Text)

The status of the parameter group. For example, if you made a change to a parameter group name-value pair, then the change could be pending a reboot of an associated cluster.