Network.AWS.Redshift.RestoreTableFromClusterSnapshot

Description

Creates a new table from a table in an Amazon Redshift cluster snapshot. You must create the new table within the Amazon Redshift cluster that the snapshot was taken from.

You cannot use RestoreTableFromClusterSnapshot to restore a table with the same name as an existing table in an Amazon Redshift cluster. That is, you cannot overwrite an existing table in a cluster with a restored table. If you want to replace your original table with a new, restored table, then rename or drop your original table before you call RestoreTableFromClusterSnapshot . When you have renamed your original table, then you can pass the original name of the table as the NewTableName parameter value in the call to RestoreTableFromClusterSnapshot . This way, you can replace the original table with the table created from the snapshot.

Creating a Request

restoreTableFromClusterSnapshot

Arguments

:: Text
rtfcsClusterIdentifier

-> Text
rtfcsSnapshotIdentifier

-> Text
rtfcsSourceDatabaseName

-> Text
rtfcsSourceTableName

-> Text
rtfcsNewTableName

-> RestoreTableFromClusterSnapshot

Creates a value of RestoreTableFromClusterSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtfcsTargetSchemaName - The name of the schema to restore the table to.

rtfcsTargetDatabaseName - The name of the database to restore the table to.

rtfcsSourceSchemaName - The name of the source schema that contains the table to restore from. If you do not specify a SourceSchemaName value, the default is public .

rtfcsClusterIdentifier - The identifier of the Amazon Redshift cluster to restore the table to.

rtfcsSnapshotIdentifier - The identifier of the snapshot to restore the table from. This snapshot must have been created from the Amazon Redshift cluster specified by the ClusterIdentifier parameter.

rtfcsSourceDatabaseName - The name of the source database that contains the table to restore from.

rtfcsSourceTableName - The name of the source table to restore from.

rtfcsNewTableName - The name of the table to create as a result of the current request.

data RestoreTableFromClusterSnapshot

See: restoreTableFromClusterSnapshot smart constructor.

Request Lenses

rtfcsTargetSchemaName :: Lens' RestoreTableFromClusterSnapshot (Maybe Text)

The name of the schema to restore the table to.

rtfcsTargetDatabaseName :: Lens' RestoreTableFromClusterSnapshot (Maybe Text)

The name of the database to restore the table to.

rtfcsSourceSchemaName :: Lens' RestoreTableFromClusterSnapshot (Maybe Text)

The name of the source schema that contains the table to restore from. If you do not specify a SourceSchemaName value, the default is public .

rtfcsClusterIdentifier :: Lens' RestoreTableFromClusterSnapshot Text

The identifier of the Amazon Redshift cluster to restore the table to.

rtfcsSnapshotIdentifier :: Lens' RestoreTableFromClusterSnapshot Text

The identifier of the snapshot to restore the table from. This snapshot must have been created from the Amazon Redshift cluster specified by the ClusterIdentifier parameter.

rtfcsSourceDatabaseName :: Lens' RestoreTableFromClusterSnapshot Text

The name of the source database that contains the table to restore from.

rtfcsSourceTableName :: Lens' RestoreTableFromClusterSnapshot Text

The name of the source table to restore from.

rtfcsNewTableName :: Lens' RestoreTableFromClusterSnapshot Text

The name of the table to create as a result of the current request.

Destructuring the Response

restoreTableFromClusterSnapshotResponse

Arguments

:: Int
rtfcsrsResponseStatus

-> RestoreTableFromClusterSnapshotResponse

Creates a value of RestoreTableFromClusterSnapshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtfcsrsTableRestoreStatus - Undocumented member.

rtfcsrsResponseStatus - -- | The response status code.

data RestoreTableFromClusterSnapshotResponse

See: restoreTableFromClusterSnapshotResponse smart constructor.

Response Lenses

rtfcsrsTableRestoreStatus :: Lens' RestoreTableFromClusterSnapshotResponse (Maybe TableRestoreStatus)

Undocumented member.

rtfcsrsResponseStatus :: Lens' RestoreTableFromClusterSnapshotResponse Int

- | The response status code.