Network.AWS.Redshift.RevokeClusterSecurityGroupIngress

Description

Revokes an ingress rule in an Amazon Redshift security group for a previously authorized IP range or Amazon EC2 security group. To add an ingress rule, see AuthorizeClusterSecurityGroupIngress. For information about managing security groups, go to Amazon Redshift Cluster Security Groups in the Amazon Redshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_RevokeClusterSecurityGroupIngress.html

Request

data RevokeClusterSecurityGroupIngress

Request constructor

revokeClusterSecurityGroupIngress

Arguments

:: Text
rcsgiClusterSecurityGroupName

-> RevokeClusterSecurityGroupIngress

RevokeClusterSecurityGroupIngress constructor.

The fields accessible through corresponding lenses are:

rcsgiCIDRIP :: Maybe Text

rcsgiClusterSecurityGroupName :: Text

rcsgiEC2SecurityGroupName :: Maybe Text

rcsgiEC2SecurityGroupOwnerId :: Maybe Text

Request lenses

rcsgiCIDRIP :: Lens' RevokeClusterSecurityGroupIngress (Maybe Text)

The IP range for which to revoke access. This range must be a valid Classless Inter-Domain Routing (CIDR) block of IP addresses. If CIDRIP is specified, EC2SecurityGroupName and EC2SecurityGroupOwnerId cannot be provided.

rcsgiClusterSecurityGroupName :: Lens' RevokeClusterSecurityGroupIngress Text

The name of the security Group from which to revoke the ingress rule.

rcsgiEC2SecurityGroupName :: Lens' RevokeClusterSecurityGroupIngress (Maybe Text)

The name of the EC2 Security Group whose access is to be revoked. If EC2SecurityGroupName is specified, EC2SecurityGroupOwnerId must also be provided and CIDRIP cannot be provided.

rcsgiEC2SecurityGroupOwnerId :: Lens' RevokeClusterSecurityGroupIngress (Maybe Text)

The AWS account number of the owner of the security group specified in the EC2SecurityGroupName parameter. The AWS access key ID is not an acceptable value. If EC2SecurityGroupOwnerId is specified, EC2SecurityGroupName must also be provided. and CIDRIP cannot be provided.

Example: '111122223333'

Response

data RevokeClusterSecurityGroupIngressResponse

Response constructor

revokeClusterSecurityGroupIngressResponse :: RevokeClusterSecurityGroupIngressResponse

RevokeClusterSecurityGroupIngressResponse constructor.

The fields accessible through corresponding lenses are:

rcsgirClusterSecurityGroup :: Maybe ClusterSecurityGroup

Response lenses

rcsgirClusterSecurityGroup :: Lens' RevokeClusterSecurityGroupIngressResponse (Maybe ClusterSecurityGroup)