Network.AWS.Redshift.RevokeSnapshotAccess

Description

Removes the ability of the specified AWS customer account to restore the specified snapshot. If the account is currently restoring the snapshot, the restore will run to completion.

For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Cluster Management Guide .

Creating a Request

revokeSnapshotAccess

Arguments

:: Text
rsaSnapshotIdentifier

-> Text
rsaAccountWithRestoreAccess

-> RevokeSnapshotAccess

Creates a value of RevokeSnapshotAccess with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsaSnapshotClusterIdentifier - The identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

rsaSnapshotIdentifier - The identifier of the snapshot that the account can no longer access.

rsaAccountWithRestoreAccess - The identifier of the AWS customer account that can no longer restore the specified snapshot.

data RevokeSnapshotAccess

See: revokeSnapshotAccess smart constructor.

Request Lenses

rsaSnapshotClusterIdentifier :: Lens' RevokeSnapshotAccess (Maybe Text)

The identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

rsaSnapshotIdentifier :: Lens' RevokeSnapshotAccess Text

The identifier of the snapshot that the account can no longer access.

rsaAccountWithRestoreAccess :: Lens' RevokeSnapshotAccess Text

The identifier of the AWS customer account that can no longer restore the specified snapshot.

Destructuring the Response

revokeSnapshotAccessResponse

Arguments

:: Int
rsarsResponseStatus

-> RevokeSnapshotAccessResponse

Creates a value of RevokeSnapshotAccessResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsarsSnapshot - Undocumented member.

rsarsResponseStatus - -- | The response status code.

data RevokeSnapshotAccessResponse

See: revokeSnapshotAccessResponse smart constructor.

Response Lenses

rsarsSnapshot :: Lens' RevokeSnapshotAccessResponse (Maybe Snapshot)

Undocumented member.

rsarsResponseStatus :: Lens' RevokeSnapshotAccessResponse Int

- | The response status code.