Network.AWS.Redshift.RotateEncryptionKey

Description

Rotates the encryption keys for a cluster.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_RotateEncryptionKey.html

Request

data RotateEncryptionKey

Request constructor

rotateEncryptionKey

Arguments

:: Text
rekClusterIdentifier

-> RotateEncryptionKey

RotateEncryptionKey constructor.

The fields accessible through corresponding lenses are:

rekClusterIdentifier :: Text

Request lenses

rekClusterIdentifier :: Lens' RotateEncryptionKey Text

The unique identifier of the cluster that you want to rotate the encryption keys for.

Constraints: Must be the name of valid cluster that has encryption enabled.

Response

data RotateEncryptionKeyResponse

Response constructor

rotateEncryptionKeyResponse :: RotateEncryptionKeyResponse

RotateEncryptionKeyResponse constructor.

The fields accessible through corresponding lenses are:

rekrCluster :: Maybe Cluster

Response lenses

rekrCluster :: Lens' RotateEncryptionKeyResponse (Maybe Cluster)