Network.AWS.Redshift.Types

Service

data Redshift

Version 2012-12-01 of the Amazon Redshift service.

Error

data RESTError :: *

XML

ns :: Text

Snapshot

data Snapshot

snapshot :: Snapshot

Snapshot constructor.

The fields accessible through corresponding lenses are:

sAccountsWithRestoreAccess :: [AccountWithRestoreAccess]

sActualIncrementalBackupSizeInMegaBytes :: Maybe Double

sAvailabilityZone :: Maybe Text

sBackupProgressInMegaBytes :: Maybe Double

sClusterCreateTime :: Maybe UTCTime

sClusterIdentifier :: Maybe Text

sClusterVersion :: Maybe Text

sCurrentBackupRateInMegaBytesPerSecond :: Maybe Double

sDBName :: Maybe Text

sElapsedTimeInSeconds :: Maybe Integer

sEncrypted :: Maybe Bool

sEncryptedWithHSM :: Maybe Bool

sEstimatedSecondsToCompletion :: Maybe Integer

sKmsKeyId :: Maybe Text

sMasterUsername :: Maybe Text

sNodeType :: Maybe Text

sNumberOfNodes :: Maybe Int

sOwnerAccount :: Maybe Text

sPort :: Maybe Int

sSnapshotCreateTime :: Maybe UTCTime

sSnapshotIdentifier :: Maybe Text

sSnapshotType :: Maybe Text

sSourceRegion :: Maybe Text

sStatus :: Maybe Text

sTags :: [Tag]

sTotalBackupSizeInMegaBytes :: Maybe Double

sVpcId :: Maybe Text

sAccountsWithRestoreAccess :: Lens' Snapshot [AccountWithRestoreAccess]

A list of the AWS customer accounts authorized to restore the snapshot. Returns null if no accounts are authorized. Visible only to the snapshot owner.

sActualIncrementalBackupSizeInMegaBytes :: Lens' Snapshot (Maybe Double)

The size of the incremental backup.

sAvailabilityZone :: Lens' Snapshot (Maybe Text)

The Availability Zone in which the cluster was created.

sBackupProgressInMegaBytes :: Lens' Snapshot (Maybe Double)

The number of megabytes that have been transferred to the snapshot backup.

sClusterCreateTime :: Lens' Snapshot (Maybe UTCTime)

The time (UTC) when the cluster was originally created.

sClusterIdentifier :: Lens' Snapshot (Maybe Text)

The identifier of the cluster for which the snapshot was taken.

sClusterVersion :: Lens' Snapshot (Maybe Text)

The version ID of the Amazon Redshift engine that is running on the cluster.

sCurrentBackupRateInMegaBytesPerSecond :: Lens' Snapshot (Maybe Double)

The number of megabytes per second being transferred to the snapshot backup. Returns '0' for a completed backup.

sDBName :: Lens' Snapshot (Maybe Text)

The name of the database that was created when the cluster was created.

sElapsedTimeInSeconds :: Lens' Snapshot (Maybe Integer)

The amount of time an in-progress snapshot backup has been running, or the amount of time it took a completed backup to finish.

sEncrypted :: Lens' Snapshot (Maybe Bool)

If true, the data in the snapshot is encrypted at rest.

sEncryptedWithHSM :: Lens' Snapshot (Maybe Bool)

A boolean that indicates whether the snapshot data is encrypted using the HSM keys of the source cluster. true indicates that the data is encrypted using HSM keys.

sEstimatedSecondsToCompletion :: Lens' Snapshot (Maybe Integer)

The estimate of the time remaining before the snapshot backup will complete. Returns '0' for a completed backup.

sKmsKeyId :: Lens' Snapshot (Maybe Text)

The AWS Key Management Service (KMS) key ID of the encryption key that was used to encrypt data in the cluster from which the snapshot was taken.

sMasterUsername :: Lens' Snapshot (Maybe Text)

The master user name for the cluster.

sNodeType :: Lens' Snapshot (Maybe Text)

The node type of the nodes in the cluster.

sNumberOfNodes :: Lens' Snapshot (Maybe Int)

The number of nodes in the cluster.

sOwnerAccount :: Lens' Snapshot (Maybe Text)

For manual snapshots, the AWS customer account used to create or copy the snapshot. For automatic snapshots, the owner of the cluster. The owner can perform all snapshot actions, such as sharing a manual snapshot.

sPort :: Lens' Snapshot (Maybe Int)

The port that the cluster is listening on.

sSnapshotCreateTime :: Lens' Snapshot (Maybe UTCTime)

The time (UTC) when Amazon Redshift began the snapshot. A snapshot contains a copy of the cluster data as of this exact time.

sSnapshotIdentifier :: Lens' Snapshot (Maybe Text)

The snapshot identifier that is provided in the request.

sSnapshotType :: Lens' Snapshot (Maybe Text)

The snapshot type. Snapshots created using CreateClusterSnapshot and CopyClusterSnapshot will be of type "manual".

sSourceRegion :: Lens' Snapshot (Maybe Text)

The source region from which the snapshot was copied.

sStatus :: Lens' Snapshot (Maybe Text)

The snapshot status. The value of the status depends on the API operation used. CreateClusterSnapshot and CopyClusterSnapshot returns status as "creating". DescribeClusterSnapshots returns status as "creating", "available", "final snapshot", or "failed". DeleteClusterSnapshot returns status as "deleted".

sTags :: Lens' Snapshot [Tag]

The list of tags for the cluster snapshot.

sTotalBackupSizeInMegaBytes :: Lens' Snapshot (Maybe Double)

The size of the complete set of backup data that would be used to restore the cluster.

sVpcId :: Lens' Snapshot (Maybe Text)

The VPC identifier of the cluster if the snapshot is from a cluster in a VPC. Otherwise, this field is not in the output.

ClusterParameterGroup

data ClusterParameterGroup

clusterParameterGroup :: ClusterParameterGroup

ClusterParameterGroup constructor.

The fields accessible through corresponding lenses are:

cpgDescription :: Maybe Text

cpgParameterGroupFamily :: Maybe Text

cpgParameterGroupName :: Maybe Text

cpgTags :: [Tag]

cpgDescription :: Lens' ClusterParameterGroup (Maybe Text)

The description of the parameter group.

cpgParameterGroupFamily :: Lens' ClusterParameterGroup (Maybe Text)

The name of the cluster parameter group family that this cluster parameter group is compatible with.

cpgParameterGroupName :: Lens' ClusterParameterGroup (Maybe Text)

The name of the cluster parameter group.

cpgTags :: Lens' ClusterParameterGroup [Tag]

The list of tags for the cluster parameter group.

RestoreStatus

data RestoreStatus

restoreStatus :: RestoreStatus

RestoreStatus constructor.

The fields accessible through corresponding lenses are:

rsCurrentRestoreRateInMegaBytesPerSecond :: Maybe Double

rsElapsedTimeInSeconds :: Maybe Integer

rsEstimatedTimeToCompletionInSeconds :: Maybe Integer

rsProgressInMegaBytes :: Maybe Integer

rsSnapshotSizeInMegaBytes :: Maybe Integer

rsStatus :: Maybe Text

rsCurrentRestoreRateInMegaBytesPerSecond :: Lens' RestoreStatus (Maybe Double)

The number of megabytes per second being transferred from the backup storage. Returns the average rate for a completed backup.

rsElapsedTimeInSeconds :: Lens' RestoreStatus (Maybe Integer)

The amount of time an in-progress restore has been running, or the amount of time it took a completed restore to finish.

rsEstimatedTimeToCompletionInSeconds :: Lens' RestoreStatus (Maybe Integer)

The estimate of the time remaining before the restore will complete. Returns 0 for a completed restore.

rsProgressInMegaBytes :: Lens' RestoreStatus (Maybe Integer)

The number of megabytes that have been transferred from snapshot storage.

rsSnapshotSizeInMegaBytes :: Lens' RestoreStatus (Maybe Integer)

The size of the set of snapshot data used to restore the cluster.

rsStatus :: Lens' RestoreStatus (Maybe Text)

The status of the restore action. Returns starting, restoring, completed, or failed.

Event

data Event

event :: Event

Event constructor.

The fields accessible through corresponding lenses are:

eDate :: Maybe UTCTime

eEventCategories :: [Text]

eEventId :: Maybe Text

eMessage :: Maybe Text

eSeverity :: Maybe Text

eSourceIdentifier :: Maybe Text

eSourceType :: Maybe SourceType

eDate :: Lens' Event (Maybe UTCTime)

The date and time of the event.

eEventCategories :: Lens' Event [Text]

A list of the event categories.

eEventId :: Lens' Event (Maybe Text)

The identifier of the event.

eMessage :: Lens' Event (Maybe Text)

The text of this event.

eSeverity :: Lens' Event (Maybe Text)

The severity of the event.

Values: ERROR, INFO

eSourceIdentifier :: Lens' Event (Maybe Text)

The identifier for the source of the event.

eSourceType :: Lens' Event (Maybe SourceType)

The source type for this event.

ClusterSnapshotCopyStatus

data ClusterSnapshotCopyStatus

clusterSnapshotCopyStatus :: ClusterSnapshotCopyStatus

ClusterSnapshotCopyStatus constructor.

The fields accessible through corresponding lenses are:

cscsDestinationRegion :: Maybe Text

cscsRetentionPeriod :: Maybe Integer

cscsDestinationRegion :: Lens' ClusterSnapshotCopyStatus (Maybe Text)

The destination region that snapshots are automatically copied to when cross-region snapshot copy is enabled.

cscsRetentionPeriod :: Lens' ClusterSnapshotCopyStatus (Maybe Integer)

The number of days that automated snapshots are retained in the destination region after they are copied from a source region.

Tag

data Tag

tag :: Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Maybe Text

tagValue :: Maybe Text

tagKey :: Lens' Tag (Maybe Text)

The key, or name, for the resource tag.

tagValue :: Lens' Tag (Maybe Text)

The value for the resource tag.

HsmClientCertificate

data HsmClientCertificate

hsmClientCertificate :: HsmClientCertificate

HsmClientCertificate constructor.

The fields accessible through corresponding lenses are:

hccHsmClientCertificateIdentifier :: Maybe Text

hccHsmClientCertificatePublicKey :: Maybe Text

hccTags :: [Tag]

hccHsmClientCertificateIdentifier :: Lens' HsmClientCertificate (Maybe Text)

The identifier of the HSM client certificate.

hccHsmClientCertificatePublicKey :: Lens' HsmClientCertificate (Maybe Text)

The public key that the Amazon Redshift cluster will use to connect to the HSM. You must register the public key in the HSM.

hccTags :: Lens' HsmClientCertificate [Tag]

The list of tags for the HSM client certificate.

Cluster

data Cluster

cluster :: Cluster

Cluster constructor.

The fields accessible through corresponding lenses are:

cAllowVersionUpgrade :: Maybe Bool

cAutomatedSnapshotRetentionPeriod :: Maybe Int

cAvailabilityZone :: Maybe Text

cClusterCreateTime :: Maybe UTCTime

cClusterIdentifier :: Maybe Text

cClusterNodes :: [ClusterNode]

cClusterParameterGroups :: [ClusterParameterGroupStatus]

cClusterPublicKey :: Maybe Text

cClusterRevisionNumber :: Maybe Text

cClusterSecurityGroups :: [ClusterSecurityGroupMembership]

cClusterSnapshotCopyStatus :: Maybe ClusterSnapshotCopyStatus

cClusterStatus :: Maybe Text

cClusterSubnetGroupName :: Maybe Text

cClusterVersion :: Maybe Text

cDBName :: Maybe Text

cElasticIpStatus :: Maybe ElasticIpStatus

cEncrypted :: Maybe Bool

cEndpoint :: Maybe Endpoint

cHsmStatus :: Maybe HsmStatus

cKmsKeyId :: Maybe Text

cMasterUsername :: Maybe Text

cModifyStatus :: Maybe Text

cNodeType :: Maybe Text

cNumberOfNodes :: Maybe Int

cPendingModifiedValues :: Maybe PendingModifiedValues

cPreferredMaintenanceWindow :: Maybe Text

cPubliclyAccessible :: Maybe Bool

cRestoreStatus :: Maybe RestoreStatus

cTags :: [Tag]

cVpcId :: Maybe Text

cVpcSecurityGroups :: [VpcSecurityGroupMembership]

cAllowVersionUpgrade :: Lens' Cluster (Maybe Bool)

If true, major version upgrades will be applied automatically to the cluster during the maintenance window.

cAutomatedSnapshotRetentionPeriod :: Lens' Cluster (Maybe Int)

The number of days that automatic cluster snapshots are retained.

cAvailabilityZone :: Lens' Cluster (Maybe Text)

The name of the Availability Zone in which the cluster is located.

cClusterCreateTime :: Lens' Cluster (Maybe UTCTime)

The date and time that the cluster was created.

cClusterIdentifier :: Lens' Cluster (Maybe Text)

The unique identifier of the cluster.

cClusterNodes :: Lens' Cluster [ClusterNode]

The nodes in a cluster.

cClusterParameterGroups :: Lens' Cluster [ClusterParameterGroupStatus]

The list of cluster parameter groups that are associated with this cluster.

cClusterPublicKey :: Lens' Cluster (Maybe Text)

The public key for the cluster.

cClusterRevisionNumber :: Lens' Cluster (Maybe Text)

The specific revision number of the database in the cluster.

cClusterSecurityGroups :: Lens' Cluster [ClusterSecurityGroupMembership]

A list of cluster security group that are associated with the cluster. Each security group is represented by an element that contains Name and Status subelements.

Cluster security groups are used when the cluster is not created in a VPC. Clusters that are created in a VPC use VPC security groups, which are listed by the VpcSecurityGroups parameter.

cClusterSnapshotCopyStatus :: Lens' Cluster (Maybe ClusterSnapshotCopyStatus)

Returns the destination region and retention period that are configured for cross-region snapshot copy.

cClusterStatus :: Lens' Cluster (Maybe Text)

The current state of this cluster. Possible values include available, creating, deleting, rebooting, renaming, and resizing.

cClusterSubnetGroupName :: Lens' Cluster (Maybe Text)

The name of the subnet group that is associated with the cluster. This parameter is valid only when the cluster is in a VPC.

cClusterVersion :: Lens' Cluster (Maybe Text)

The version ID of the Amazon Redshift engine that is running on the cluster.

cDBName :: Lens' Cluster (Maybe Text)

The name of the initial database that was created when the cluster was created. This same name is returned for the life of the cluster. If an initial database was not specified, a database named "dev" was created by default.

cElasticIpStatus :: Lens' Cluster (Maybe ElasticIpStatus)

Describes the status of the elastic IP (EIP) address.

cEncrypted :: Lens' Cluster (Maybe Bool)

If true, data in the cluster is encrypted at rest.

cEndpoint :: Lens' Cluster (Maybe Endpoint)

The connection endpoint.

cHsmStatus :: Lens' Cluster (Maybe HsmStatus)

Reports whether the Amazon Redshift cluster has finished applying any HSM settings changes specified in a modify cluster command.

Values: active, applying

cKmsKeyId :: Lens' Cluster (Maybe Text)

The AWS Key Management Service (KMS) key ID of the encryption key used to encrypt data in the cluster.

cMasterUsername :: Lens' Cluster (Maybe Text)

The master user name for the cluster. This name is used to connect to the database that is specified in DBName.

cModifyStatus :: Lens' Cluster (Maybe Text)

The status of a modify operation, if any, initiated for the cluster.

cNodeType :: Lens' Cluster (Maybe Text)

The node type for the nodes in the cluster.

cNumberOfNodes :: Lens' Cluster (Maybe Int)

The number of compute nodes in the cluster.

cPendingModifiedValues :: Lens' Cluster (Maybe PendingModifiedValues)

If present, changes to the cluster are pending. Specific pending changes are identified by subelements.

cPreferredMaintenanceWindow :: Lens' Cluster (Maybe Text)

The weekly time range (in UTC) during which system maintenance can occur.

cPubliclyAccessible :: Lens' Cluster (Maybe Bool)

If true, the cluster can be accessed from a public network.

cRestoreStatus :: Lens' Cluster (Maybe RestoreStatus)

Describes the status of a cluster restore action. Returns null if the cluster was not created by restoring a snapshot.

cTags :: Lens' Cluster [Tag]

The list of tags for the cluster.

cVpcId :: Lens' Cluster (Maybe Text)

The identifier of the VPC the cluster is in, if the cluster is in a VPC.

cVpcSecurityGroups :: Lens' Cluster [VpcSecurityGroupMembership]

A list of Virtual Private Cloud (VPC) security groups that are associated with the cluster. This parameter is returned only if the cluster is in a VPC.

ClusterNode

data ClusterNode

clusterNode :: ClusterNode

ClusterNode constructor.

The fields accessible through corresponding lenses are:

cnNodeRole :: Maybe Text

cnPrivateIPAddress :: Maybe Text

cnPublicIPAddress :: Maybe Text

cnNodeRole :: Lens' ClusterNode (Maybe Text)

Whether the node is a leader node or a compute node.

cnPrivateIPAddress :: Lens' ClusterNode (Maybe Text)

The private IP address of a node within a cluster.

cnPublicIPAddress :: Lens' ClusterNode (Maybe Text)

The public IP address of a node within a cluster.

EC2SecurityGroup

data EC2SecurityGroup

ec2SecurityGroup :: EC2SecurityGroup

EC2SecurityGroup constructor.

The fields accessible through corresponding lenses are:

ecsgEC2SecurityGroupName :: Maybe Text

ecsgEC2SecurityGroupOwnerId :: Maybe Text

ecsgStatus :: Maybe Text

ecsgTags :: [Tag]

ecsgEC2SecurityGroupName :: Lens' EC2SecurityGroup (Maybe Text)

The name of the EC2 Security Group.

ecsgEC2SecurityGroupOwnerId :: Lens' EC2SecurityGroup (Maybe Text)

The AWS ID of the owner of the EC2 security group specified in the EC2SecurityGroupName field.

ecsgStatus :: Lens' EC2SecurityGroup (Maybe Text)

The status of the EC2 security group.

ecsgTags :: Lens' EC2SecurityGroup [Tag]

The list of tags for the EC2 security group.

OrderableClusterOption

data OrderableClusterOption

orderableClusterOption :: OrderableClusterOption

OrderableClusterOption constructor.

The fields accessible through corresponding lenses are:

ocoAvailabilityZones :: [AvailabilityZone]

ocoClusterType :: Maybe Text

ocoClusterVersion :: Maybe Text

ocoNodeType :: Maybe Text

ocoAvailabilityZones :: Lens' OrderableClusterOption [AvailabilityZone]

A list of availability zones for the orderable cluster.

ocoClusterType :: Lens' OrderableClusterOption (Maybe Text)

The cluster type, for example 'multi-node'.

ocoClusterVersion :: Lens' OrderableClusterOption (Maybe Text)

The version of the orderable cluster.

ocoNodeType :: Lens' OrderableClusterOption (Maybe Text)

The node type for the orderable cluster.

SourceType

data SourceType

Constructors

STCluster
cluster

STClusterParameterGroup
cluster-parameter-group

STClusterSecurityGroup
cluster-security-group

STClusterSnapshot
cluster-snapshot

ClusterParameterGroupStatus

data ClusterParameterGroupStatus

clusterParameterGroupStatus :: ClusterParameterGroupStatus

ClusterParameterGroupStatus constructor.

The fields accessible through corresponding lenses are:

cpgsParameterApplyStatus :: Maybe Text

cpgsParameterGroupName :: Maybe Text

cpgsParameterApplyStatus :: Lens' ClusterParameterGroupStatus (Maybe Text)

The status of parameter updates.

cpgsParameterGroupName :: Lens' ClusterParameterGroupStatus (Maybe Text)

The name of the cluster parameter group.

Subnet

data Subnet

subnet :: Subnet

Subnet constructor.

The fields accessible through corresponding lenses are:

sSubnetAvailabilityZone :: Maybe AvailabilityZone

sSubnetIdentifier :: Maybe Text

sSubnetStatus :: Maybe Text

sSubnetAvailabilityZone :: Lens' Subnet (Maybe AvailabilityZone)

sSubnetIdentifier :: Lens' Subnet (Maybe Text)

The identifier of the subnet.

sSubnetStatus :: Lens' Subnet (Maybe Text)

The status of the subnet.

ClusterSecurityGroup

data ClusterSecurityGroup

clusterSecurityGroup :: ClusterSecurityGroup

ClusterSecurityGroup constructor.

The fields accessible through corresponding lenses are:

csgClusterSecurityGroupName :: Maybe Text

csgDescription :: Maybe Text

csgEC2SecurityGroups :: [EC2SecurityGroup]

csgIPRanges :: [IPRange]

csgTags :: [Tag]

csgClusterSecurityGroupName :: Lens' ClusterSecurityGroup (Maybe Text)

The name of the cluster security group to which the operation was applied.

csgDescription :: Lens' ClusterSecurityGroup (Maybe Text)

A description of the security group.

csgEC2SecurityGroups :: Lens' ClusterSecurityGroup [EC2SecurityGroup]

A list of EC2 security groups that are permitted to access clusters associated with this cluster security group.

csgIPRanges :: Lens' ClusterSecurityGroup [IPRange]

A list of IP ranges (CIDR blocks) that are permitted to access clusters associated with this cluster security group.

csgTags :: Lens' ClusterSecurityGroup [Tag]

The list of tags for the cluster security group.

DefaultClusterParameters

data DefaultClusterParameters

defaultClusterParameters :: DefaultClusterParameters

DefaultClusterParameters constructor.

The fields accessible through corresponding lenses are:

dcpMarker :: Maybe Text

dcpParameterGroupFamily :: Maybe Text

dcpParameters :: [Parameter]

dcpMarker :: Lens' DefaultClusterParameters (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcpParameterGroupFamily :: Lens' DefaultClusterParameters (Maybe Text)

The name of the cluster parameter group family to which the engine default parameters apply.

dcpParameters :: Lens' DefaultClusterParameters [Parameter]

The list of cluster default parameters.

ClusterSubnetGroup

data ClusterSubnetGroup

clusterSubnetGroup :: ClusterSubnetGroup

ClusterSubnetGroup constructor.

The fields accessible through corresponding lenses are:

csg1ClusterSubnetGroupName :: Maybe Text

csg1Description :: Maybe Text

csg1SubnetGroupStatus :: Maybe Text

csg1Subnets :: [Subnet]

csg1Tags :: [Tag]

csg1VpcId :: Maybe Text

csg1ClusterSubnetGroupName :: Lens' ClusterSubnetGroup (Maybe Text)

The name of the cluster subnet group.

csg1Description :: Lens' ClusterSubnetGroup (Maybe Text)

The description of the cluster subnet group.

csg1SubnetGroupStatus :: Lens' ClusterSubnetGroup (Maybe Text)

The status of the cluster subnet group. Possible values are Complete, Incomplete and Invalid.

csg1Subnets :: Lens' ClusterSubnetGroup [Subnet]

A list of the VPC Subnet elements.

csg1Tags :: Lens' ClusterSubnetGroup [Tag]

The list of tags for the cluster subnet group.

csg1VpcId :: Lens' ClusterSubnetGroup (Maybe Text)

The VPC ID of the cluster subnet group.

EventInfoMap

data EventInfoMap

eventInfoMap :: EventInfoMap

EventInfoMap constructor.

The fields accessible through corresponding lenses are:

eimEventCategories :: [Text]

eimEventDescription :: Maybe Text

eimEventId :: Maybe Text

eimSeverity :: Maybe Text

eimEventCategories :: Lens' EventInfoMap [Text]

The category of an Amazon Redshift event.

eimEventDescription :: Lens' EventInfoMap (Maybe Text)

The description of an Amazon Redshift event.

eimEventId :: Lens' EventInfoMap (Maybe Text)

The identifier of an Amazon Redshift event.

eimSeverity :: Lens' EventInfoMap (Maybe Text)

The severity of the event.

Values: ERROR, INFO

ClusterSecurityGroupMembership

data ClusterSecurityGroupMembership

clusterSecurityGroupMembership :: ClusterSecurityGroupMembership

ClusterSecurityGroupMembership constructor.

The fields accessible through corresponding lenses are:

csgmClusterSecurityGroupName :: Maybe Text

csgmStatus :: Maybe Text

csgmClusterSecurityGroupName :: Lens' ClusterSecurityGroupMembership (Maybe Text)

The name of the cluster security group.

csgmStatus :: Lens' ClusterSecurityGroupMembership (Maybe Text)

The status of the cluster security group.

ReservedNodeOffering

data ReservedNodeOffering

reservedNodeOffering :: ReservedNodeOffering

ReservedNodeOffering constructor.

The fields accessible through corresponding lenses are:

rnoCurrencyCode :: Maybe Text

rnoDuration :: Maybe Int

rnoFixedPrice :: Maybe Double

rnoNodeType :: Maybe Text

rnoOfferingType :: Maybe Text

rnoRecurringCharges :: [RecurringCharge]

rnoReservedNodeOfferingId :: Maybe Text

rnoUsagePrice :: Maybe Double

rnoCurrencyCode :: Lens' ReservedNodeOffering (Maybe Text)

The currency code for the compute nodes offering.

rnoDuration :: Lens' ReservedNodeOffering (Maybe Int)

The duration, in seconds, for which the offering will reserve the node.

rnoFixedPrice :: Lens' ReservedNodeOffering (Maybe Double)

The upfront fixed charge you will pay to purchase the specific reserved node offering.

rnoNodeType :: Lens' ReservedNodeOffering (Maybe Text)

The node type offered by the reserved node offering.

rnoOfferingType :: Lens' ReservedNodeOffering (Maybe Text)

The anticipated utilization of the reserved node, as defined in the reserved node offering.

rnoRecurringCharges :: Lens' ReservedNodeOffering [RecurringCharge]

The charge to your account regardless of whether you are creating any clusters using the node offering. Recurring charges are only in effect for heavy-utilization reserved nodes.

rnoReservedNodeOfferingId :: Lens' ReservedNodeOffering (Maybe Text)

The offering identifier.

rnoUsagePrice :: Lens' ReservedNodeOffering (Maybe Double)

The rate you are charged for each hour the cluster that is using the offering is running.

ReservedNode

data ReservedNode

reservedNode :: ReservedNode

ReservedNode constructor.

The fields accessible through corresponding lenses are:

rnCurrencyCode :: Maybe Text

rnDuration :: Maybe Int

rnFixedPrice :: Maybe Double

rnNodeCount :: Maybe Int

rnNodeType :: Maybe Text

rnOfferingType :: Maybe Text

rnRecurringCharges :: [RecurringCharge]

rnReservedNodeId :: Maybe Text

rnReservedNodeOfferingId :: Maybe Text

rnStartTime :: Maybe UTCTime

rnState :: Maybe Text

rnUsagePrice :: Maybe Double

rnCurrencyCode :: Lens' ReservedNode (Maybe Text)

The currency code for the reserved cluster.

rnDuration :: Lens' ReservedNode (Maybe Int)

The duration of the node reservation in seconds.

rnFixedPrice :: Lens' ReservedNode (Maybe Double)

The fixed cost Amazon Redshift charged you for this reserved node.

rnNodeCount :: Lens' ReservedNode (Maybe Int)

The number of reserved compute nodes.

rnNodeType :: Lens' ReservedNode (Maybe Text)

The node type of the reserved node.

rnOfferingType :: Lens' ReservedNode (Maybe Text)

The anticipated utilization of the reserved node, as defined in the reserved node offering.

rnRecurringCharges :: Lens' ReservedNode [RecurringCharge]

The recurring charges for the reserved node.

rnReservedNodeId :: Lens' ReservedNode (Maybe Text)

The unique identifier for the reservation.

rnReservedNodeOfferingId :: Lens' ReservedNode (Maybe Text)

The identifier for the reserved node offering.

rnStartTime :: Lens' ReservedNode (Maybe UTCTime)

The time the reservation started. You purchase a reserved node offering for a duration. This is the start time of that duration.

rnState :: Lens' ReservedNode (Maybe Text)

The state of the reserved compute node.

Possible Values:

pending-payment-This reserved node has recently been purchased, and the sale has been approved, but payment has not yet been confirmed. active-This reserved node is owned by the caller and is available for use. payment-failed-Payment failed for the purchase attempt.

rnUsagePrice :: Lens' ReservedNode (Maybe Double)

The hourly rate Amazon Redshift charge you for this reserved node.

LoggingStatus

data LoggingStatus

loggingStatus :: LoggingStatus

LoggingStatus constructor.

The fields accessible through corresponding lenses are:

lsBucketName :: Maybe Text

lsLastFailureMessage :: Maybe Text

lsLastFailureTime :: Maybe UTCTime

lsLastSuccessfulDeliveryTime :: Maybe UTCTime

lsLoggingEnabled :: Maybe Bool

lsS3KeyPrefix :: Maybe Text

lsBucketName :: Lens' LoggingStatus (Maybe Text)

The name of the S3 bucket where the log files are stored.

lsLastFailureMessage :: Lens' LoggingStatus (Maybe Text)

The message indicating that logs failed to be delivered.

lsLastFailureTime :: Lens' LoggingStatus (Maybe UTCTime)

The last time when logs failed to be delivered.

lsLastSuccessfulDeliveryTime :: Lens' LoggingStatus (Maybe UTCTime)

The last time when logs were delivered.

lsLoggingEnabled :: Lens' LoggingStatus (Maybe Bool)

true if logging is on, false if logging is off.

lsS3KeyPrefix :: Lens' LoggingStatus (Maybe Text)

The prefix applied to the log file names.

AccountWithRestoreAccess

data AccountWithRestoreAccess

accountWithRestoreAccess :: AccountWithRestoreAccess

AccountWithRestoreAccess constructor.

The fields accessible through corresponding lenses are:

awraAccountId :: Maybe Text

awraAccountId :: Lens' AccountWithRestoreAccess (Maybe Text)

The identifier of an AWS customer account authorized to restore a snapshot.

AvailabilityZone

data AvailabilityZone

availabilityZone :: AvailabilityZone

AvailabilityZone constructor.

The fields accessible through corresponding lenses are:

azName :: Maybe Text

azName :: Lens' AvailabilityZone (Maybe Text)

The name of the availability zone.

EventSubscription

data EventSubscription

eventSubscription :: EventSubscription

EventSubscription constructor.

The fields accessible through corresponding lenses are:

esCustSubscriptionId :: Maybe Text

esCustomerAwsId :: Maybe Text

esEnabled :: Maybe Bool

esEventCategoriesList :: [Text]

esSeverity :: Maybe Text

esSnsTopicArn :: Maybe Text

esSourceIdsList :: [Text]

esSourceType :: Maybe Text

esStatus :: Maybe Text

esSubscriptionCreationTime :: Maybe UTCTime

esTags :: [Tag]

esCustSubscriptionId :: Lens' EventSubscription (Maybe Text)

The name of the Amazon Redshift event notification subscription.

esCustomerAwsId :: Lens' EventSubscription (Maybe Text)

The AWS customer account associated with the Amazon Redshift event notification subscription.

esEnabled :: Lens' EventSubscription (Maybe Bool)

A Boolean value indicating whether the subscription is enabled. true indicates the subscription is enabled.

esEventCategoriesList :: Lens' EventSubscription [Text]

The list of Amazon Redshift event categories specified in the event notification subscription.

Values: Configuration, Management, Monitoring, Security

esSeverity :: Lens' EventSubscription (Maybe Text)

The event severity specified in the Amazon Redshift event notification subscription.

Values: ERROR, INFO

esSnsTopicArn :: Lens' EventSubscription (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic used by the event notification subscription.

esSourceIdsList :: Lens' EventSubscription [Text]

A list of the sources that publish events to the Amazon Redshift event notification subscription.

esSourceType :: Lens' EventSubscription (Maybe Text)

The source type of the events returned the Amazon Redshift event notification, such as cluster, or cluster-snapshot.

esStatus :: Lens' EventSubscription (Maybe Text)

The status of the Amazon Redshift event notification subscription.

Constraints:

Can be one of the following: active | no-permission | topic-not-exist The status "no-permission" indicates that Amazon Redshift no longer has permission to post to the Amazon SNS topic. The status "topic-not-exist" indicates that the topic was deleted after the subscription was created.

esSubscriptionCreationTime :: Lens' EventSubscription (Maybe UTCTime)

The date and time the Amazon Redshift event notification subscription was created.

esTags :: Lens' EventSubscription [Tag]

The list of tags for the event subscription.

HsmStatus

data HsmStatus

hsmStatus :: HsmStatus

HsmStatus constructor.

The fields accessible through corresponding lenses are:

hsHsmClientCertificateIdentifier :: Maybe Text

hsHsmConfigurationIdentifier :: Maybe Text

hsStatus :: Maybe Text

hsHsmClientCertificateIdentifier :: Lens' HsmStatus (Maybe Text)

Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

hsHsmConfigurationIdentifier :: Lens' HsmStatus (Maybe Text)

Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

hsStatus :: Lens' HsmStatus (Maybe Text)

Reports whether the Amazon Redshift cluster has finished applying any HSM settings changes specified in a modify cluster command.

Values: active, applying

ClusterParameterGroupNameMessage

data ClusterParameterGroupNameMessage

clusterParameterGroupNameMessage :: ClusterParameterGroupNameMessage

ClusterParameterGroupNameMessage constructor.

The fields accessible through corresponding lenses are:

cpgnmParameterGroupName :: Maybe Text

cpgnmParameterGroupStatus :: Maybe Text

cpgnmParameterGroupName :: Lens' ClusterParameterGroupNameMessage (Maybe Text)

The name of the cluster parameter group.

cpgnmParameterGroupStatus :: Lens' ClusterParameterGroupNameMessage (Maybe Text)

The status of the parameter group. For example, if you made a change to a parameter group name-value pair, then the change could be pending a reboot of an associated cluster.

ElasticIpStatus

data ElasticIpStatus

elasticIpStatus :: ElasticIpStatus

ElasticIpStatus constructor.

The fields accessible through corresponding lenses are:

eisElasticIp :: Maybe Text

eisStatus :: Maybe Text

eisElasticIp :: Lens' ElasticIpStatus (Maybe Text)

The elastic IP (EIP) address for the cluster.

eisStatus :: Lens' ElasticIpStatus (Maybe Text)

Describes the status of the elastic IP (EIP) address.

ClusterVersion

data ClusterVersion

clusterVersion :: ClusterVersion

ClusterVersion constructor.

The fields accessible through corresponding lenses are:

cvClusterParameterGroupFamily :: Maybe Text

cvClusterVersion :: Maybe Text

cvDescription :: Maybe Text

cvClusterParameterGroupFamily :: Lens' ClusterVersion (Maybe Text)

The name of the cluster parameter group family for the cluster.

cvClusterVersion :: Lens' ClusterVersion (Maybe Text)

The version number used by the cluster.

cvDescription :: Lens' ClusterVersion (Maybe Text)

The description of the cluster version.

RecurringCharge

data RecurringCharge

recurringCharge :: RecurringCharge

RecurringCharge constructor.

The fields accessible through corresponding lenses are:

rcRecurringChargeAmount :: Maybe Double

rcRecurringChargeFrequency :: Maybe Text

rcRecurringChargeAmount :: Lens' RecurringCharge (Maybe Double)

The amount charged per the period of time specified by the recurring charge frequency.

rcRecurringChargeFrequency :: Lens' RecurringCharge (Maybe Text)

The frequency at which the recurring charge amount is applied.

Endpoint

data Endpoint

endpoint :: Endpoint

Endpoint constructor.

The fields accessible through corresponding lenses are:

eAddress :: Maybe Text

ePort :: Maybe Int

eAddress :: Lens' Endpoint (Maybe Text)

The DNS address of the Cluster.

ePort :: Lens' Endpoint (Maybe Int)

The port that the database engine is listening on.

IPRange

data IPRange

iprange :: IPRange

IPRange constructor.

The fields accessible through corresponding lenses are:

iprCIDRIP :: Maybe Text

iprStatus :: Maybe Text

iprTags :: [Tag]

iprCIDRIP :: Lens' IPRange (Maybe Text)

The IP range in Classless Inter-Domain Routing (CIDR) notation.

iprStatus :: Lens' IPRange (Maybe Text)

The status of the IP range, for example, "authorized".

iprTags :: Lens' IPRange [Tag]

The list of tags for the IP range.

TaggedResource

data TaggedResource

taggedResource :: TaggedResource

TaggedResource constructor.

The fields accessible through corresponding lenses are:

trResourceName :: Maybe Text

trResourceType :: Maybe Text

trTag :: Maybe Tag

trResourceName :: Lens' TaggedResource (Maybe Text)

The Amazon Resource Name (ARN) with which the tag is associated. For example, 'arn:aws:redshift:us-east-1:123456789:cluster:t1'.

trResourceType :: Lens' TaggedResource (Maybe Text)

The type of resource with which the tag is associated. Valid resource types are: Cluster CIDR/IP EC2 security group Snapshot Cluster security group Subnet group HSM connection HSM certificate Parameter group

For more information about Amazon Redshift resource types and constructing ARNs, go to Constructing an Amazon Redshift Amazon Resource Name (ARN) in the Amazon Redshift Cluster Management Guide.

trTag :: Lens' TaggedResource (Maybe Tag)

The tag for the resource.

EventCategoriesMap

data EventCategoriesMap

eventCategoriesMap :: EventCategoriesMap

EventCategoriesMap constructor.

The fields accessible through corresponding lenses are:

ecmEvents :: [EventInfoMap]

ecmSourceType :: Maybe Text

ecmEvents :: Lens' EventCategoriesMap [EventInfoMap]

The events in the event category.

ecmSourceType :: Lens' EventCategoriesMap (Maybe Text)

The Amazon Redshift source type, such as cluster or cluster-snapshot, that the returned categories belong to.

HsmConfiguration

data HsmConfiguration

hsmConfiguration :: HsmConfiguration

HsmConfiguration constructor.

The fields accessible through corresponding lenses are:

hcDescription :: Maybe Text

hcHsmConfigurationIdentifier :: Maybe Text

hcHsmIpAddress :: Maybe Text

hcHsmPartitionName :: Maybe Text

hcTags :: [Tag]

hcDescription :: Lens' HsmConfiguration (Maybe Text)

A text description of the HSM configuration.

hcHsmConfigurationIdentifier :: Lens' HsmConfiguration (Maybe Text)

The name of the Amazon Redshift HSM configuration.

hcHsmIpAddress :: Lens' HsmConfiguration (Maybe Text)

The IP address that the Amazon Redshift cluster must use to access the HSM.

hcHsmPartitionName :: Lens' HsmConfiguration (Maybe Text)

The name of the partition in the HSM where the Amazon Redshift clusters will store their database encryption keys.

hcTags :: Lens' HsmConfiguration [Tag]

The list of tags for the HSM configuration.

PendingModifiedValues

data PendingModifiedValues

pendingModifiedValues :: PendingModifiedValues

PendingModifiedValues constructor.

The fields accessible through corresponding lenses are:

pmvAutomatedSnapshotRetentionPeriod :: Maybe Int

pmvClusterIdentifier :: Maybe Text

pmvClusterType :: Maybe Text

pmvClusterVersion :: Maybe Text

pmvMasterUserPassword :: Maybe Text

pmvNodeType :: Maybe Text

pmvNumberOfNodes :: Maybe Int

pmvAutomatedSnapshotRetentionPeriod :: Lens' PendingModifiedValues (Maybe Int)

The pending or in-progress change of the automated snapshot retention period.

pmvClusterIdentifier :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the new identifier for the cluster.

pmvClusterType :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the cluster type.

pmvClusterVersion :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the service version.

pmvMasterUserPassword :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the master user password for the cluster.

pmvNodeType :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the cluster's node type.

pmvNumberOfNodes :: Lens' PendingModifiedValues (Maybe Int)

The pending or in-progress change of the number of nodes in the cluster.

VpcSecurityGroupMembership

data VpcSecurityGroupMembership

vpcSecurityGroupMembership :: VpcSecurityGroupMembership

VpcSecurityGroupMembership constructor.

The fields accessible through corresponding lenses are:

vsgmStatus :: Maybe Text

vsgmVpcSecurityGroupId :: Maybe Text

vsgmStatus :: Lens' VpcSecurityGroupMembership (Maybe Text)

vsgmVpcSecurityGroupId :: Lens' VpcSecurityGroupMembership (Maybe Text)

Parameter

data Parameter

parameter :: Parameter

Parameter constructor.

The fields accessible through corresponding lenses are:

pAllowedValues :: Maybe Text

pDataType :: Maybe Text

pDescription :: Maybe Text

pIsModifiable :: Maybe Bool

pMinimumEngineVersion :: Maybe Text

pParameterName :: Maybe Text

pParameterValue :: Maybe Text

pSource :: Maybe Text

pAllowedValues :: Lens' Parameter (Maybe Text)

The valid range of values for the parameter.

pDataType :: Lens' Parameter (Maybe Text)

The data type of the parameter.

pDescription :: Lens' Parameter (Maybe Text)

A description of the parameter.

pIsModifiable :: Lens' Parameter (Maybe Bool)

If true, the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

pMinimumEngineVersion :: Lens' Parameter (Maybe Text)

The earliest engine version to which the parameter can apply.

pParameterName :: Lens' Parameter (Maybe Text)

The name of the parameter.

pParameterValue :: Lens' Parameter (Maybe Text)

The value of the parameter.

pSource :: Lens' Parameter (Maybe Text)

The source of the parameter value, such as "engine-default" or "user".