Network.AWS.Redshift.V2012_12_01

Description

Amazon Redshift is a fast, fully managed, petabyte-scale data warehouse service that makes it simple and cost-effective to efficiently analyze all your data using your existing business intelligence tools. You can start small for just $0.25 per hour with no commitments or upfront costs and scale to a petabyte or more for $1,000 per terabyte per year, less than a tenth of most other data warehousing solutions.

Documentation

module Network.AWS.Redshift.V2012_12_01.AuthorizeClusterSecurityGroupIngress

module Network.AWS.Redshift.V2012_12_01.AuthorizeSnapshotAccess

module Network.AWS.Redshift.V2012_12_01.CopyClusterSnapshot

module Network.AWS.Redshift.V2012_12_01.CreateCluster

module Network.AWS.Redshift.V2012_12_01.CreateClusterParameterGroup

module Network.AWS.Redshift.V2012_12_01.CreateClusterSecurityGroup

module Network.AWS.Redshift.V2012_12_01.CreateClusterSnapshot

module Network.AWS.Redshift.V2012_12_01.CreateClusterSubnetGroup

module Network.AWS.Redshift.V2012_12_01.CreateEventSubscription

module Network.AWS.Redshift.V2012_12_01.CreateHsmClientCertificate

module Network.AWS.Redshift.V2012_12_01.CreateHsmConfiguration

module Network.AWS.Redshift.V2012_12_01.DeleteCluster

module Network.AWS.Redshift.V2012_12_01.DeleteClusterParameterGroup

module Network.AWS.Redshift.V2012_12_01.DeleteClusterSecurityGroup

module Network.AWS.Redshift.V2012_12_01.DeleteClusterSnapshot

module Network.AWS.Redshift.V2012_12_01.DeleteClusterSubnetGroup

module Network.AWS.Redshift.V2012_12_01.DeleteEventSubscription

module Network.AWS.Redshift.V2012_12_01.DeleteHsmClientCertificate

module Network.AWS.Redshift.V2012_12_01.DeleteHsmConfiguration

module Network.AWS.Redshift.V2012_12_01.DescribeClusterParameterGroups

module Network.AWS.Redshift.V2012_12_01.DescribeClusterParameters

module Network.AWS.Redshift.V2012_12_01.DescribeClusterSecurityGroups

module Network.AWS.Redshift.V2012_12_01.DescribeClusterSnapshots

module Network.AWS.Redshift.V2012_12_01.DescribeClusterSubnetGroups

module Network.AWS.Redshift.V2012_12_01.DescribeClusterVersions

module Network.AWS.Redshift.V2012_12_01.DescribeClusters

module Network.AWS.Redshift.V2012_12_01.DescribeDefaultClusterParameters

module Network.AWS.Redshift.V2012_12_01.DescribeEventCategories

module Network.AWS.Redshift.V2012_12_01.DescribeEventSubscriptions

module Network.AWS.Redshift.V2012_12_01.DescribeEvents

module Network.AWS.Redshift.V2012_12_01.DescribeHsmClientCertificates

module Network.AWS.Redshift.V2012_12_01.DescribeHsmConfigurations

module Network.AWS.Redshift.V2012_12_01.DescribeLoggingStatus

module Network.AWS.Redshift.V2012_12_01.DescribeOrderableClusterOptions

module Network.AWS.Redshift.V2012_12_01.DescribeReservedNodeOfferings

module Network.AWS.Redshift.V2012_12_01.DescribeReservedNodes

module Network.AWS.Redshift.V2012_12_01.DescribeResize

module Network.AWS.Redshift.V2012_12_01.DisableLogging

module Network.AWS.Redshift.V2012_12_01.DisableSnapshotCopy

module Network.AWS.Redshift.V2012_12_01.EnableLogging

module Network.AWS.Redshift.V2012_12_01.EnableSnapshotCopy

module Network.AWS.Redshift.V2012_12_01.ModifyCluster

module Network.AWS.Redshift.V2012_12_01.ModifyClusterParameterGroup

module Network.AWS.Redshift.V2012_12_01.ModifyClusterSubnetGroup

module Network.AWS.Redshift.V2012_12_01.ModifyEventSubscription

module Network.AWS.Redshift.V2012_12_01.ModifySnapshotCopyRetentionPeriod

module Network.AWS.Redshift.V2012_12_01.PurchaseReservedNodeOffering

module Network.AWS.Redshift.V2012_12_01.RebootCluster

module Network.AWS.Redshift.V2012_12_01.ResetClusterParameterGroup

module Network.AWS.Redshift.V2012_12_01.RestoreFromClusterSnapshot

module Network.AWS.Redshift.V2012_12_01.RevokeClusterSecurityGroupIngress

module Network.AWS.Redshift.V2012_12_01.RevokeSnapshotAccess

module Network.AWS.Redshift.V2012_12_01.RotateEncryptionKey

module Network.AWS.Redshift.V2012_12_01.Types