Network.AWS.Redshift.V2012_12_01.AuthorizeClusterSecurityGroupIngress

Description

Adds an inbound (ingress) rule to an Amazon Redshift security group. Depending on whether the application accessing your cluster is running on the Internet or an EC2 instance, you can authorize inbound access to either a Classless Interdomain Routing (CIDR) IP address range or an EC2 security group. EC2SecurityGroupName and EC2SecurityGroupOwnerId --> You can add as many as 20 ingress rules to an Amazon Redshift security group. The EC2 security group must be defined in the AWS region where the cluster resides. For an overview of CIDR blocks, see the Wikipedia article on Classless Inter-Domain Routing. You must also associate the security group with a cluster so that clients running on these IP addresses or the EC2 instance are authorized to connect to the cluster. For information about managing security groups, go to Working with Security Groups in the Amazon Redshift Management Guide. https:/redshift.us-east-1.amazonaws.com ?Action=AuthorizeClusterSecurityGroupIngress &CIDRIP=192.168.40.3/32 &ClusterSecurityGroupName=securitygroup1 &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20130123us-east-1redshiftaws4_request &x-amz-date=20130123T020649Z &x-amz-signedheaders=content-type;host;x-amz-date 192.168.40.3/32 authorized my security group securitygroup1 8c7cd4c8-6501-11e2-a8da-655adc216806.

Request

data AuthorizeClusterSecurityGroupIngress

???.

Request constructor

mkAuthorizeClusterSecurityGroupIngress

Arguments

:: Text
acsgiClusterSecurityGroupName

-> AuthorizeClusterSecurityGroupIngress

Smart constructor for the minimum required parameters to construct a valid AuthorizeClusterSecurityGroupIngress request.

Request lenses

acsgiClusterSecurityGroupName :: Lens' AuthorizeClusterSecurityGroupIngress Text

The name of the security group to which the ingress rule is added.

acsgiCIDRIP :: Lens' AuthorizeClusterSecurityGroupIngress (Maybe Text)

The IP range to be added the Amazon Redshift security group.

acsgiEC2SecurityGroupName :: Lens' AuthorizeClusterSecurityGroupIngress (Maybe Text)

The EC2 security group to be added the Amazon Redshift security group.

acsgiEC2SecurityGroupOwnerId :: Lens' AuthorizeClusterSecurityGroupIngress (Maybe Text)

The AWS account number of the owner of the security group specified by the EC2SecurityGroupName parameter. The AWS Access Key ID is not an acceptable value. Example: 111122223333.

Response

data AuthorizeClusterSecurityGroupIngressResponse

Response lenses

acsgirsClusterSecurityGroup :: Lens' AuthorizeClusterSecurityGroupIngressResponse (Maybe ClusterSecurityGroup)

Describes a security group.