Network.AWS.Redshift.V2012_12_01.CreateClusterSecurityGroup

Description

Creates a new Amazon Redshift security group. You use security groups to control access to non-VPC clusters. For information about managing security groups, go to Amazon Redshift Cluster Security Groups in the Amazon Redshift Management Guide. https:/redshift.us-east-1.amazonaws.com ?Action=CreateClusterSecurityGroup &ClusterSecurityGroupName=securitygroup1 &Description=my security group &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20130123us-east-1redshiftaws4_request &x-amz-date=20130123T005817Z &x-amz-signedheaders=content-type;host;x-amz-date my security group securitygroup1 f9ee270f-64f7-11e2-a8da-655adc216806.

Request

data CreateClusterSecurityGroup

???.

Request constructor

mkCreateClusterSecurityGroup

Arguments

:: Text
ccsgClusterSecurityGroupName

-> Text
ccsgDescription

-> CreateClusterSecurityGroup

Smart constructor for the minimum required parameters to construct a valid CreateClusterSecurityGroup request.

Request lenses

ccsgClusterSecurityGroupName :: Lens' CreateClusterSecurityGroup Text

The name for the security group. Amazon Redshift stores the value as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters or hyphens. Must not be Default. Must be unique for all security groups that are created by your AWS account. Example: examplesecuritygroup.

ccsgDescription :: Lens' CreateClusterSecurityGroup Text

A description for the security group.

Response

data CreateClusterSecurityGroupResponse

Response lenses

ccsgrClusterSecurityGroup :: Lens' CreateClusterSecurityGroupResponse (Maybe ClusterSecurityGroup)

Describes a security group.