Network.AWS.Redshift.V2012_12_01.CreateClusterSubnetGroup

Description

Creates a new Amazon Redshift subnet group. You must provide a list of one or more subnets in your existing Amazon Virtual Private Cloud (Amazon VPC) when creating Amazon Redshift subnet group. For information about subnet groups, go to Amazon Redshift Cluster Subnet Groups in the Amazon Redshift Management Guide. https:/redshift.us-east-1.amazonaws.com ?Action=CreateClusterSubnetGroup &ClusterSubnetGroupName=mysubnetgroup1 &Description=My subnet group 1 &SubnetIds.member.1=subnet-756a591f &SubnetIds.member.1=subnet-716a591b &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20130129us-east-1redshiftaws4_request &x-amz-date=20130129T192820Z &x-amz-signedheaders=content-type;host;x-amz-date vpc-796a5913 My subnet group 1 mysubnetgroup1 Complete Active subnet-756a591f us-east-1c 0a60660f-6a4a-11e2-aad2-71d00c36728e.

Request

data CreateClusterSubnetGroup

Request constructor

mkCreateClusterSubnetGroup

Arguments

:: Text
ccsg1ClusterSubnetGroupName

-> Text
ccsg1Description

-> [Text]
ccsg1SubnetIds

-> CreateClusterSubnetGroup

Smart constructor for the minimum required parameters to construct a valid CreateClusterSubnetGroup request.

Request lenses

ccsg1ClusterSubnetGroupName :: Lens' CreateClusterSubnetGroup Text

The name for the subnet group. Amazon Redshift stores the value as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters or hyphens. Must not be Default. Must be unique for all subnet groups that are created by your AWS account. Example: examplesubnetgroup.

ccsg1Description :: Lens' CreateClusterSubnetGroup Text

A description for the subnet group.

ccsg1SubnetIds :: Lens' CreateClusterSubnetGroup [Text]

An array of VPC subnet IDs. A maximum of 20 subnets can be modified in a single request.

Response

data CreateClusterSubnetGroupResponse

Response constructor

mkCreateClusterSubnetGroupResponse :: CreateClusterSubnetGroupResponse

Smart constructor for the minimum required parameters to construct a valid CreateClusterSubnetGroupResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

ccsgrrClusterSubnetGroup :: Lens' CreateClusterSubnetGroupResponse (Maybe ClusterSubnetGroup)

Describes a subnet group.