Network.AWS.Redshift.V2012_12_01.CreateHsmClientCertificate

Description

Creates an HSM client certificate that an Amazon Redshift cluster will use to connect to the client's HSM in order to store and retrieve the keys used to encrypt the cluster databases. The command returns a public key, which you must store in the HSM. In addition to creating the HSM certificate, you must create an Amazon Redshift HSM configuration that provides a cluster the information needed to store and use encryption keys in the HSM. For more information, go to Hardware Security Modules in the Amazon Redshift Management Guide.

Documentation

data CreateHsmClientCertificate

Constructors

CreateHsmClientCertificate

Fields

_chccmHsmClientCertificateIdentifier :: Text

The identifier to be assigned to the new HSM client certificate that the cluster will use to connect to the HSM to use the database encryption keys.

chccmHsmClientCertificateIdentifier :: Iso' CreateHsmClientCertificate Text

data CreateHsmClientCertificateResponse

Constructors

CreateHsmClientCertificateResponse

Fields

_hccwHsmClientCertificate :: Maybe HsmClientCertificate

Returns information about an HSM client certificate. The certificate is stored in a secure Hardware Storage Module (HSM), and used by the Amazon Redshift cluster to encrypt data files.

hccwHsmClientCertificate :: Iso' CreateHsmClientCertificateResponse (Maybe HsmClientCertificate)