Network.AWS.Redshift.V2012_12_01.DeleteClusterSubnetGroup

Description

Deletes the specified cluster subnet group. https:/redshift.us-east-1.amazonaws.com ?Action=DeleteClusterSubnetGroup &ClusterSubnetGroupName=my-subnet-group-2 &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20130130us-east-1redshiftaws4_request &x-amz-date=20130130T154635Z &x-amz-signedheaders=content-type;host;x-amz-date 3a63806b-6af4-11e2-b27b-4d850b1c672d.

Request

data DeleteClusterSubnetGroup

Request constructor

mkDeleteClusterSubnetGroup

Arguments

:: Text
dcsg1ClusterSubnetGroupName

-> DeleteClusterSubnetGroup

Smart constructor for the minimum required parameters to construct a valid DeleteClusterSubnetGroup request.

The fields accessible through corresponding lenses are:

ClusterSubnetGroupName :: Text

Request lenses

dcsg1ClusterSubnetGroupName :: Lens' DeleteClusterSubnetGroup Text

The name of the cluster subnet group name to be deleted.

Response

data DeleteClusterSubnetGroupResponse

Response constructor

mkDeleteClusterSubnetGroupResponse :: DeleteClusterSubnetGroupResponse

Smart constructor for the minimum required parameters to construct a valid DeleteClusterSubnetGroupResponse response.

This constructor is provided for convenience and testing purposes.