Network.AWS.Redshift.V2012_12_01.DescribeEventSubscriptions

Description

Lists descriptions of all the Amazon Redshift event notifications subscription for a customer account. If you specify a subscription name, lists the description for that subscription.

Documentation

describeEventSubscriptions :: DescribeEventSubscriptions

Minimum specification for a DescribeEventSubscriptions request.

data DescribeEventSubscriptions

Constructors

DescribeEventSubscriptions

Fields

_desnMaxRecords :: Maybe Integer

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

_desnSubscriptionName :: Maybe Text

The name of the Amazon Redshift event notification subscription to be described.

_desnMarker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeEventSubscriptions request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

desnSubscriptionName :: Lens' DescribeEventSubscriptions (Maybe Text)

desnMaxRecords :: Lens' DescribeEventSubscriptions (Maybe Integer)

desnMarker :: Lens' DescribeEventSubscriptions (Maybe Text)

data DescribeEventSubscriptionsResponse

Constructors

DescribeEventSubscriptionsResponse

Fields

_esmEventSubscriptionsList :: [EventSubscription]

A list of event subscriptions.

_esmMarker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

esmMarker :: Lens' DescribeEventSubscriptionsResponse (Maybe Text)

esmEventSubscriptionsList :: Lens' DescribeEventSubscriptionsResponse [EventSubscription]