Network.AWS.Redshift.V2012_12_01.DescribeHsmClientCertificates

Description

Returns information about the specified HSM client certificate. If no certificate ID is specified, returns information about all the HSM certificates owned by your AWS customer account.

Request

data DescribeHsmClientCertificates

Request constructor

mkDescribeHsmClientCertificates :: DescribeHsmClientCertificates

Smart constructor for the minimum required parameters to construct a valid DescribeHsmClientCertificates request.

Request lenses

dhcc1HsmClientCertificateIdentifier :: Lens' DescribeHsmClientCertificates (Maybe Text)

The identifier of a specific HSM client certificate for which you want information. If no identifier is specified, information is returned for all HSM client certificates owned by your AWS customer account.

dhcc1MaxRecords :: Lens' DescribeHsmClientCertificates (Maybe Integer)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

dhcc1Marker :: Lens' DescribeHsmClientCertificates (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeHsmClientCertificates request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

Response

data DescribeHsmClientCertificatesResponse

Response constructor

mkDescribeHsmClientCertificatesResponse :: DescribeHsmClientCertificatesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeHsmClientCertificatesResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

dhccrMarker :: Lens' DescribeHsmClientCertificatesResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dhccrHsmClientCertificates :: Lens' DescribeHsmClientCertificatesResponse [HsmClientCertificate]

A list of the identifiers for one or more HSM client certificates used by Amazon Redshift clusters to store and retrieve database encryption keys in an HSM.