Network.AWS.Redshift.V2012_12_01.DescribeLoggingStatus

Description

Describes whether information, such as queries and connection attempts, is being logged for the specified Amazon Redshift cluster.

Documentation

data DescribeLoggingStatus

Constructors

DescribeLoggingStatus

Fields

_dlsmClusterIdentifier :: Text

The identifier of the cluster to get the logging status from. Example: examplecluster.

dlsmClusterIdentifier :: Iso' DescribeLoggingStatus Text

data DescribeLoggingStatusResponse

Constructors

DescribeLoggingStatusResponse

Fields

_llzLoggingEnabled :: Maybe Bool

true if logging is on, false if logging is off.

_llzS3KeyPrefix :: Maybe Text

The prefix applied to the log file names.

_llzBucketName :: Maybe Text

The name of the S3 bucket where the log files are stored.

_llzLastFailureMessage :: Maybe Text

The message indicating that logs failed to be delivered.

_llzLastFailureTime :: Maybe ISO8601

The last time when logs failed to be delivered.

_llzLastSuccessfulDeliveryTime :: Maybe ISO8601

The last time when logs were delivered.

llzS3KeyPrefix :: Lens' DescribeLoggingStatusResponse (Maybe Text)

llzLoggingEnabled :: Lens' DescribeLoggingStatusResponse (Maybe Bool)

llzLastSuccessfulDeliveryTime :: Lens' DescribeLoggingStatusResponse (Maybe ISO8601)

llzLastFailureTime :: Lens' DescribeLoggingStatusResponse (Maybe ISO8601)

llzLastFailureMessage :: Lens' DescribeLoggingStatusResponse (Maybe Text)

llzBucketName :: Lens' DescribeLoggingStatusResponse (Maybe Text)