Network.AWS.Redshift.V2012_12_01.DescribeResize

Description

Returns information about the last resize operation for the specified cluster. If no resize operation has ever been initiated for the specified cluster, a HTTP 404 error is returned. If a resize operation was initiated and completed, the status of the resize remains as SUCCEEDED until the next resize. A resize operation can be requested using ModifyCluster and specifying a different number or type of nodes for the cluster. https:/redshift.us-east-1.amazonaws.com ?Action=DescribeResize &ClusterIdentifier=examplecluster &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20121207us-east-1redshiftaws4_request &x-amz-date=20121207T232427Z &x-amz-signedheaders=content-type;host;x-amz-date multi-node SUCCEEDED 6.5263 66922 0 users venue sales listing event date category 10254 dw1.xlarge 2 a6d59c61-a162-11e2-b2bc-fb54c9d11e09.

Request

data DescribeResize

Request constructor

mkDescribeResize

Arguments

:: Text
drClusterIdentifier

-> DescribeResize

Smart constructor for the minimum required parameters to construct a valid DescribeResize request.

Request lenses

drClusterIdentifier :: Lens' DescribeResize Text

The unique identifier of a cluster whose resize progress you are requesting. This parameter isn't case-sensitive. By default, resize operations for all clusters defined for an AWS account are returned.

Response

data DescribeResizeResponse

Describes the result of a cluster resize operation.

Response constructor

mkDescribeResizeResponse :: DescribeResizeResponse

Smart constructor for the minimum required parameters to construct a valid DescribeResizeResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

drrTargetNodeType :: Lens' DescribeResizeResponse (Maybe Text)

The node type that the cluster will have after the resize operation is complete.

drrTargetNumberOfNodes :: Lens' DescribeResizeResponse (Maybe Integer)

The number of nodes that the cluster will have after the resize operation is complete.

drrTargetClusterType :: Lens' DescribeResizeResponse (Maybe Text)

The cluster type after the resize operation is complete. Valid Values: multi-node | single-node.

drrStatus :: Lens' DescribeResizeResponse (Maybe Text)

The status of the resize operation. Valid Values: NONE | IN_PROGRESS | FAILED | SUCCEEDED.

drrImportTablesCompleted :: Lens' DescribeResizeResponse [Text]

The names of tables that have been completely imported . Valid Values: List of table names.

drrImportTablesInProgress :: Lens' DescribeResizeResponse [Text]

The names of tables that are being currently imported. Valid Values: List of table names.

drrImportTablesNotStarted :: Lens' DescribeResizeResponse [Text]

The names of tables that have not been yet imported. Valid Values: List of table names.

drrAvgResizeRateInMegaBytesPerSecond :: Lens' DescribeResizeResponse (Maybe Double)

The average rate of the resize operation over the last few minutes, measured in megabytes per second. After the resize operation completes, this value shows the average rate of the entire resize operation.

drrTotalResizeDataInMegaBytes :: Lens' DescribeResizeResponse (Maybe Integer)

The estimated total amount of data, in megabytes, on the cluster before the resize operation began.

drrProgressInMegaBytes :: Lens' DescribeResizeResponse (Maybe Integer)

While the resize operation is in progress, this value shows the current amount of data, in megabytes, that has been processed so far. When the resize operation is complete, this value shows the total amount of data, in megabytes, on the cluster, which may be more or less than TotalResizeDataInMegaBytes (the estimated total amount of data before resize).

drrElapsedTimeInSeconds :: Lens' DescribeResizeResponse (Maybe Integer)

The amount of seconds that have elapsed since the resize operation began. After the resize operation completes, this value shows the total actual time, in seconds, for the resize operation.

drrEstimatedTimeToCompletionInSeconds :: Lens' DescribeResizeResponse (Maybe Integer)

The estimated time remaining, in seconds, until the resize operation is complete. This value is calculated based on the average resize rate and the estimated amount of data remaining to be processed. Once the resize operation is complete, this value will be 0.