Network.AWS.Redshift.V2012_12_01.DisableLogging

Description

Stops logging information, such as queries and connection attempts, for the specified Amazon Redshift cluster.

Documentation

data DisableLogging

Constructors

DisableLogging

Fields

_dlmClusterIdentifier :: Text

The identifier of the cluster on which logging is to be stopped. Example: examplecluster.

dlmClusterIdentifier :: Iso' DisableLogging Text

data DisableLoggingResponse

Constructors

DisableLoggingResponse

Fields

_lsLoggingEnabled :: Maybe Bool

true if logging is on, false if logging is off.

_lsS3KeyPrefix :: Maybe Text

The prefix applied to the log file names.

_lsBucketName :: Maybe Text

The name of the S3 bucket where the log files are stored.

_lsLastFailureMessage :: Maybe Text

The message indicating that logs failed to be delivered.

_lsLastFailureTime :: Maybe ISO8601

The last time when logs failed to be delivered.

_lsLastSuccessfulDeliveryTime :: Maybe ISO8601

The last time when logs were delivered.

lsS3KeyPrefix :: Lens' DisableLoggingResponse (Maybe Text)

lsLoggingEnabled :: Lens' DisableLoggingResponse (Maybe Bool)

lsLastSuccessfulDeliveryTime :: Lens' DisableLoggingResponse (Maybe ISO8601)

lsLastFailureTime :: Lens' DisableLoggingResponse (Maybe ISO8601)

lsLastFailureMessage :: Lens' DisableLoggingResponse (Maybe Text)

lsBucketName :: Lens' DisableLoggingResponse (Maybe Text)