Network.AWS.Redshift.V2012_12_01.EnableLogging

Description

Starts logging information, such as queries and connection attempts, for the specified Amazon Redshift cluster.

Documentation

enableLogging

Arguments

:: Text
_elmClusterIdentifier

-> Text
_elmBucketName

-> EnableLogging

Minimum specification for a EnableLogging request.

data EnableLogging

Constructors

EnableLogging

Fields

_elmClusterIdentifier :: Text

The identifier of the cluster on which logging is to be started. Example: examplecluster.

_elmBucketName :: Text

The name of an existing S3 bucket where the log files are to be stored. Constraints: Must be in the same region as the cluster The cluster must have read bucket and put object permissions.

_elmS3KeyPrefix :: Maybe Text

The prefix applied to the log file names. Constraints: Cannot exceed 512 characters Cannot contain spaces( ), double quotes ("), single quotes ('), a backslash (), or control characters. The hexadecimal codes for invalid characters are: x00 to x20 x22 x27 x5c x7f or larger.

elmS3KeyPrefix :: Lens' EnableLogging (Maybe Text)

elmClusterIdentifier :: Lens' EnableLogging Text

elmBucketName :: Lens' EnableLogging Text

data EnableLoggingResponse

Constructors

EnableLoggingResponse

Fields

_ltLoggingEnabled :: Maybe Bool

true if logging is on, false if logging is off.

_ltS3KeyPrefix :: Maybe Text

The prefix applied to the log file names.

_ltBucketName :: Maybe Text

The name of the S3 bucket where the log files are stored.

_ltLastFailureMessage :: Maybe Text

The message indicating that logs failed to be delivered.

_ltLastFailureTime :: Maybe ISO8601

The last time when logs failed to be delivered.

_ltLastSuccessfulDeliveryTime :: Maybe ISO8601

The last time when logs were delivered.

ltS3KeyPrefix :: Lens' EnableLoggingResponse (Maybe Text)

ltLoggingEnabled :: Lens' EnableLoggingResponse (Maybe Bool)

ltLastSuccessfulDeliveryTime :: Lens' EnableLoggingResponse (Maybe ISO8601)

ltLastFailureTime :: Lens' EnableLoggingResponse (Maybe ISO8601)

ltLastFailureMessage :: Lens' EnableLoggingResponse (Maybe Text)

ltBucketName :: Lens' EnableLoggingResponse (Maybe Text)