Network.AWS.Redshift.V2012_12_01.EnableSnapshotCopy

Description

Enables the automatic copy of snapshots from one region to another region for a specified cluster.

Documentation

enableSnapshotCopy

Arguments

:: Text
_escmClusterIdentifier

-> Text
_escmDestinationRegion

-> EnableSnapshotCopy

Minimum specification for a EnableSnapshotCopy request.

data EnableSnapshotCopy

Constructors

EnableSnapshotCopy

Fields

_escmClusterIdentifier :: Text

The unique identifier of the source cluster to copy snapshots from. Constraints: Must be the valid name of an existing cluster that does not already have cross-region snapshot copy enabled.

_escmDestinationRegion :: Text

The destination region that you want to copy snapshots to. Constraints: Must be the name of a valid region. For more information, see Regions and Endpoints in the Amazon Web Services General Reference.

_escmRetentionPeriod :: Maybe Integer

The number of days to retain automated snapshots in the destination region after they are copied from the source region. Default: 7. Constraints: Must be at least 1 and no more than 35.

escmRetentionPeriod :: Lens' EnableSnapshotCopy (Maybe Integer)

escmDestinationRegion :: Lens' EnableSnapshotCopy Text

escmClusterIdentifier :: Lens' EnableSnapshotCopy Text

data EnableSnapshotCopyResponse

Constructors

EnableSnapshotCopyResponse

Fields

_cwCluster :: Maybe Cluster

Describes a cluster.

cwCluster :: Iso' EnableSnapshotCopyResponse (Maybe Cluster)