Network.AWS.Redshift.V2012_12_01.ModifyClusterParameterGroup

Description

Modifies the parameters of a parameter group. For more information about managing parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Management Guide. https:/redshift.us-east-1.amazonaws.com ?Action=ModifyClusterParameterGroup &ParameterGroupName=parametergroup1 &Parameters.member.1.ParameterName=extra_float_digits &Parameters.member.1.ParameterValue=2 &Parameters.member.2.ParameterName=wlm_json_configuration &Parameters.member.2.ParameterValue=[{"user_group":["example_user_group1"],"query_group":["example_query_group1"],"query_concurrency":7},{"query_concurrency":5}] &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20121208us-east-1redshiftaws4_request &x-amz-date=20121208T022525Z &x-amz-signedheaders=content-type;host;x-amz-date Your parameter group has been updated but changes won't get applied until you reboot the associated Clusters. parametergroup1 86e64043-40de-11e2-8a25-eb010998df4e.

Documentation

data ModifyClusterParameterGroup

Constructors

ModifyClusterParameterGroup

Fields

_mcpgmParameters :: [Parameter]

An array of parameters to be modified. A maximum of 20 parameters can be modified in a single request. For each parameter to be modified, you must supply at least the parameter name and parameter value; other name-value pairs of the parameter are optional. For the workload management (WLM) configuration, you must supply all the name-value pairs in the wlm_json_configuration parameter.

_mcpgmParameterGroupName :: Text

The name of the parameter group to be modified.

mcpgmParameters :: Lens' ModifyClusterParameterGroup [Parameter]

mcpgmParameterGroupName :: Lens' ModifyClusterParameterGroup Text

data ModifyClusterParameterGroupResponse

Constructors

ModifyClusterParameterGroupResponse

Fields

_cpgnnParameterGroupStatus :: Maybe Text

The status of the parameter group. For example, if you made a change to a parameter group name-value pair, then the change could be pending a reboot of an associated cluster.

_cpgnnParameterGroupName :: Maybe Text

The name of the cluster parameter group.

cpgnnParameterGroupStatus :: Lens' ModifyClusterParameterGroupResponse (Maybe Text)

cpgnnParameterGroupName :: Lens' ModifyClusterParameterGroupResponse (Maybe Text)