Network.AWS.Redshift.V2012_12_01.ModifySnapshotCopyRetentionPeriod

Description

Modifies the number of days to retain automated snapshots in the destination region after they are copied from the source region.

Documentation

data ModifySnapshotCopyRetentionPeriod

Constructors

ModifySnapshotCopyRetentionPeriod

Fields

_mscrpmRetentionPeriod :: Integer

The number of days to retain automated snapshots in the destination region after they are copied from the source region. If you decrease the retention period for automated snapshots that are copied to a destination region, Amazon Redshift will delete any existing automated snapshots that were copied to the destination region and that fall outside of the new retention period. Constraints: Must be at least 1 and no more than 35.

_mscrpmClusterIdentifier :: Text

The unique identifier of the cluster for which you want to change the retention period for automated snapshots that are copied to a destination region. Constraints: Must be the valid name of an existing cluster that has cross-region snapshot copy enabled.

mscrpmRetentionPeriod :: Lens' ModifySnapshotCopyRetentionPeriod Integer

mscrpmClusterIdentifier :: Lens' ModifySnapshotCopyRetentionPeriod Text

data ModifySnapshotCopyRetentionPeriodResponse

Constructors

ModifySnapshotCopyRetentionPeriodResponse

Fields

_cxCluster :: Maybe Cluster

Describes a cluster.

cxCluster :: Iso' ModifySnapshotCopyRetentionPeriodResponse (Maybe Cluster)