Network.AWS.Redshift.V2012_12_01.RebootCluster

Description

Reboots a cluster. This action is taken as soon as possible. It results in a momentary outage to the cluster, during which the cluster status is set to rebooting. A cluster event is created when the reboot is completed. Any pending cluster modifications (see ModifyCluster) are applied at this reboot. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Management Guide https:/redshift.us-east-1.amazonaws.com ?Action=RebootCluster &ClusterIdentifier=examplecluster &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20130123us-east-1redshiftaws4_request &x-amz-date=20130123T021951Z &x-amz-signedheaders=content-type;host;x-amz-date 1.0 5439 examplecluster.cobaosmlqshn.us-east-1.redshift.amazonaws.com rebooting 2 1 true false dev sun:06:30-sun:07:00 in-sync default.redshift-1.0 2013-01-22T19:23:59.368Z active default us-east-1c dw1.xlarge examplecluster true adminuser 5edee79e-6503-11e2-9e70-918437dd236d.

Documentation

data RebootCluster

Constructors

RebootCluster

Fields

_rcoClusterIdentifier :: Text

The cluster identifier.

rcoClusterIdentifier :: Iso' RebootCluster Text

data RebootClusterResponse

Constructors

RebootClusterResponse

Fields

_cyCluster :: Maybe Cluster

Describes a cluster.

cyCluster :: Iso' RebootClusterResponse (Maybe Cluster)