Network.AWS.Redshift.V2012_12_01.RevokeSnapshotAccess

Description

Removes the ability of the specified AWS customer account to restore the specified snapshot. If the account is currently restoring the snapshot, the restore will run to completion. For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Management Guide.

Documentation

revokeSnapshotAccess

Arguments

:: Text
_rsamSnapshotIdentifier

-> Text
_rsamAccountWithRestoreAccess

-> RevokeSnapshotAccess

Minimum specification for a RevokeSnapshotAccess request.

data RevokeSnapshotAccess

Constructors

RevokeSnapshotAccess

Fields

_rsamSnapshotIdentifier :: Text

The identifier of the snapshot that the account can no longer access.

_rsamAccountWithRestoreAccess :: Text

The identifier of the AWS customer account that can no longer restore the specified snapshot.

_rsamSnapshotClusterIdentifier :: Maybe Text

The identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

rsamSnapshotIdentifier :: Lens' RevokeSnapshotAccess Text

rsamSnapshotClusterIdentifier :: Lens' RevokeSnapshotAccess (Maybe Text)

rsamAccountWithRestoreAccess :: Lens' RevokeSnapshotAccess Text

data RevokeSnapshotAccessResponse

Constructors

RevokeSnapshotAccessResponse

Fields

_sssssssssssssssssssssssssssssssssssrSnapshot :: Maybe Snapshot

Describes a snapshot.

sssssssssssssssssssssssssssssssssssrSnapshot :: Iso' RevokeSnapshotAccessResponse (Maybe Snapshot)