Network.AWS.Redshift.V2012_12_01.RotateEncryptionKey

Description

Rotates the encryption keys for a cluster.

Documentation

data RotateEncryptionKey

Constructors

RotateEncryptionKey

Fields

_rekmClusterIdentifier :: Text

The unique identifier of the cluster that you want to rotate the encryption keys for. Constraints: Must be the name of valid cluster that has encryption enabled.

rekmClusterIdentifier :: Iso' RotateEncryptionKey Text

data RotateEncryptionKeyResponse

Constructors

RotateEncryptionKeyResponse

Fields

_cczCluster :: Maybe Cluster

Describes a cluster.

cczCluster :: Iso' RotateEncryptionKeyResponse (Maybe Cluster)