Network.AWS.Redshift.V2012_12_01.Types

Description

Amazon Redshift is a fast, fully managed, petabyte-scale data warehouse service that makes it simple and cost-effective to efficiently analyze all your data using your existing business intelligence tools. You can start small for just $0.25 per hour with no commitments or upfront costs and scale to a petabyte or more for $1,000 per terabyte per year, less than a tenth of most other data warehousing solutions.

Documentation

data Redshift

Supported version (2012-12-01) of the Amazon Redshift service.

xmlOptions :: Tagged a XMLOptions

data SourceType

The event source to retrieve events for. If no value is specified, all events are returned. Constraints: If SourceType is supplied, SourceIdentifier must also be provided. Specify cluster when SourceIdentifier is a cluster identifier. Specify cluster-security-group when SourceIdentifier is a cluster security group name. Specify cluster-parameter-group when SourceIdentifier is a cluster parameter group name. Specify cluster-snapshot when SourceIdentifier is a cluster snapshot identifier.

Constructors

SourceTypeCluster
cluster

SourceTypeClusterParameterGroup
cluster-parameter-group

SourceTypeClusterSecurityGroup
cluster-security-group

SourceTypeClusterSnapshot
cluster-snapshot

newtype AccountWithRestoreAccess

Describes an AWS customer account authorized to restore a snapshot.

Constructors

AccountWithRestoreAccess

Fields

_awraAccountId :: Maybe Text

The identifier of an AWS customer account authorized to restore a snapshot.

newtype AvailabilityZone

Describes an availability zone.

Constructors

AvailabilityZone

Fields

_azName :: Maybe Text

The name of the availability zone.

data Cluster

Describes a cluster.

Constructors

Cluster

Fields

_cRestoreStatus :: Maybe RestoreStatus

Describes the status of a cluster restore action. Returns null if the cluster was not created by restoring a snapshot.

_cClusterSnapshotCopyStatus :: Maybe ClusterSnapshotCopyStatus

Returns the destination region and retention period that are configured for cross-region snapshot copy.

_cClusterRevisionNumber :: Maybe Text

The specific revision number of the database in the cluster.

_cPubliclyAccessible :: Maybe Bool

If true, the cluster can be accessed from a public network.

_cMasterUsername :: Maybe Text

The master user name for the cluster. This name is used to connect to the database that is specified in DBName.

_cVpcId :: Maybe Text

The identifier of the VPC the cluster is in, if the cluster is in a VPC.

_cClusterSecurityGroups :: [ClusterSecurityGroupMembership]

A list of cluster security group that are associated with the cluster. Each security group is represented by an element that contains ClusterSecurityGroup.Name and ClusterSecurityGroup.Status subelements. Cluster security groups are used when the cluster is not created in a VPC. Clusters that are created in a VPC use VPC security groups, which are listed by the VpcSecurityGroups parameter.

_cAutomatedSnapshotRetentionPeriod :: Maybe Integer

The number of days that automatic cluster snapshots are retained.

_cEncrypted :: Maybe Bool

If true, data in the cluster is encrypted at rest.

_cClusterSubnetGroupName :: Maybe Text

The name of the subnet group that is associated with the cluster. This parameter is valid only when the cluster is in a VPC.

_cClusterIdentifier :: Maybe Text

The unique identifier of the cluster.

_cNumberOfNodes :: Maybe Integer

The number of compute nodes in the cluster.

_cClusterPublicKey :: Maybe Text

The public key for the cluster.

_cPreferredMaintenanceWindow :: Maybe Text

The weekly time range (in UTC) during which system maintenance can occur.

_cModifyStatus :: Maybe Text

The status of a modify operation, if any, initiated for the cluster.

_cClusterParameterGroups :: [ClusterParameterGroupStatus]

The list of cluster parameter groups that are associated with this cluster.

_cAvailabilityZone :: Maybe Text

The name of the Availability Zone in which the cluster is located.

_cVpcSecurityGroups :: [VpcSecurityGroupMembership]

A list of Virtual Private Cloud (VPC) security groups that are associated with the cluster. This parameter is returned only if the cluster is in a VPC.

_cHsmStatus :: Maybe HsmStatus

Reports whether the Amazon Redshift cluster has finished applying any HSM settings changes specified in a modify cluster command. Values: active, applying.

_cElasticIpStatus :: Maybe ElasticIpStatus

Describes the status of the elastic IP (EIP) address.

_cClusterVersion :: Maybe Text

The version ID of the Amazon Redshift engine that is running on the cluster.

_cNodeType :: Maybe Text

The node type for the nodes in the cluster.

_cClusterCreateTime :: Maybe ISO8601

The date and time that the cluster was created.

_cEndpoint :: Maybe Endpoint

The connection endpoint.

_cAllowVersionUpgrade :: Maybe Bool

If true, version upgrades will be applied automatically to the cluster during the maintenance window.

_cClusterStatus :: Maybe Text

The current state of this cluster. Possible values include available, creating, deleting, rebooting, renaming, and resizing.

_cPendingModifiedValues :: Maybe PendingModifiedValues

If present, changes to the cluster are pending. Specific pending changes are identified by subelements.

_cClusterNodes :: [ClusterNode]

The nodes in a cluster.

_cDBName :: Maybe Text

The name of the initial database that was created when the cluster was created. This same name is returned for the life of the cluster. If an initial database was not specified, a database named "dev" was created by default.

data ClusterNode

The identifier of a node in a cluster. --&gt;.

Constructors

ClusterNode

Fields

_cnNodeRole :: Maybe Text

Whether the node is a leader node or a compute node.

_cnPrivateIPAddress :: Maybe Text

The private IP address of a node within a cluster.

_cnPublicIPAddress :: Maybe Text

The public IP address of a node within a cluster.

data ClusterParameterGroup

Describes a parameter group.

Constructors

ClusterParameterGroup

Fields

_cpgParameterGroupFamily :: Maybe Text

The name of the cluster parameter group family that this cluster parameter group is compatible with.

_cpgDescription :: Maybe Text

The description of the parameter group.

_cpgParameterGroupName :: Maybe Text

The name of the cluster parameter group.

data ClusterParameterGroupStatus

Describes the status of a parameter group.

Constructors

ClusterParameterGroupStatus

Fields

_cpgsParameterApplyStatus :: Maybe Text

The status of parameter updates.

_cpgsParameterGroupName :: Maybe Text

The name of the cluster parameter group.

data ClusterSecurityGroup

Describes a security group.

Constructors

ClusterSecurityGroup

Fields

_csiClusterSecurityGroupName :: Maybe Text

The name of the cluster security group to which the operation was applied.

_csiIPRanges :: [IPRange]

A list of IP ranges (CIDR blocks) that are permitted to access clusters associated with this cluster security group.

_csiEC2SecurityGroups :: [EC2SecurityGroup]

A list of EC2 security groups that are permitted to access clusters associated with this cluster security group.

_csiDescription :: Maybe Text

A description of the security group.

data ClusterSecurityGroupMembership

Describes a security group.

Constructors

ClusterSecurityGroupMembership

Fields

_csgmStatus :: Maybe Text

The status of the cluster security group.

_csgmClusterSecurityGroupName :: Maybe Text

The name of the cluster security group.

data ClusterSnapshotCopyStatus

Returns the destination region and retention period that are configured for cross-region snapshot copy.

Constructors

ClusterSnapshotCopyStatus

Fields

_cscsRetentionPeriod :: Maybe Integer

The number of days that automated snapshots are retained in the destination region after they are copied from a source region.

_cscsDestinationRegion :: Maybe Text

The destination region that snapshots are automatically copied to when cross-region snapshot copy is enabled.

data ClusterSubnetGroup

Describes a subnet group.

Constructors

ClusterSubnetGroup

Fields

_csgVpcId :: Maybe Text

The VPC ID of the cluster subnet group.

_csgSubnets :: [Subnet]

A list of the VPC Subnet elements.

_csgClusterSubnetGroupName :: Maybe Text

The name of the cluster subnet group.

_csgSubnetGroupStatus :: Maybe Text

The status of the cluster subnet group. Possible values are Complete, Incomplete and Invalid.

_csgDescription :: Maybe Text

The description of the cluster subnet group.

data ClusterVersion

Describes a cluster version, including the parameter group family and description of the version.

Constructors

ClusterVersion

Fields

_ccnClusterParameterGroupFamily :: Maybe Text

The name of the cluster parameter group family for the cluster.

_ccnClusterVersion :: Maybe Text

The version number used by the cluster.

_ccnDescription :: Maybe Text

The description of the cluster version.

data DefaultClusterParameters

Describes the default cluster parameters for a parameter group family.

Constructors

DefaultClusterParameters

Fields

_dcpMarker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

_dcpParameters :: [Parameter]

The list of cluster default parameters.

_dcpParameterGroupFamily :: Maybe Text

The name of the cluster parameter group family to which the engine default parameters apply.

data EC2SecurityGroup

Describes an Amazon EC2 security group.

Constructors

EC2SecurityGroup

Fields

_ecsgStatus :: Maybe Text

The status of the EC2 security group.

_ecsgEC2SecurityGroupOwnerId :: Maybe Text

The AWS ID of the owner of the EC2 security group specified in the EC2SecurityGroupName field.

_ecsgEC2SecurityGroupName :: Maybe Text

The name of the EC2 Security Group.

data ElasticIpStatus

Describes the status of the elastic IP (EIP) address.

Constructors

ElasticIpStatus

Fields

_eisStatus :: Maybe Text

Describes the status of the elastic IP (EIP) address.

_eisElasticIp :: Maybe Text

The elastic IP (EIP) address for the cluster.

data Endpoint

The connection endpoint.

Constructors

Endpoint

Fields

_fAddress :: Maybe Text

The DNS address of the Cluster.

_fPort :: Maybe Integer

The port that the database engine is listening on.

data Event

Describes an event.

Constructors

Event

Fields

_etSourceType :: Maybe SourceType

The source type for this event.

_etSeverity :: Maybe Text

The severity of the event. Values: ERROR, INFO.

_etSourceIdentifier :: Maybe Text

The identifier for the source of the event.

_etDate :: Maybe ISO8601

The date and time of the event.

_etEventCategories :: [Text]

A list of the event categories.

_etMessage :: Maybe Text

The text of this event.

_etEventId :: Maybe Text

The identifier of the event.

data EventCategoriesMap

Constructors

EventCategoriesMap

Fields

_ecsSourceType :: Maybe Text

The Amazon Redshift source type, such as cluster or cluster-snapshot, that the returned categories belong to.

_ecsEvents :: [EventInfoMap]

The events in the event category.

data EventInfoMap

Constructors

EventInfoMap

Fields

_eimEventDescription :: Maybe Text

The description of an Amazon Redshift event.

_eimSeverity :: Maybe Text

The severity of the event. Values: ERROR, INFO.

_eimEventCategories :: [Text]

The category of an Amazon Redshift event.

_eimEventId :: Maybe Text

The identifier of an Amazon Redshift event.

data EventSubscription

Constructors

EventSubscription

Fields

_esStatus :: Maybe Text

The status of the Amazon Redshift event notification subscription. Constraints: Can be one of the following: active | no-permission | topic-not-exist The status "no-permission" indicates that Amazon Redshift no longer has permission to post to the Amazon SNS topic. The status "topic-not-exist" indicates that the topic was deleted after the subscription was created.

_esCustomerAwsId :: Maybe Text

The AWS customer account associated with the Amazon Redshift event notification subscription.

_esCustSubscriptionId :: Maybe Text

The name of the Amazon Redshift event notification subscription.

_esSnsTopicArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon SNS topic used by the event notification subscription.

_esEnabled :: Maybe Bool

A Boolean value indicating whether the subscription is enabled. true indicates the subscription is enabled.

_esSourceType :: Maybe Text

The source type of the events returned the Amazon Redshift event notification, such as cluster, or cluster-snapshot.

_esSeverity :: Maybe Text

The event severity specified in the Amazon Redshift event notification subscription. Values: ERROR, INFO.

_esSubscriptionCreationTime :: Maybe ISO8601

The date and time the Amazon Redshift event notification subscription was created.

_esEventCategoriesList :: [Text]

The list of Amazon Redshift event categories specified in the event notification subscription. Values: Configuration, Management, Monitoring, Security.

_esSourceIdsList :: [Text]

A list of the sources that publish events to the Amazon Redshift event notification subscription.

data HsmClientCertificate

Returns information about an HSM client certificate. The certificate is stored in a secure Hardware Storage Module (HSM), and used by the Amazon Redshift cluster to encrypt data files.

Constructors

HsmClientCertificate

Fields

_hccHsmClientCertificateIdentifier :: Maybe Text

The identifier of the HSM client certificate.

_hccHsmClientCertificatePublicKey :: Maybe Text

The public key that the Amazon Redshift cluster will use to connect to the HSM. You must register the public key in the HSM.

data HsmConfiguration

Returns information about an HSM configuration, which is an object that describes to Amazon Redshift clusters the information they require to connect to an HSM where they can store database encryption keys.

Constructors

HsmConfiguration

Fields

_hcHsmConfigurationIdentifier :: Maybe Text

The name of the Amazon Redshift HSM configuration.

_hcHsmPartitionName :: Maybe Text

The name of the partition in the HSM where the Amazon Redshift clusters will store their database encryption keys.

_hcDescription :: Maybe Text

A text description of the HSM configuration.

_hcHsmIpAddress :: Maybe Text

The IP address that the Amazon Redshift cluster must use to access the HSM.

data HsmStatus

Reports whether the Amazon Redshift cluster has finished applying any HSM settings changes specified in a modify cluster command. Values: active, applying.

Constructors

HsmStatus

Fields

_hsStatus :: Maybe Text

Reports whether the Amazon Redshift cluster has finished applying any HSM settings changes specified in a modify cluster command. Values: active, applying.

_hsHsmConfigurationIdentifier :: Maybe Text

Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

_hsHsmClientCertificateIdentifier :: Maybe Text

Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

data IPRange

Describes an IP range used in a security group.

Constructors

IPRange

Fields

_iprStatus :: Maybe Text

The status of the IP range, for example, "authorized".

_iprCIDRIP :: Maybe Text

The IP range in Classless Inter-Domain Routing (CIDR) notation.

data OrderableClusterOption

Describes an orderable cluster option.

Constructors

OrderableClusterOption

Fields

_ocoAvailabilityZones :: [AvailabilityZone]

A list of availability zones for the orderable cluster.

_ocoClusterType :: Maybe Text

The cluster type, for example multi-node.

_ocoClusterVersion :: Maybe Text

The version of the orderable cluster.

_ocoNodeType :: Maybe Text

The node type for the orderable cluster.

data Parameter

Describes a parameter in a cluster parameter group.

Constructors

Parameter

Fields

_prParameterValue :: Maybe Text

The value of the parameter.

_prMinimumEngineVersion :: Maybe Text

The earliest engine version to which the parameter can apply.

_prSource :: Maybe Text

The source of the parameter value, such as "engine-default" or "user".

_prIsModifiable :: Maybe Bool

If true, the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

_prDataType :: Maybe Text

The data type of the parameter.

_prAllowedValues :: Maybe Text

The valid range of values for the parameter.

_prParameterName :: Maybe Text

The name of the parameter.

_prDescription :: Maybe Text

A description of the parameter.

data PendingModifiedValues

If present, changes to the cluster are pending. Specific pending changes are identified by subelements.

Constructors

PendingModifiedValues

Fields

_pmvMasterUserPassword :: Maybe Text

The pending or in-progress change of the master user password for the cluster.

_pmvAutomatedSnapshotRetentionPeriod :: Maybe Integer

The pending or in-progress change of the automated snapshot retention period.

_pmvClusterIdentifier :: Maybe Text

The pending or in-progress change of the new identifier for the cluster.

_pmvNumberOfNodes :: Maybe Integer

The pending or in-progress change of the number of nodes in the cluster.

_pmvClusterType :: Maybe Text

The pending or in-progress change of the cluster type.

_pmvClusterVersion :: Maybe Text

The pending or in-progress change of the service version.

_pmvNodeType :: Maybe Text

The pending or in-progress change of the cluster's node type.

data RecurringCharge

Describes a recurring charge.

Constructors

RecurringCharge

Fields

_rcRecurringChargeFrequency :: Maybe Text

The frequency at which the recurring charge amount is applied.

_rcRecurringChargeAmount :: Maybe Double

The amount charged per the period of time specified by the recurring charge frequency.

data ReservedNode

Describes a reserved node.

Constructors

ReservedNode

Fields

_rnState :: Maybe Text

The state of the reserved Compute Node. Possible Values: pending-payment-This reserved node has recently been purchased, and the sale has been approved, but payment has not yet been confirmed. active-This reserved node is owned by the caller and is available for use. payment-failed-Payment failed for the purchase attempt.

_rnCurrencyCode :: Maybe Text

The currency code for the reserved cluster.

_rnStartTime :: Maybe ISO8601

The time the reservation started. You purchase a reserved node offering for a duration. This is the start time of that duration.

_rnNodeCount :: Maybe Integer

The number of reserved compute nodes.

_rnReservedNodeId :: Maybe Text

The unique identifier for the reservation.

_rnReservedNodeOfferingId :: Maybe Text

The identifier for the reserved node offering.

_rnRecurringCharges :: [RecurringCharge]

The recurring charges for the reserved node.

_rnOfferingType :: Maybe Text

The anticipated utilization of the reserved node, as defined in the reserved node offering.

_rnUsagePrice :: Maybe Double

The hourly rate Amazon Redshift charge you for this reserved node.

_rnNodeType :: Maybe Text

The node type of the reserved node.

_rnFixedPrice :: Maybe Double

The fixed cost Amazon Redshift charged you for this reserved node.

_rnDuration :: Maybe Integer

The duration of the node reservation in seconds.

data ReservedNodeOffering

Describes a reserved node offering.

Constructors

ReservedNodeOffering

Fields

_rnoCurrencyCode :: Maybe Text

The currency code for the compute nodes offering.

_rnoReservedNodeOfferingId :: Maybe Text

The offering identifier.

_rnoRecurringCharges :: [RecurringCharge]

The charge to your account regardless of whether you are creating any clusters using the node offering. Recurring charges are only in effect for heavy-utilization reserved nodes.

_rnoOfferingType :: Maybe Text

The anticipated utilization of the reserved node, as defined in the reserved node offering.

_rnoUsagePrice :: Maybe Double

The rate you are charged for each hour the cluster that is using the offering is running.

_rnoNodeType :: Maybe Text

The node type offered by the reserved node offering.

_rnoFixedPrice :: Maybe Double

The upfront fixed charge you will pay to purchase the specific reserved node offering.

_rnoDuration :: Maybe Integer

The duration, in seconds, for which the offering will reserve the node.

data RestoreStatus

Describes the status of a cluster restore action. Returns null if the cluster was not created by restoring a snapshot.

Constructors

RestoreStatus

Fields

_rsStatus :: Maybe Text

The status of the restore action. Returns starting, restoring, completed, or failed.

_rsEstimatedTimeToCompletionInSeconds :: Maybe Integer

The estimate of the time remaining before the restore will complete. Returns 0 for a completed restore.

_rsCurrentRestoreRateInMegaBytesPerSecond :: Maybe Double

The number of megabytes per second being transferred from the backup storage. Returns the average rate for a completed backup.

_rsProgressInMegaBytes :: Maybe Integer

The number of megabytes that have been transferred from snapshot storage.

_rsElapsedTimeInSeconds :: Maybe Integer

The amount of time an in-progress restore has been running, or the amount of time it took a completed restore to finish.

_rsSnapshotSizeInMegaBytes :: Maybe Integer

The size of the set of snapshot data used to restore the cluster.

data Snapshot

Describes a snapshot.

Constructors

Snapshot

Fields

_ssstStatus :: Maybe Text

The snapshot status. The value of the status depends on the API operation used. CreateClusterSnapshot and CopyClusterSnapshot returns status as "creating". DescribeClusterSnapshots returns status as "creating", "available", "final snapshot", or "failed". DeleteClusterSnapshot returns status as "deleted".

_ssstAccountsWithRestoreAccess :: [AccountWithRestoreAccess]

A list of the AWS customer accounts authorized to restore the snapshot. Returns null if no accounts are authorized. Visible only to the snapshot owner.

_ssstSnapshotIdentifier :: Maybe Text

The snapshot identifier that is provided in the request.

_ssstEncryptedWithHSM :: Maybe Bool

A boolean that indicates whether the snapshot data is encrypted using the HSM keys of the source cluster. true indicates that the data is encrypted using HSM keys.

_ssstMasterUsername :: Maybe Text

The master user name for the cluster.

_ssstSourceRegion :: Maybe Text

The source region from which the snapshot was copied.

_ssstVpcId :: Maybe Text

The VPC identifier of the cluster if the snapshot is from a cluster in a VPC. Otherwise, this field is not in the output.

_ssstBackupProgressInMegaBytes :: Maybe Double

The number of megabytes that have been transferred to the snapshot backup.

_ssstEncrypted :: Maybe Bool

If true, the data in the snapshot is encrypted at rest.

_ssstClusterIdentifier :: Maybe Text

The identifier of the cluster for which the snapshot was taken.

_ssstNumberOfNodes :: Maybe Integer

The number of nodes in the cluster.

_ssstSnapshotType :: Maybe Text

The snapshot type. Snapshots created using CreateClusterSnapshot and CopyClusterSnapshot will be of type "manual".

_ssstAvailabilityZone :: Maybe Text

The Availability Zone in which the cluster was created.

_ssstCurrentBackupRateInMegaBytesPerSecond :: Maybe Double

The number of megabytes per second being transferred to the snapshot backup. Returns 0 for a completed backup.

_ssstSnapshotCreateTime :: Maybe ISO8601

The time (UTC) when Amazon Redshift began the snapshot. A snapshot contains a copy of the cluster data as of this exact time.

_ssstClusterVersion :: Maybe Text

The version ID of the Amazon Redshift engine that is running on the cluster.

_ssstOwnerAccount :: Maybe Text

For manual snapshots, the AWS customer account used to create or copy the snapshot. For automatic snapshots, the owner of the cluster. The owner can perform all snapshot actions, such as sharing a manual snapshot.

_ssstNodeType :: Maybe Text

The node type of the nodes in the cluster.

_ssstElapsedTimeInSeconds :: Maybe Integer

The amount of time an in-progress snapshot backup has been running, or the amount of time it took a completed backup to finish.

_ssstClusterCreateTime :: Maybe ISO8601

The time (UTC) when the cluster was originally created.

_ssstEstimatedSecondsToCompletion :: Maybe Integer

The estimate of the time remaining before the snapshot backup will complete. Returns 0 for a completed backup.

_ssstActualIncrementalBackupSizeInMegaBytes :: Maybe Double

The size of the incremental backup.

_ssstPort :: Maybe Integer

The port that the cluster is listening on.

_ssstTotalBackupSizeInMegaBytes :: Maybe Double

The size of the complete set of backup data that would be used to restore the cluster.

_ssstDBName :: Maybe Text

The name of the database that was created when the cluster was created.

data Subnet

Describes a subnet.

Constructors

Subnet

Fields

_ssssssxSubnetStatus :: Maybe Text

The status of the subnet.

_ssssssxSubnetIdentifier :: Maybe Text

The identifier of the subnet.

_ssssssxSubnetAvailabilityZone :: Maybe AvailabilityZone

Describes an availability zone.

data VpcSecurityGroupMembership

Describes the members of a VPC security group.

Constructors

VpcSecurityGroupMembership

Fields

_vsgmStatus :: Maybe Text

_vsgmVpcSecurityGroupId :: Maybe Text

awraAccountId :: Iso' AccountWithRestoreAccess (Maybe Text)

azName :: Iso' AvailabilityZone (Maybe Text)

cVpcSecurityGroups :: Lens' Cluster [VpcSecurityGroupMembership]

cVpcId :: Lens' Cluster (Maybe Text)

cRestoreStatus :: Lens' Cluster (Maybe RestoreStatus)

cPubliclyAccessible :: Lens' Cluster (Maybe Bool)

cPreferredMaintenanceWindow :: Lens' Cluster (Maybe Text)

cPendingModifiedValues :: Lens' Cluster (Maybe PendingModifiedValues)

cNumberOfNodes :: Lens' Cluster (Maybe Integer)

cNodeType :: Lens' Cluster (Maybe Text)

cModifyStatus :: Lens' Cluster (Maybe Text)

cMasterUsername :: Lens' Cluster (Maybe Text)

cHsmStatus :: Lens' Cluster (Maybe HsmStatus)

cEndpoint :: Lens' Cluster (Maybe Endpoint)

cEncrypted :: Lens' Cluster (Maybe Bool)

cElasticIpStatus :: Lens' Cluster (Maybe ElasticIpStatus)

cDBName :: Lens' Cluster (Maybe Text)

cClusterVersion :: Lens' Cluster (Maybe Text)

cClusterSubnetGroupName :: Lens' Cluster (Maybe Text)

cClusterStatus :: Lens' Cluster (Maybe Text)

cClusterSnapshotCopyStatus :: Lens' Cluster (Maybe ClusterSnapshotCopyStatus)

cClusterSecurityGroups :: Lens' Cluster [ClusterSecurityGroupMembership]

cClusterRevisionNumber :: Lens' Cluster (Maybe Text)

cClusterPublicKey :: Lens' Cluster (Maybe Text)

cClusterParameterGroups :: Lens' Cluster [ClusterParameterGroupStatus]

cClusterNodes :: Lens' Cluster [ClusterNode]

cClusterIdentifier :: Lens' Cluster (Maybe Text)

cClusterCreateTime :: Lens' Cluster (Maybe ISO8601)

cAvailabilityZone :: Lens' Cluster (Maybe Text)

cAutomatedSnapshotRetentionPeriod :: Lens' Cluster (Maybe Integer)

cAllowVersionUpgrade :: Lens' Cluster (Maybe Bool)

cnPublicIPAddress :: Lens' ClusterNode (Maybe Text)

cnPrivateIPAddress :: Lens' ClusterNode (Maybe Text)

cnNodeRole :: Lens' ClusterNode (Maybe Text)

cpgParameterGroupName :: Lens' ClusterParameterGroup (Maybe Text)

cpgParameterGroupFamily :: Lens' ClusterParameterGroup (Maybe Text)

cpgDescription :: Lens' ClusterParameterGroup (Maybe Text)

cpgsParameterGroupName :: Lens' ClusterParameterGroupStatus (Maybe Text)

cpgsParameterApplyStatus :: Lens' ClusterParameterGroupStatus (Maybe Text)

csiIPRanges :: Lens' ClusterSecurityGroup [IPRange]

csiEC2SecurityGroups :: Lens' ClusterSecurityGroup [EC2SecurityGroup]

csiDescription :: Lens' ClusterSecurityGroup (Maybe Text)

csiClusterSecurityGroupName :: Lens' ClusterSecurityGroup (Maybe Text)

csgmStatus :: Lens' ClusterSecurityGroupMembership (Maybe Text)

csgmClusterSecurityGroupName :: Lens' ClusterSecurityGroupMembership (Maybe Text)

cscsRetentionPeriod :: Lens' ClusterSnapshotCopyStatus (Maybe Integer)

cscsDestinationRegion :: Lens' ClusterSnapshotCopyStatus (Maybe Text)

csgVpcId :: Lens' ClusterSubnetGroup (Maybe Text)

csgSubnets :: Lens' ClusterSubnetGroup [Subnet]

csgSubnetGroupStatus :: Lens' ClusterSubnetGroup (Maybe Text)

csgDescription :: Lens' ClusterSubnetGroup (Maybe Text)

csgClusterSubnetGroupName :: Lens' ClusterSubnetGroup (Maybe Text)

ccnDescription :: Lens' ClusterVersion (Maybe Text)

ccnClusterVersion :: Lens' ClusterVersion (Maybe Text)

ccnClusterParameterGroupFamily :: Lens' ClusterVersion (Maybe Text)

dcpParameters :: Lens' DefaultClusterParameters [Parameter]

dcpParameterGroupFamily :: Lens' DefaultClusterParameters (Maybe Text)

dcpMarker :: Lens' DefaultClusterParameters (Maybe Text)

ecsgStatus :: Lens' EC2SecurityGroup (Maybe Text)

ecsgEC2SecurityGroupOwnerId :: Lens' EC2SecurityGroup (Maybe Text)

ecsgEC2SecurityGroupName :: Lens' EC2SecurityGroup (Maybe Text)

eisStatus :: Lens' ElasticIpStatus (Maybe Text)

eisElasticIp :: Lens' ElasticIpStatus (Maybe Text)

fPort :: Lens' Endpoint (Maybe Integer)

fAddress :: Lens' Endpoint (Maybe Text)

etSourceType :: Lens' Event (Maybe SourceType)

etSourceIdentifier :: Lens' Event (Maybe Text)

etSeverity :: Lens' Event (Maybe Text)

etMessage :: Lens' Event (Maybe Text)

etEventId :: Lens' Event (Maybe Text)

etEventCategories :: Lens' Event [Text]

etDate :: Lens' Event (Maybe ISO8601)

ecsSourceType :: Lens' EventCategoriesMap (Maybe Text)

ecsEvents :: Lens' EventCategoriesMap [EventInfoMap]

eimSeverity :: Lens' EventInfoMap (Maybe Text)

eimEventId :: Lens' EventInfoMap (Maybe Text)

eimEventDescription :: Lens' EventInfoMap (Maybe Text)

eimEventCategories :: Lens' EventInfoMap [Text]

esSubscriptionCreationTime :: Lens' EventSubscription (Maybe ISO8601)

esStatus :: Lens' EventSubscription (Maybe Text)

esSourceType :: Lens' EventSubscription (Maybe Text)

esSourceIdsList :: Lens' EventSubscription [Text]

esSnsTopicArn :: Lens' EventSubscription (Maybe Text)

esSeverity :: Lens' EventSubscription (Maybe Text)

esEventCategoriesList :: Lens' EventSubscription [Text]

esEnabled :: Lens' EventSubscription (Maybe Bool)

esCustomerAwsId :: Lens' EventSubscription (Maybe Text)

esCustSubscriptionId :: Lens' EventSubscription (Maybe Text)

hccHsmClientCertificatePublicKey :: Lens' HsmClientCertificate (Maybe Text)

hccHsmClientCertificateIdentifier :: Lens' HsmClientCertificate (Maybe Text)

hcHsmPartitionName :: Lens' HsmConfiguration (Maybe Text)

hcHsmIpAddress :: Lens' HsmConfiguration (Maybe Text)

hcHsmConfigurationIdentifier :: Lens' HsmConfiguration (Maybe Text)

hcDescription :: Lens' HsmConfiguration (Maybe Text)

hsStatus :: Lens' HsmStatus (Maybe Text)

hsHsmConfigurationIdentifier :: Lens' HsmStatus (Maybe Text)

hsHsmClientCertificateIdentifier :: Lens' HsmStatus (Maybe Text)

iprStatus :: Lens' IPRange (Maybe Text)

iprCIDRIP :: Lens' IPRange (Maybe Text)

ocoNodeType :: Lens' OrderableClusterOption (Maybe Text)

ocoClusterVersion :: Lens' OrderableClusterOption (Maybe Text)

ocoClusterType :: Lens' OrderableClusterOption (Maybe Text)

ocoAvailabilityZones :: Lens' OrderableClusterOption [AvailabilityZone]

prSource :: Lens' Parameter (Maybe Text)

prParameterValue :: Lens' Parameter (Maybe Text)

prParameterName :: Lens' Parameter (Maybe Text)

prMinimumEngineVersion :: Lens' Parameter (Maybe Text)

prIsModifiable :: Lens' Parameter (Maybe Bool)

prDescription :: Lens' Parameter (Maybe Text)

prDataType :: Lens' Parameter (Maybe Text)

prAllowedValues :: Lens' Parameter (Maybe Text)

pmvNumberOfNodes :: Lens' PendingModifiedValues (Maybe Integer)

pmvNodeType :: Lens' PendingModifiedValues (Maybe Text)

pmvMasterUserPassword :: Lens' PendingModifiedValues (Maybe Text)

pmvClusterVersion :: Lens' PendingModifiedValues (Maybe Text)

pmvClusterType :: Lens' PendingModifiedValues (Maybe Text)

pmvClusterIdentifier :: Lens' PendingModifiedValues (Maybe Text)

pmvAutomatedSnapshotRetentionPeriod :: Lens' PendingModifiedValues (Maybe Integer)

rcRecurringChargeFrequency :: Lens' RecurringCharge (Maybe Text)

rcRecurringChargeAmount :: Lens' RecurringCharge (Maybe Double)

rnUsagePrice :: Lens' ReservedNode (Maybe Double)

rnState :: Lens' ReservedNode (Maybe Text)

rnStartTime :: Lens' ReservedNode (Maybe ISO8601)

rnReservedNodeOfferingId :: Lens' ReservedNode (Maybe Text)

rnReservedNodeId :: Lens' ReservedNode (Maybe Text)

rnRecurringCharges :: Lens' ReservedNode [RecurringCharge]

rnOfferingType :: Lens' ReservedNode (Maybe Text)

rnNodeType :: Lens' ReservedNode (Maybe Text)

rnNodeCount :: Lens' ReservedNode (Maybe Integer)

rnFixedPrice :: Lens' ReservedNode (Maybe Double)

rnDuration :: Lens' ReservedNode (Maybe Integer)

rnCurrencyCode :: Lens' ReservedNode (Maybe Text)

rnoUsagePrice :: Lens' ReservedNodeOffering (Maybe Double)

rnoReservedNodeOfferingId :: Lens' ReservedNodeOffering (Maybe Text)

rnoRecurringCharges :: Lens' ReservedNodeOffering [RecurringCharge]

rnoOfferingType :: Lens' ReservedNodeOffering (Maybe Text)

rnoNodeType :: Lens' ReservedNodeOffering (Maybe Text)

rnoFixedPrice :: Lens' ReservedNodeOffering (Maybe Double)

rnoDuration :: Lens' ReservedNodeOffering (Maybe Integer)

rnoCurrencyCode :: Lens' ReservedNodeOffering (Maybe Text)

rsStatus :: Lens' RestoreStatus (Maybe Text)

rsSnapshotSizeInMegaBytes :: Lens' RestoreStatus (Maybe Integer)

rsProgressInMegaBytes :: Lens' RestoreStatus (Maybe Integer)

rsEstimatedTimeToCompletionInSeconds :: Lens' RestoreStatus (Maybe Integer)

rsElapsedTimeInSeconds :: Lens' RestoreStatus (Maybe Integer)

rsCurrentRestoreRateInMegaBytesPerSecond :: Lens' RestoreStatus (Maybe Double)

ssstVpcId :: Lens' Snapshot (Maybe Text)

ssstTotalBackupSizeInMegaBytes :: Lens' Snapshot (Maybe Double)

ssstStatus :: Lens' Snapshot (Maybe Text)

ssstSourceRegion :: Lens' Snapshot (Maybe Text)

ssstSnapshotType :: Lens' Snapshot (Maybe Text)

ssstSnapshotIdentifier :: Lens' Snapshot (Maybe Text)

ssstSnapshotCreateTime :: Lens' Snapshot (Maybe ISO8601)

ssstPort :: Lens' Snapshot (Maybe Integer)

ssstOwnerAccount :: Lens' Snapshot (Maybe Text)

ssstNumberOfNodes :: Lens' Snapshot (Maybe Integer)

ssstNodeType :: Lens' Snapshot (Maybe Text)

ssstMasterUsername :: Lens' Snapshot (Maybe Text)

ssstEstimatedSecondsToCompletion :: Lens' Snapshot (Maybe Integer)

ssstEncryptedWithHSM :: Lens' Snapshot (Maybe Bool)

ssstEncrypted :: Lens' Snapshot (Maybe Bool)

ssstElapsedTimeInSeconds :: Lens' Snapshot (Maybe Integer)

ssstDBName :: Lens' Snapshot (Maybe Text)

ssstCurrentBackupRateInMegaBytesPerSecond :: Lens' Snapshot (Maybe Double)

ssstClusterVersion :: Lens' Snapshot (Maybe Text)

ssstClusterIdentifier :: Lens' Snapshot (Maybe Text)

ssstClusterCreateTime :: Lens' Snapshot (Maybe ISO8601)

ssstBackupProgressInMegaBytes :: Lens' Snapshot (Maybe Double)

ssstAvailabilityZone :: Lens' Snapshot (Maybe Text)

ssstActualIncrementalBackupSizeInMegaBytes :: Lens' Snapshot (Maybe Double)

ssstAccountsWithRestoreAccess :: Lens' Snapshot [AccountWithRestoreAccess]

ssssssxSubnetStatus :: Lens' Subnet (Maybe Text)

ssssssxSubnetIdentifier :: Lens' Subnet (Maybe Text)

ssssssxSubnetAvailabilityZone :: Lens' Subnet (Maybe AvailabilityZone)

vsgmVpcSecurityGroupId :: Lens' VpcSecurityGroupMembership (Maybe Text)

vsgmStatus :: Lens' VpcSecurityGroupMembership (Maybe Text)