Network.AWS.Redshift.AuthorizeClusterSecurityGroupIngress

Description

Adds an inbound (ingress) rule to an Amazon Redshift security group. Depending on whether the application accessing your cluster is running on the Internet or an Amazon EC2 instance, you can authorize inbound access to either a Classless Interdomain Routing (CIDR)/Internet Protocol (IP) range or to an Amazon EC2 security group. You can add as many as 20 ingress rules to an Amazon Redshift security group.

If you authorize access to an Amazon EC2 security group, specify EC2SecurityGroupName and EC2SecurityGroupOwnerId . The Amazon EC2 security group and Amazon Redshift cluster must be in the same AWS region.

If you authorize access to a CIDRIP address range, specify CIDRIP . For an overview of CIDR blocks, see the Wikipedia article on <http:en.wikipedia.orgwiki/Classless_Inter-Domain_Routing Classless Inter-Domain Routing> .

You must also associate the security group with a cluster so that clients running on these IP addresses or the EC2 instance are authorized to connect to the cluster. For information about managing security groups, go to Working with Security Groups in the Amazon Redshift Cluster Management Guide .

Creating a Request

authorizeClusterSecurityGroupIngress

Arguments

:: Text
acsgiClusterSecurityGroupName

-> AuthorizeClusterSecurityGroupIngress

Creates a value of AuthorizeClusterSecurityGroupIngress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acsgiEC2SecurityGroupOwnerId - The AWS account number of the owner of the security group specified by the EC2SecurityGroupName parameter. The AWS Access Key ID is not an acceptable value. Example: 111122223333

acsgiEC2SecurityGroupName - The EC2 security group to be added the Amazon Redshift security group.

acsgiCIdRIP - The IP range to be added the Amazon Redshift security group.

acsgiClusterSecurityGroupName - The name of the security group to which the ingress rule is added.

data AuthorizeClusterSecurityGroupIngress

See: authorizeClusterSecurityGroupIngress smart constructor.

Request Lenses

acsgiEC2SecurityGroupOwnerId :: Lens' AuthorizeClusterSecurityGroupIngress (Maybe Text)

The AWS account number of the owner of the security group specified by the EC2SecurityGroupName parameter. The AWS Access Key ID is not an acceptable value. Example: 111122223333

acsgiEC2SecurityGroupName :: Lens' AuthorizeClusterSecurityGroupIngress (Maybe Text)

The EC2 security group to be added the Amazon Redshift security group.

acsgiCIdRIP :: Lens' AuthorizeClusterSecurityGroupIngress (Maybe Text)

The IP range to be added the Amazon Redshift security group.

acsgiClusterSecurityGroupName :: Lens' AuthorizeClusterSecurityGroupIngress Text

The name of the security group to which the ingress rule is added.

Destructuring the Response

authorizeClusterSecurityGroupIngressResponse

Arguments

:: Int
acsgirsResponseStatus

-> AuthorizeClusterSecurityGroupIngressResponse

Creates a value of AuthorizeClusterSecurityGroupIngressResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acsgirsClusterSecurityGroup - Undocumented member.

acsgirsResponseStatus - -- | The response status code.

data AuthorizeClusterSecurityGroupIngressResponse

See: authorizeClusterSecurityGroupIngressResponse smart constructor.

Response Lenses

acsgirsClusterSecurityGroup :: Lens' AuthorizeClusterSecurityGroupIngressResponse (Maybe ClusterSecurityGroup)

Undocumented member.

acsgirsResponseStatus :: Lens' AuthorizeClusterSecurityGroupIngressResponse Int

- | The response status code.