Network.AWS.Redshift.CopyClusterSnapshot

Description

Copies the specified automated cluster snapshot to a new manual cluster snapshot. The source must be an automated snapshot and it must be in the available state.

When you delete a cluster, Amazon Redshift deletes any automated snapshots of the cluster. Also, when the retention period of the snapshot expires, Amazon Redshift automatically deletes it. If you want to keep an automated snapshot for a longer period, you can make a manual copy of the snapshot. Manual snapshots are retained until you delete them.

For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Cluster Management Guide .

Creating a Request

copyClusterSnapshot

Arguments

:: Text
ccsSourceSnapshotIdentifier

-> Text
ccsTargetSnapshotIdentifier

-> CopyClusterSnapshot

Creates a value of CopyClusterSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccsSourceSnapshotClusterIdentifier - The identifier of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name. Constraints: * Must be the identifier for a valid cluster.

ccsSourceSnapshotIdentifier - The identifier for the source snapshot. Constraints: * Must be the identifier for a valid automated snapshot whose state is available .

ccsTargetSnapshotIdentifier - The identifier given to the new manual snapshot. Constraints: * Cannot be null, empty, or blank. * Must contain from 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. * Must be unique for the AWS account that is making the request.

data CopyClusterSnapshot

See: copyClusterSnapshot smart constructor.

Request Lenses

ccsSourceSnapshotClusterIdentifier :: Lens' CopyClusterSnapshot (Maybe Text)

The identifier of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name. Constraints: * Must be the identifier for a valid cluster.

ccsSourceSnapshotIdentifier :: Lens' CopyClusterSnapshot Text

The identifier for the source snapshot. Constraints: * Must be the identifier for a valid automated snapshot whose state is available .

ccsTargetSnapshotIdentifier :: Lens' CopyClusterSnapshot Text

The identifier given to the new manual snapshot. Constraints: * Cannot be null, empty, or blank. * Must contain from 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. * Must be unique for the AWS account that is making the request.

Destructuring the Response

copyClusterSnapshotResponse

Arguments

:: Int
ccsrsResponseStatus

-> CopyClusterSnapshotResponse

Creates a value of CopyClusterSnapshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccsrsSnapshot - Undocumented member.

ccsrsResponseStatus - -- | The response status code.

data CopyClusterSnapshotResponse

See: copyClusterSnapshotResponse smart constructor.

Response Lenses

ccsrsSnapshot :: Lens' CopyClusterSnapshotResponse (Maybe Snapshot)

Undocumented member.

ccsrsResponseStatus :: Lens' CopyClusterSnapshotResponse Int

- | The response status code.