Network.AWS.Redshift.CopyClusterSnapshot

Description

Copies the specified automated cluster snapshot to a new manual cluster snapshot. The source must be an automated snapshot and it must be in the available state.

When you delete a cluster, Amazon Redshift deletes any automated snapshots of the cluster. Also, when the retention period of the snapshot expires, Amazon Redshift automatically deletes it. If you want to keep an automated snapshot for a longer period, you can make a manual copy of the snapshot. Manual snapshots are retained until you delete them.

For more information about working with snapshots, go to Amazon RedshiftSnapshots in the Amazon Redshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_CopyClusterSnapshot.html

Request

data CopyClusterSnapshot

Request constructor

copyClusterSnapshot

Arguments

:: Text
ccsSourceSnapshotIdentifier

-> Text
ccsTargetSnapshotIdentifier

-> CopyClusterSnapshot

CopyClusterSnapshot constructor.

The fields accessible through corresponding lenses are:

ccsSourceSnapshotClusterIdentifier :: Maybe Text

ccsSourceSnapshotIdentifier :: Text

ccsTargetSnapshotIdentifier :: Text

Request lenses

ccsSourceSnapshotClusterIdentifier :: Lens' CopyClusterSnapshot (Maybe Text)

The identifier of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

Constraints:

Must be the identifier for a valid cluster.

ccsSourceSnapshotIdentifier :: Lens' CopyClusterSnapshot Text

The identifier for the source snapshot.

Constraints:

Must be the identifier for a valid automated snapshot whose state is available.

ccsTargetSnapshotIdentifier :: Lens' CopyClusterSnapshot Text

The identifier given to the new manual snapshot.

Constraints:

Cannot be null, empty, or blank. Must contain from 1 to 255 alphanumeric characters or hyphens. First character must be a letter. Cannot end with a hyphen or contain two consecutive hyphens. Must be unique for the AWS account that is making the request.

Response

data CopyClusterSnapshotResponse

Response constructor

copyClusterSnapshotResponse :: CopyClusterSnapshotResponse

CopyClusterSnapshotResponse constructor.

The fields accessible through corresponding lenses are:

ccsrSnapshot :: Maybe Snapshot

Response lenses

ccsrSnapshot :: Lens' CopyClusterSnapshotResponse (Maybe Snapshot)