Network.AWS.Redshift.CreateCluster

Description

Creates a new cluster.

To create the cluster in Virtual Private Cloud (VPC), you must provide a cluster subnet group name. The cluster subnet group identifies the subnets of your VPC that Amazon Redshift uses when creating the cluster. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Cluster Management Guide .

Creating a Request

createCluster

Arguments

:: Text
ccClusterIdentifier

-> Text
ccNodeType

-> Text
ccMasterUsername

-> Text
ccMasterUserPassword

-> CreateCluster

Creates a value of CreateCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccEnhancedVPCRouting - An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide. If this option is true , enhanced VPC routing is enabled. Default: false

ccAdditionalInfo - Reserved.

ccPubliclyAccessible - If true , the cluster can be accessed from a public network.

ccHSMConfigurationIdentifier - Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

ccClusterSecurityGroups - A list of security groups to be associated with this cluster. Default: The default cluster security group for Amazon Redshift.

ccAutomatedSnapshotRetentionPeriod - The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot . Default: 1 Constraints: Must be a value from 0 to 35.

ccEncrypted - If true , the data in the cluster is encrypted at rest. Default: false

ccClusterSubnetGroupName - The name of a cluster subnet group to be associated with this cluster. If this parameter is not provided the resulting cluster will be deployed outside virtual private cloud (VPC).

ccHSMClientCertificateIdentifier - Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

ccNumberOfNodes - The number of compute nodes in the cluster. This parameter is required when the ClusterType parameter is specified as multi-node . For information about determining how many nodes you need, go to Working with Clusters in the Amazon Redshift Cluster Management Guide . If you don't specify this parameter, you get a single-node cluster. When requesting a multi-node cluster, you must specify the number of nodes that you want in the cluster. Default: 1 Constraints: Value must be at least 1 and no more than 100.

ccElasticIP - The Elastic IP (EIP) address for the cluster. Constraints: The cluster must be provisioned in EC2-VPC and publicly-accessible through an Internet gateway. For more information about provisioning clusters in EC2-VPC, go to Supported Platforms to Launch Your Cluster in the Amazon Redshift Cluster Management Guide.

ccPreferredMaintenanceWindow - The weekly time range (in UTC) during which automated cluster maintenance can occur. Format: ddd:hh24:mi-ddd:hh24:mi Default: A 30-minute window selected at random from an 8-hour block of time per region, occurring on a random day of the week. For more information about the time blocks for each region, see Maintenance Windows in Amazon Redshift Cluster Management Guide. Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun Constraints: Minimum 30-minute window.

ccKMSKeyId - The AWS Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster.

ccAvailabilityZone - The EC2 Availability Zone (AZ) in which you want Amazon Redshift to provision the cluster. For example, if you have several EC2 instances running in a specific Availability Zone, then you might want the cluster to be provisioned in the same zone in order to decrease network latency. Default: A random, system-chosen Availability Zone in the region that is specified by the endpoint. Example: us-east-1d Constraint: The specified Availability Zone must be in the same region as the current endpoint.

ccVPCSecurityGroupIds - A list of Virtual Private Cloud (VPC) security groups to be associated with the cluster. Default: The default VPC security group is associated with the cluster.

ccIAMRoles - A list of AWS Identity and Access Management (IAM) roles that can be used by the cluster to access other AWS services. You must supply the IAM roles in their Amazon Resource Name (ARN) format. You can supply up to 10 IAM roles in a single request. A cluster can have up to 10 IAM roles associated with it at any time.

ccClusterType - The type of the cluster. When cluster type is specified as * single-node , the NumberOfNodes parameter is not required. * multi-node , the NumberOfNodes parameter is required. Valid Values: multi-node | single-node Default: multi-node

ccClusterVersion - The version of the Amazon Redshift engine software that you want to deploy on the cluster. The version selected runs on all the nodes in the cluster. Constraints: Only version 1.0 is currently available. Example: 1.0

ccAllowVersionUpgrade - If true , major version upgrades can be applied during the maintenance window to the Amazon Redshift engine that is running on the cluster. When a new major version of the Amazon Redshift engine is released, you can request that the service automatically apply upgrades during the maintenance window to the Amazon Redshift engine that is running on your cluster. Default: true

ccClusterParameterGroupName - The name of the parameter group to be associated with this cluster. Default: The default Amazon Redshift cluster parameter group. For information about the default parameter group, go to Working with Amazon Redshift Parameter Groups Constraints: * Must be 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

ccTags - A list of tag instances.

ccPort - The port number on which the cluster accepts incoming connections. The cluster is accessible only via the JDBC and ODBC connection strings. Part of the connection string requires the port on which the cluster will listen for incoming connections. Default: 5439 Valid Values: 1150-65535

ccDBName - The name of the first database to be created when the cluster is created. To create additional databases after the cluster is created, connect to the cluster with a SQL client and use SQL commands to create a database. For more information, go to Create a Database in the Amazon Redshift Database Developer Guide. Default: dev Constraints: * Must contain 1 to 64 alphanumeric characters. * Must contain only lowercase letters. * Cannot be a word that is reserved by the service. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

ccClusterIdentifier - A unique identifier for the cluster. You use this identifier to refer to the cluster for any subsequent cluster operations such as deleting or modifying. The identifier also appears in the Amazon Redshift console. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * Alphabetic characters must be lowercase. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. * Must be unique for all clusters within an AWS account. Example: myexamplecluster

ccNodeType - The node type to be provisioned for the cluster. For information about node types, go to Working with Clusters in the Amazon Redshift Cluster Management Guide . Valid Values: ds1.xlarge | ds1.8xlarge | ds2.xlarge | ds2.8xlarge | dc1.large | dc1.8xlarge .

ccMasterUsername - The user name associated with the master user account for the cluster that is being created. Constraints: * Must be 1 - 128 alphanumeric characters. * First character must be a letter. * Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

ccMasterUserPassword - The password associated with the master user account for the cluster that is being created. Constraints: * Must be between 8 and 64 characters in length. * Must contain at least one uppercase letter. * Must contain at least one lowercase letter. * Must contain one number. * Can be any printable ASCII character (ASCII code 33 to 126) except ' (single quote), " (double quote), , /, @, or space.

data CreateCluster

See: createCluster smart constructor.

Request Lenses

ccEnhancedVPCRouting :: Lens' CreateCluster (Maybe Bool)

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide. If this option is true , enhanced VPC routing is enabled. Default: false

ccAdditionalInfo :: Lens' CreateCluster (Maybe Text)

Reserved.

ccPubliclyAccessible :: Lens' CreateCluster (Maybe Bool)

If true , the cluster can be accessed from a public network.

ccHSMConfigurationIdentifier :: Lens' CreateCluster (Maybe Text)

Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

ccClusterSecurityGroups :: Lens' CreateCluster [Text]

A list of security groups to be associated with this cluster. Default: The default cluster security group for Amazon Redshift.

ccAutomatedSnapshotRetentionPeriod :: Lens' CreateCluster (Maybe Int)

The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot . Default: 1 Constraints: Must be a value from 0 to 35.

ccEncrypted :: Lens' CreateCluster (Maybe Bool)

If true , the data in the cluster is encrypted at rest. Default: false

ccClusterSubnetGroupName :: Lens' CreateCluster (Maybe Text)

The name of a cluster subnet group to be associated with this cluster. If this parameter is not provided the resulting cluster will be deployed outside virtual private cloud (VPC).

ccHSMClientCertificateIdentifier :: Lens' CreateCluster (Maybe Text)

Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

ccNumberOfNodes :: Lens' CreateCluster (Maybe Int)

The number of compute nodes in the cluster. This parameter is required when the ClusterType parameter is specified as multi-node . For information about determining how many nodes you need, go to Working with Clusters in the Amazon Redshift Cluster Management Guide . If you don't specify this parameter, you get a single-node cluster. When requesting a multi-node cluster, you must specify the number of nodes that you want in the cluster. Default: 1 Constraints: Value must be at least 1 and no more than 100.

ccElasticIP :: Lens' CreateCluster (Maybe Text)

The Elastic IP (EIP) address for the cluster. Constraints: The cluster must be provisioned in EC2-VPC and publicly-accessible through an Internet gateway. For more information about provisioning clusters in EC2-VPC, go to Supported Platforms to Launch Your Cluster in the Amazon Redshift Cluster Management Guide.

ccPreferredMaintenanceWindow :: Lens' CreateCluster (Maybe Text)

The weekly time range (in UTC) during which automated cluster maintenance can occur. Format: ddd:hh24:mi-ddd:hh24:mi Default: A 30-minute window selected at random from an 8-hour block of time per region, occurring on a random day of the week. For more information about the time blocks for each region, see Maintenance Windows in Amazon Redshift Cluster Management Guide. Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun Constraints: Minimum 30-minute window.

ccKMSKeyId :: Lens' CreateCluster (Maybe Text)

The AWS Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster.

ccAvailabilityZone :: Lens' CreateCluster (Maybe Text)

The EC2 Availability Zone (AZ) in which you want Amazon Redshift to provision the cluster. For example, if you have several EC2 instances running in a specific Availability Zone, then you might want the cluster to be provisioned in the same zone in order to decrease network latency. Default: A random, system-chosen Availability Zone in the region that is specified by the endpoint. Example: us-east-1d Constraint: The specified Availability Zone must be in the same region as the current endpoint.

ccVPCSecurityGroupIds :: Lens' CreateCluster [Text]

A list of Virtual Private Cloud (VPC) security groups to be associated with the cluster. Default: The default VPC security group is associated with the cluster.

ccIAMRoles :: Lens' CreateCluster [Text]

A list of AWS Identity and Access Management (IAM) roles that can be used by the cluster to access other AWS services. You must supply the IAM roles in their Amazon Resource Name (ARN) format. You can supply up to 10 IAM roles in a single request. A cluster can have up to 10 IAM roles associated with it at any time.

ccClusterType :: Lens' CreateCluster (Maybe Text)

The type of the cluster. When cluster type is specified as * single-node , the NumberOfNodes parameter is not required. * multi-node , the NumberOfNodes parameter is required. Valid Values: multi-node | single-node Default: multi-node

ccClusterVersion :: Lens' CreateCluster (Maybe Text)

The version of the Amazon Redshift engine software that you want to deploy on the cluster. The version selected runs on all the nodes in the cluster. Constraints: Only version 1.0 is currently available. Example: 1.0

ccAllowVersionUpgrade :: Lens' CreateCluster (Maybe Bool)

If true , major version upgrades can be applied during the maintenance window to the Amazon Redshift engine that is running on the cluster. When a new major version of the Amazon Redshift engine is released, you can request that the service automatically apply upgrades during the maintenance window to the Amazon Redshift engine that is running on your cluster. Default: true

ccClusterParameterGroupName :: Lens' CreateCluster (Maybe Text)

The name of the parameter group to be associated with this cluster. Default: The default Amazon Redshift cluster parameter group. For information about the default parameter group, go to Working with Amazon Redshift Parameter Groups Constraints: * Must be 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

ccTags :: Lens' CreateCluster [Tag]

A list of tag instances.

ccPort :: Lens' CreateCluster (Maybe Int)

The port number on which the cluster accepts incoming connections. The cluster is accessible only via the JDBC and ODBC connection strings. Part of the connection string requires the port on which the cluster will listen for incoming connections. Default: 5439 Valid Values: 1150-65535

ccDBName :: Lens' CreateCluster (Maybe Text)

The name of the first database to be created when the cluster is created. To create additional databases after the cluster is created, connect to the cluster with a SQL client and use SQL commands to create a database. For more information, go to Create a Database in the Amazon Redshift Database Developer Guide. Default: dev Constraints: * Must contain 1 to 64 alphanumeric characters. * Must contain only lowercase letters. * Cannot be a word that is reserved by the service. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

ccClusterIdentifier :: Lens' CreateCluster Text

A unique identifier for the cluster. You use this identifier to refer to the cluster for any subsequent cluster operations such as deleting or modifying. The identifier also appears in the Amazon Redshift console. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * Alphabetic characters must be lowercase. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. * Must be unique for all clusters within an AWS account. Example: myexamplecluster

ccNodeType :: Lens' CreateCluster Text

The node type to be provisioned for the cluster. For information about node types, go to Working with Clusters in the Amazon Redshift Cluster Management Guide . Valid Values: ds1.xlarge | ds1.8xlarge | ds2.xlarge | ds2.8xlarge | dc1.large | dc1.8xlarge .

ccMasterUsername :: Lens' CreateCluster Text

The user name associated with the master user account for the cluster that is being created. Constraints: * Must be 1 - 128 alphanumeric characters. * First character must be a letter. * Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

ccMasterUserPassword :: Lens' CreateCluster Text

The password associated with the master user account for the cluster that is being created. Constraints: * Must be between 8 and 64 characters in length. * Must contain at least one uppercase letter. * Must contain at least one lowercase letter. * Must contain one number. * Can be any printable ASCII character (ASCII code 33 to 126) except ' (single quote), " (double quote), , /, @, or space.

Destructuring the Response

createClusterResponse

Arguments

:: Int
ccrsResponseStatus

-> CreateClusterResponse

Creates a value of CreateClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccrsCluster - Undocumented member.

ccrsResponseStatus - -- | The response status code.

data CreateClusterResponse

See: createClusterResponse smart constructor.

Response Lenses

ccrsCluster :: Lens' CreateClusterResponse (Maybe Cluster)

Undocumented member.

ccrsResponseStatus :: Lens' CreateClusterResponse Int

- | The response status code.