Network.AWS.Redshift.CreateClusterParameterGroup

Description

Creates an Amazon Redshift parameter group.

Creating parameter groups is independent of creating clusters. You can associate a cluster with a parameter group when you create the cluster. You can also associate an existing cluster with a parameter group after the cluster is created by using ModifyCluster .

Parameters in the parameter group define specific behavior that applies to the databases you create on the cluster. For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide .

Creating a Request

createClusterParameterGroup

Arguments

:: Text
ccpgParameterGroupName

-> Text
ccpgParameterGroupFamily

-> Text
ccpgDescription

-> CreateClusterParameterGroup

Creates a value of CreateClusterParameterGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccpgTags - A list of tag instances.

ccpgParameterGroupName - The name of the cluster parameter group. Constraints: * Must be 1 to 255 alphanumeric characters or hyphens * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. * Must be unique withing your AWS account.

ccpgParameterGroupFamily - The Amazon Redshift engine version to which the cluster parameter group applies. The cluster engine version determines the set of parameters. To get a list of valid parameter group family names, you can call DescribeClusterParameterGroups . By default, Amazon Redshift returns a list of all the parameter groups that are owned by your AWS account, including the default parameter groups for each Amazon Redshift engine version. The parameter group family names associated with the default parameter groups provide you the valid values. For example, a valid family name is "redshift-1.0".

ccpgDescription - A description of the parameter group.

data CreateClusterParameterGroup

See: createClusterParameterGroup smart constructor.

Request Lenses

ccpgTags :: Lens' CreateClusterParameterGroup [Tag]

A list of tag instances.

ccpgParameterGroupName :: Lens' CreateClusterParameterGroup Text

The name of the cluster parameter group. Constraints: * Must be 1 to 255 alphanumeric characters or hyphens * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. * Must be unique withing your AWS account.

ccpgParameterGroupFamily :: Lens' CreateClusterParameterGroup Text

The Amazon Redshift engine version to which the cluster parameter group applies. The cluster engine version determines the set of parameters. To get a list of valid parameter group family names, you can call DescribeClusterParameterGroups . By default, Amazon Redshift returns a list of all the parameter groups that are owned by your AWS account, including the default parameter groups for each Amazon Redshift engine version. The parameter group family names associated with the default parameter groups provide you the valid values. For example, a valid family name is "redshift-1.0".

ccpgDescription :: Lens' CreateClusterParameterGroup Text

A description of the parameter group.

Destructuring the Response

createClusterParameterGroupResponse

Arguments

:: Int
ccpgrsResponseStatus

-> CreateClusterParameterGroupResponse

Creates a value of CreateClusterParameterGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccpgrsClusterParameterGroup - Undocumented member.

ccpgrsResponseStatus - -- | The response status code.

data CreateClusterParameterGroupResponse

See: createClusterParameterGroupResponse smart constructor.

Response Lenses

ccpgrsClusterParameterGroup :: Lens' CreateClusterParameterGroupResponse (Maybe ClusterParameterGroup)

Undocumented member.

ccpgrsResponseStatus :: Lens' CreateClusterParameterGroupResponse Int

- | The response status code.