Network.AWS.Redshift.CreateClusterParameterGroup

Description

Creates an Amazon Redshift parameter group.

Creating parameter groups is independent of creating clusters. You can associate a cluster with a parameter group when you create the cluster. You can also associate an existing cluster with a parameter group after the cluster is created by using ModifyCluster.

Parameters in the parameter group define specific behavior that applies to the databases you create on the cluster. For more information about managing parameter groups, go to Amazon Redshift Parameter Groups in the AmazonRedshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_CreateClusterParameterGroup.html

Request

data CreateClusterParameterGroup

Request constructor

createClusterParameterGroup

Arguments

:: Text
ccpgParameterGroupName

-> Text
ccpgParameterGroupFamily

-> Text
ccpgDescription

-> CreateClusterParameterGroup

CreateClusterParameterGroup constructor.

The fields accessible through corresponding lenses are:

ccpgDescription :: Text

ccpgParameterGroupFamily :: Text

ccpgParameterGroupName :: Text

ccpgTags :: [Tag]

Request lenses

ccpgDescription :: Lens' CreateClusterParameterGroup Text

A description of the parameter group.

ccpgParameterGroupFamily :: Lens' CreateClusterParameterGroup Text

The Amazon Redshift engine version to which the cluster parameter group applies. The cluster engine version determines the set of parameters.

To get a list of valid parameter group family names, you can call DescribeClusterParameterGroups. By default, Amazon Redshift returns a list of all the parameter groups that are owned by your AWS account, including the default parameter groups for each Amazon Redshift engine version. The parameter group family names associated with the default parameter groups provide you the valid values. For example, a valid family name is "redshift-1.0".

ccpgParameterGroupName :: Lens' CreateClusterParameterGroup Text

The name of the cluster parameter group.

Constraints:

Must be 1 to 255 alphanumeric characters or hyphens First character must be a letter. Cannot end with a hyphen or contain two consecutive hyphens. Must be unique withing your AWS account. This value is stored as a lower-case string.

ccpgTags :: Lens' CreateClusterParameterGroup [Tag]

A list of tag instances.

Response

data CreateClusterParameterGroupResponse

Response constructor

createClusterParameterGroupResponse :: CreateClusterParameterGroupResponse

CreateClusterParameterGroupResponse constructor.

The fields accessible through corresponding lenses are:

ccpgrClusterParameterGroup :: Maybe ClusterParameterGroup

Response lenses

ccpgrClusterParameterGroup :: Lens' CreateClusterParameterGroupResponse (Maybe ClusterParameterGroup)