Network.AWS.Redshift.CreateEventSubscription

Description

Creates an Amazon Redshift event notification subscription. This action requires an ARN (Amazon Resource Name) of an Amazon SNS topic created by either the Amazon Redshift console, the Amazon SNS console, or the Amazon SNS API. To obtain an ARN with Amazon SNS, you must create a topic in Amazon SNS and subscribe to the topic. The ARN is displayed in the SNS console.

You can specify the source type, and lists of Amazon Redshift source IDs, event categories, and event severities. Notifications will be sent for all events you want that match those criteria. For example, you can specify source type = cluster, source ID = my-cluster-1 and mycluster2, event categories = Availability, Backup, and severity = ERROR. The subscription will only send notifications for those ERROR events in the Availability and Backup categories for the specified clusters.

If you specify both the source type and source IDs, such as source type = cluster and source identifier = my-cluster-1, notifications will be sent for all the cluster events for my-cluster-1. If you specify a source type but do not specify a source identifier, you will receive notice of the events for the objects of that type in your AWS account. If you do not specify either the SourceType nor the SourceIdentifier, you will be notified of events generated from all Amazon Redshift sources belonging to your AWS account. You must specify a source type if you specify a source ID.

Creating a Request

createEventSubscription

Arguments

:: Text
cesSubscriptionName

-> Text
cesSNSTopicARN

-> CreateEventSubscription

Creates a value of CreateEventSubscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cesEnabled - A Boolean value; set to true to activate the subscription, set to false to create the subscription but not active it.

cesSourceType - The type of source that will be generating the events. For example, if you want to be notified of events generated by a cluster, you would set this parameter to cluster. If this value is not specified, events are returned for all Amazon Redshift objects in your AWS account. You must specify a source type in order to specify source IDs. Valid values: cluster, cluster-parameter-group, cluster-security-group, and cluster-snapshot.

cesSeverity - Specifies the Amazon Redshift event severity to be published by the event notification subscription. Values: ERROR, INFO

cesEventCategories - Specifies the Amazon Redshift event categories to be published by the event notification subscription. Values: Configuration, Management, Monitoring, Security

cesSourceIds - A list of one or more identifiers of Amazon Redshift source objects. All of the objects must be of the same type as was specified in the source type parameter. The event subscription will return only events generated by the specified objects. If not specified, then events are returned for all objects within the source type specified. Example: my-cluster-1, my-cluster-2 Example: my-snapshot-20131010

cesTags - A list of tag instances.

cesSubscriptionName - The name of the event subscription to be created. Constraints: * Cannot be null, empty, or blank. * Must contain from 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

cesSNSTopicARN - The Amazon Resource Name (ARN) of the Amazon SNS topic used to transmit the event notifications. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

data CreateEventSubscription

See: createEventSubscription smart constructor.

Request Lenses

cesEnabled :: Lens' CreateEventSubscription (Maybe Bool)

A Boolean value; set to true to activate the subscription, set to false to create the subscription but not active it.

cesSourceType :: Lens' CreateEventSubscription (Maybe Text)

The type of source that will be generating the events. For example, if you want to be notified of events generated by a cluster, you would set this parameter to cluster. If this value is not specified, events are returned for all Amazon Redshift objects in your AWS account. You must specify a source type in order to specify source IDs. Valid values: cluster, cluster-parameter-group, cluster-security-group, and cluster-snapshot.

cesSeverity :: Lens' CreateEventSubscription (Maybe Text)

Specifies the Amazon Redshift event severity to be published by the event notification subscription. Values: ERROR, INFO

cesEventCategories :: Lens' CreateEventSubscription [Text]

Specifies the Amazon Redshift event categories to be published by the event notification subscription. Values: Configuration, Management, Monitoring, Security

cesSourceIds :: Lens' CreateEventSubscription [Text]

A list of one or more identifiers of Amazon Redshift source objects. All of the objects must be of the same type as was specified in the source type parameter. The event subscription will return only events generated by the specified objects. If not specified, then events are returned for all objects within the source type specified. Example: my-cluster-1, my-cluster-2 Example: my-snapshot-20131010

cesTags :: Lens' CreateEventSubscription [Tag]

A list of tag instances.

cesSubscriptionName :: Lens' CreateEventSubscription Text

The name of the event subscription to be created. Constraints: * Cannot be null, empty, or blank. * Must contain from 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

cesSNSTopicARN :: Lens' CreateEventSubscription Text

The Amazon Resource Name (ARN) of the Amazon SNS topic used to transmit the event notifications. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

Destructuring the Response

createEventSubscriptionResponse

Arguments

:: Int
cesrsResponseStatus

-> CreateEventSubscriptionResponse

Creates a value of CreateEventSubscriptionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cesrsEventSubscription - Undocumented member.

cesrsResponseStatus - -- | The response status code.

data CreateEventSubscriptionResponse

See: createEventSubscriptionResponse smart constructor.

Response Lenses

cesrsEventSubscription :: Lens' CreateEventSubscriptionResponse (Maybe EventSubscription)

Undocumented member.

cesrsResponseStatus :: Lens' CreateEventSubscriptionResponse Int

- | The response status code.