Network.AWS.Redshift.CreateHsmConfiguration

Description

Creates an HSM configuration that contains the information required by an Amazon Redshift cluster to store and use database encryption keys in a Hardware Security Module (HSM). After creating the HSM configuration, you can specify it as a parameter when creating a cluster. The cluster will then store its encryption keys in the HSM.

In addition to creating an HSM configuration, you must also create an HSM client certificate. For more information, go to Hardware Security Modules in the Amazon Redshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_CreateHsmConfiguration.html

Request

data CreateHsmConfiguration

Request constructor

createHsmConfiguration

Arguments

:: Text
chcHsmConfigurationIdentifier

-> Text
chcDescription

-> Text
chcHsmIpAddress

-> Text
chcHsmPartitionName

-> Text
chcHsmPartitionPassword

-> Text
chcHsmServerPublicCertificate

-> CreateHsmConfiguration

CreateHsmConfiguration constructor.

The fields accessible through corresponding lenses are:

chcDescription :: Text

chcHsmConfigurationIdentifier :: Text

chcHsmIpAddress :: Text

chcHsmPartitionName :: Text

chcHsmPartitionPassword :: Text

chcHsmServerPublicCertificate :: Text

chcTags :: [Tag]

Request lenses

chcDescription :: Lens' CreateHsmConfiguration Text

A text description of the HSM configuration to be created.

chcHsmConfigurationIdentifier :: Lens' CreateHsmConfiguration Text

The identifier to be assigned to the new Amazon Redshift HSM configuration.

chcHsmIpAddress :: Lens' CreateHsmConfiguration Text

The IP address that the Amazon Redshift cluster must use to access the HSM.

chcHsmPartitionName :: Lens' CreateHsmConfiguration Text

The name of the partition in the HSM where the Amazon Redshift clusters will store their database encryption keys.

chcHsmPartitionPassword :: Lens' CreateHsmConfiguration Text

The password required to access the HSM partition.

chcHsmServerPublicCertificate :: Lens' CreateHsmConfiguration Text

The HSMs public certificate file. When using Cloud HSM, the file name is server.pem.

chcTags :: Lens' CreateHsmConfiguration [Tag]

A list of tag instances.

Response

data CreateHsmConfigurationResponse

Response constructor

createHsmConfigurationResponse :: CreateHsmConfigurationResponse

CreateHsmConfigurationResponse constructor.

The fields accessible through corresponding lenses are:

chcrHsmConfiguration :: Maybe HsmConfiguration

Response lenses

chcrHsmConfiguration :: Lens' CreateHsmConfigurationResponse (Maybe HsmConfiguration)