Network.AWS.Redshift.CreateTags

Description

Adds one or more tags to a specified resource.

A resource can have up to 10 tags. If you try to create more than 10 tags for a resource, you will receive an error and the attempt will fail.

If you specify a key that already exists for the resource, the value for that key will be updated with the new value.

Creating a Request

createTags

Arguments

:: Text
ctResourceName

-> CreateTags

Creates a value of CreateTags with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctResourceName - The Amazon Resource Name (ARN) to which you want to add the tag or tags. For example, arn:aws:redshift:us-east-1:123456789:cluster:t1 .

ctTags - One or more name/value pairs to add as tags to the specified resource. Each tag name is passed in with the parameter Key and the corresponding value is passed in with the parameter Value . The Key and Value parameters are separated by a comma (,). Separate multiple tags with a space. For example, --tags Key="owner",Value="admin" Key="environment",Value="test" Key="version",Value="1.0" .

data CreateTags

Contains the output from the CreateTags action.

See: createTags smart constructor.

Request Lenses

ctResourceName :: Lens' CreateTags Text

The Amazon Resource Name (ARN) to which you want to add the tag or tags. For example, arn:aws:redshift:us-east-1:123456789:cluster:t1 .

ctTags :: Lens' CreateTags [Tag]

One or more name/value pairs to add as tags to the specified resource. Each tag name is passed in with the parameter Key and the corresponding value is passed in with the parameter Value . The Key and Value parameters are separated by a comma (,). Separate multiple tags with a space. For example, --tags Key="owner",Value="admin" Key="environment",Value="test" Key="version",Value="1.0" .

Destructuring the Response

createTagsResponse :: CreateTagsResponse

Creates a value of CreateTagsResponse with the minimum fields required to make a request.

data CreateTagsResponse

See: createTagsResponse smart constructor.