Network.AWS.Redshift.DescribeClusterParameterGroups

Description

Returns a list of Amazon Redshift parameter groups, including parameter groups you created and the default parameter group. For each parameter group, the response includes the parameter group name, description, and parameter group family name. You can optionally specify a name to retrieve the description of a specific parameter group.

For more information about managing parameter groups, go to Amazon RedshiftParameter Groups in the Amazon Redshift Cluster Management Guide.

If you specify both tag keys and tag values in the same request, Amazon Redshift returns all parameter groups that match any combination of the specified keys and values. For example, if you have owner and environment for tag keys, and admin and test for tag values, all parameter groups that have any combination of those values are returned.

If both tag keys and values are omitted from the request, parameter groups are returned regardless of whether they have tag keys or values associated with them.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeClusterParameterGroups.html

Request

data DescribeClusterParameterGroups

Request constructor

describeClusterParameterGroups :: DescribeClusterParameterGroups

DescribeClusterParameterGroups constructor.

The fields accessible through corresponding lenses are:

dcpgMarker :: Maybe Text

dcpgMaxRecords :: Maybe Int

dcpgParameterGroupName :: Maybe Text

dcpgTagKeys :: [Text]

dcpgTagValues :: [Text]

Request lenses

dcpgMarker :: Lens' DescribeClusterParameterGroups (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterParameterGroups request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dcpgMaxRecords :: Lens' DescribeClusterParameterGroups (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: '100'

Constraints: minimum 20, maximum 100.

dcpgParameterGroupName :: Lens' DescribeClusterParameterGroups (Maybe Text)

The name of a specific parameter group for which to return details. By default, details about all parameter groups and the default parameter group are returned.

dcpgTagKeys :: Lens' DescribeClusterParameterGroups [Text]

A tag key or keys for which you want to return all matching cluster parameter groups that are associated with the specified key or keys. For example, suppose that you have parameter groups that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the parameter groups that have either or both of these tag keys associated with them.

dcpgTagValues :: Lens' DescribeClusterParameterGroups [Text]

A tag value or values for which you want to return all matching cluster parameter groups that are associated with the specified tag value or values. For example, suppose that you have parameter groups that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the parameter groups that have either or both of these tag values associated with them.

Response

data DescribeClusterParameterGroupsResponse

Response constructor

describeClusterParameterGroupsResponse :: DescribeClusterParameterGroupsResponse

DescribeClusterParameterGroupsResponse constructor.

The fields accessible through corresponding lenses are:

dcpgrMarker :: Maybe Text

dcpgrParameterGroups :: [ClusterParameterGroup]

Response lenses

dcpgrMarker :: Lens' DescribeClusterParameterGroupsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcpgrParameterGroups :: Lens' DescribeClusterParameterGroupsResponse [ClusterParameterGroup]

A list of ClusterParameterGroup instances. Each instance describes one cluster parameter group.