Network.AWS.Redshift.DescribeClusterParameters

Description

Returns a detailed list of parameters contained within the specified Amazon Redshift parameter group. For each parameter the response includes information such as parameter name, description, data type, value, whether the parameter value is modifiable, and so on.

You can specify source filter to retrieve parameters of only specific type. For example, to retrieve parameters that were modified by a user action such as from ModifyClusterParameterGroup , you can specify source equal to user .

For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide .

This operation returns paginated results.

Creating a Request

describeClusterParameters

Arguments

:: Text
dcpsParameterGroupName

-> DescribeClusterParameters

Creates a value of DescribeClusterParameters with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcpsMarker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterParameters request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dcpsMaxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

dcpsSource - The parameter types to return. Specify user to show parameters that are different form the default. Similarly, specify engine-default to show parameters that are the same as the default parameter group. Default: All parameter types returned. Valid Values: user | engine-default

dcpsParameterGroupName - The name of a cluster parameter group for which to return details.

data DescribeClusterParameters

See: describeClusterParameters smart constructor.

Request Lenses

dcpsMarker :: Lens' DescribeClusterParameters (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterParameters request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dcpsMaxRecords :: Lens' DescribeClusterParameters (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

dcpsSource :: Lens' DescribeClusterParameters (Maybe Text)

The parameter types to return. Specify user to show parameters that are different form the default. Similarly, specify engine-default to show parameters that are the same as the default parameter group. Default: All parameter types returned. Valid Values: user | engine-default

dcpsParameterGroupName :: Lens' DescribeClusterParameters Text

The name of a cluster parameter group for which to return details.

Destructuring the Response

describeClusterParametersResponse

Arguments

:: Int
dcprsResponseStatus

-> DescribeClusterParametersResponse

Creates a value of DescribeClusterParametersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcprsMarker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcprsParameters - A list of Parameter instances. Each instance lists the parameters of one cluster parameter group.

dcprsResponseStatus - -- | The response status code.

data DescribeClusterParametersResponse

Contains the output from the DescribeClusterParameters action.

See: describeClusterParametersResponse smart constructor.

Response Lenses

dcprsMarker :: Lens' DescribeClusterParametersResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcprsParameters :: Lens' DescribeClusterParametersResponse [Parameter]

A list of Parameter instances. Each instance lists the parameters of one cluster parameter group.

dcprsResponseStatus :: Lens' DescribeClusterParametersResponse Int

- | The response status code.