Network.AWS.Redshift.DescribeClusterSecurityGroups

Description

Returns information about Amazon Redshift security groups. If the name of a security group is specified, the response will contain only information about only that security group. For information about managing security groups, go to Amazon Redshift Cluster Security Groups in the Amazon Redshift Management Guide. https:/redshift.us-east-1.amazonaws.com ?Action=DescribeClusterSecurityGroups &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20130123us-east-1redshiftaws4_request &x-amz-date=20130123T010237Z &x-amz-signedheaders=content-type;host;x-amz-date 0.0.0.0/0 authorized default default my security group securitygroup1 947a8305-64f8-11e2-bec0-17624ad140dd.

Request

data DescribeClusterSecurityGroups

???.

Request constructor

describeClusterSecurityGroups :: DescribeClusterSecurityGroups

Smart constructor for the minimum required parameters to construct a valid DescribeClusterSecurityGroups request.

The fields accessible through corresponding lenses are:

ClusterSecurityGroupName :: Maybe Text

MaxRecords :: Maybe Integer

Marker :: Maybe Text

Request lenses

dcsg2ClusterSecurityGroupName :: Lens' DescribeClusterSecurityGroups (Maybe Text)

The name of a cluster security group for which you are requesting details. You can specify either the Marker parameter or a ClusterSecurityGroupName parameter, but not both. Example: securitygroup1.

dcsg2MaxRecords :: Lens' DescribeClusterSecurityGroups (Maybe Integer)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

dcsg2Marker :: Lens' DescribeClusterSecurityGroups (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterSecurityGroups request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request. Constraints: You can specify either the ClusterSecurityGroupName parameter or the Marker parameter, but not both.

Response

data DescribeClusterSecurityGroupsResponse

Contains the output from the DescribeClusterSecurityGroups action.

Response constructor

describeClusterSecurityGroupsResponse :: DescribeClusterSecurityGroupsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeClusterSecurityGroupsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

ClusterSecurityGroups :: [ClusterSecurityGroup]

Response lenses

dcsgrMarker :: Lens' DescribeClusterSecurityGroupsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcsgrClusterSecurityGroups :: Lens' DescribeClusterSecurityGroupsResponse [ClusterSecurityGroup]

A list of ClusterSecurityGroup instances.