Network.AWS.Redshift.DescribeClusterSnapshots

Description

Returns one or more snapshot objects, which contain metadata about your cluster snapshots. By default, this operation returns information about all snapshots of all clusters that are owned by you AWS customer account. No information is returned for snapshots owned by inactive AWS customer accounts.

If you specify both tag keys and tag values in the same request, Amazon Redshift returns all snapshots that match any combination of the specified keys and values. For example, if you have owner and environment for tag keys, and admin and test for tag values, all snapshots that have any combination of those values are returned. Only snapshots that you own are returned in the response; shared snapshots are not returned with the tag key and tag value request parameters.

If both tag keys and values are omitted from the request, snapshots are returned regardless of whether they have tag keys or values associated with them.

This operation returns paginated results.

Creating a Request

describeClusterSnapshots :: DescribeClusterSnapshots

Creates a value of DescribeClusterSnapshots with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dSnapshotIdentifier - The snapshot identifier of the snapshot about which to return information.

dTagValues - A tag value or values for which you want to return all matching cluster snapshots that are associated with the specified tag value or values. For example, suppose that you have snapshots that are tagged with values called admin and test . If you specify both of these tag values in the request, Amazon Redshift returns a response with the snapshots that have either or both of these tag values associated with them.

dStartTime - A value that requests only snapshots created at or after the specified time. The time value is specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page. Example: 2012-07-16T18:00:00Z

dTagKeys - A tag key or keys for which you want to return all matching cluster snapshots that are associated with the specified key or keys. For example, suppose that you have snapshots that are tagged with keys called owner and environment . If you specify both of these tag keys in the request, Amazon Redshift returns a response with the snapshots that have either or both of these tag keys associated with them.

dClusterIdentifier - The identifier of the cluster for which information about snapshots is requested.

dSnapshotType - The type of snapshots for which you are requesting information. By default, snapshots of all types are returned. Valid Values: automated | manual

dMarker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterSnapshots request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dMaxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

dEndTime - A time value that requests only snapshots created at or before the specified time. The time value is specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page. Example: 2012-07-16T18:00:00Z

dOwnerAccount - The AWS customer account used to create or copy the snapshot. Use this field to filter the results to snapshots owned by a particular account. To describe snapshots you own, either specify your AWS customer account, or do not specify the parameter.

data DescribeClusterSnapshots

See: describeClusterSnapshots smart constructor.

Request Lenses

dSnapshotIdentifier :: Lens' DescribeClusterSnapshots (Maybe Text)

The snapshot identifier of the snapshot about which to return information.

dTagValues :: Lens' DescribeClusterSnapshots [Text]

A tag value or values for which you want to return all matching cluster snapshots that are associated with the specified tag value or values. For example, suppose that you have snapshots that are tagged with values called admin and test . If you specify both of these tag values in the request, Amazon Redshift returns a response with the snapshots that have either or both of these tag values associated with them.

dStartTime :: Lens' DescribeClusterSnapshots (Maybe UTCTime)

A value that requests only snapshots created at or after the specified time. The time value is specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page. Example: 2012-07-16T18:00:00Z

dTagKeys :: Lens' DescribeClusterSnapshots [Text]

A tag key or keys for which you want to return all matching cluster snapshots that are associated with the specified key or keys. For example, suppose that you have snapshots that are tagged with keys called owner and environment . If you specify both of these tag keys in the request, Amazon Redshift returns a response with the snapshots that have either or both of these tag keys associated with them.

dClusterIdentifier :: Lens' DescribeClusterSnapshots (Maybe Text)

The identifier of the cluster for which information about snapshots is requested.

dSnapshotType :: Lens' DescribeClusterSnapshots (Maybe Text)

The type of snapshots for which you are requesting information. By default, snapshots of all types are returned. Valid Values: automated | manual

dMarker :: Lens' DescribeClusterSnapshots (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterSnapshots request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dMaxRecords :: Lens' DescribeClusterSnapshots (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

dEndTime :: Lens' DescribeClusterSnapshots (Maybe UTCTime)

A time value that requests only snapshots created at or before the specified time. The time value is specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page. Example: 2012-07-16T18:00:00Z

dOwnerAccount :: Lens' DescribeClusterSnapshots (Maybe Text)

The AWS customer account used to create or copy the snapshot. Use this field to filter the results to snapshots owned by a particular account. To describe snapshots you own, either specify your AWS customer account, or do not specify the parameter.

Destructuring the Response

describeClusterSnapshotsResponse

Arguments

:: Int
dcssrsResponseStatus

-> DescribeClusterSnapshotsResponse

Creates a value of DescribeClusterSnapshotsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcssrsSnapshots - A list of Snapshot instances.

dcssrsMarker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcssrsResponseStatus - -- | The response status code.

data DescribeClusterSnapshotsResponse

Contains the output from the DescribeClusterSnapshots action.

See: describeClusterSnapshotsResponse smart constructor.

Response Lenses

dcssrsSnapshots :: Lens' DescribeClusterSnapshotsResponse [Snapshot]

A list of Snapshot instances.

dcssrsMarker :: Lens' DescribeClusterSnapshotsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcssrsResponseStatus :: Lens' DescribeClusterSnapshotsResponse Int

- | The response status code.