Network.AWS.Redshift.DescribeClusterSubnetGroups

Description

Returns one or more cluster subnet group objects, which contain metadata about your cluster subnet groups. By default, this operation returns information about all cluster subnet groups that are defined in you AWS account.

If you specify both tag keys and tag values in the same request, Amazon Redshift returns all subnet groups that match any combination of the specified keys and values. For example, if you have owner and environment for tag keys, and admin and test for tag values, all subnet groups that have any combination of those values are returned.

If both tag keys and values are omitted from the request, subnet groups are returned regardless of whether they have tag keys or values associated with them.

This operation returns paginated results.

Creating a Request

describeClusterSubnetGroups :: DescribeClusterSubnetGroups

Creates a value of DescribeClusterSubnetGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcsgsTagValues - A tag value or values for which you want to return all matching cluster subnet groups that are associated with the specified tag value or values. For example, suppose that you have subnet groups that are tagged with values called admin and test . If you specify both of these tag values in the request, Amazon Redshift returns a response with the subnet groups that have either or both of these tag values associated with them.

dcsgsTagKeys - A tag key or keys for which you want to return all matching cluster subnet groups that are associated with the specified key or keys. For example, suppose that you have subnet groups that are tagged with keys called owner and environment . If you specify both of these tag keys in the request, Amazon Redshift returns a response with the subnet groups that have either or both of these tag keys associated with them.

dcsgsClusterSubnetGroupName - The name of the cluster subnet group for which information is requested.

dcsgsMarker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterSubnetGroups request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dcsgsMaxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

data DescribeClusterSubnetGroups

See: describeClusterSubnetGroups smart constructor.

Request Lenses

dcsgsTagValues :: Lens' DescribeClusterSubnetGroups [Text]

A tag value or values for which you want to return all matching cluster subnet groups that are associated with the specified tag value or values. For example, suppose that you have subnet groups that are tagged with values called admin and test . If you specify both of these tag values in the request, Amazon Redshift returns a response with the subnet groups that have either or both of these tag values associated with them.

dcsgsTagKeys :: Lens' DescribeClusterSubnetGroups [Text]

A tag key or keys for which you want to return all matching cluster subnet groups that are associated with the specified key or keys. For example, suppose that you have subnet groups that are tagged with keys called owner and environment . If you specify both of these tag keys in the request, Amazon Redshift returns a response with the subnet groups that have either or both of these tag keys associated with them.

dcsgsClusterSubnetGroupName :: Lens' DescribeClusterSubnetGroups (Maybe Text)

The name of the cluster subnet group for which information is requested.

dcsgsMarker :: Lens' DescribeClusterSubnetGroups (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterSubnetGroups request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dcsgsMaxRecords :: Lens' DescribeClusterSubnetGroups (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

Destructuring the Response

describeClusterSubnetGroupsResponse

Arguments

:: Int
dcsgrsResponseStatus

-> DescribeClusterSubnetGroupsResponse

Creates a value of DescribeClusterSubnetGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcsgrsClusterSubnetGroups - A list of ClusterSubnetGroup instances.

dcsgrsMarker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcsgrsResponseStatus - -- | The response status code.

data DescribeClusterSubnetGroupsResponse

Contains the output from the DescribeClusterSubnetGroups action.

See: describeClusterSubnetGroupsResponse smart constructor.

Response Lenses

dcsgrsClusterSubnetGroups :: Lens' DescribeClusterSubnetGroupsResponse [ClusterSubnetGroup]

A list of ClusterSubnetGroup instances.

dcsgrsMarker :: Lens' DescribeClusterSubnetGroupsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcsgrsResponseStatus :: Lens' DescribeClusterSubnetGroupsResponse Int

- | The response status code.