Network.AWS.Redshift.DescribeClusterVersions

Description

Returns descriptions of the available Amazon Redshift cluster versions. You can call this operation even before creating any clusters to learn more about the Amazon Redshift versions. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Cluster Management Guide .

This operation returns paginated results.

Creating a Request

describeClusterVersions :: DescribeClusterVersions

Creates a value of DescribeClusterVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcvClusterParameterGroupFamily - The name of a specific cluster parameter group family to return details for. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

dcvMarker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterVersions request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dcvMaxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

dcvClusterVersion - The specific cluster version to return. Example: 1.0

data DescribeClusterVersions

See: describeClusterVersions smart constructor.

Request Lenses

dcvClusterParameterGroupFamily :: Lens' DescribeClusterVersions (Maybe Text)

The name of a specific cluster parameter group family to return details for. Constraints: * Must be 1 to 255 alphanumeric characters * First character must be a letter * Cannot end with a hyphen or contain two consecutive hyphens

dcvMarker :: Lens' DescribeClusterVersions (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterVersions request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dcvMaxRecords :: Lens' DescribeClusterVersions (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

dcvClusterVersion :: Lens' DescribeClusterVersions (Maybe Text)

The specific cluster version to return. Example: 1.0

Destructuring the Response

describeClusterVersionsResponse

Arguments

:: Int
dcvrsResponseStatus

-> DescribeClusterVersionsResponse

Creates a value of DescribeClusterVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcvrsClusterVersions - A list of Version elements.

dcvrsMarker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcvrsResponseStatus - -- | The response status code.

data DescribeClusterVersionsResponse

Contains the output from the DescribeClusterVersions action.

See: describeClusterVersionsResponse smart constructor.

Response Lenses

dcvrsClusterVersions :: Lens' DescribeClusterVersionsResponse [ClusterVersion]

A list of Version elements.

dcvrsMarker :: Lens' DescribeClusterVersionsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcvrsResponseStatus :: Lens' DescribeClusterVersionsResponse Int

- | The response status code.