Network.AWS.Redshift.DescribeDefaultClusterParameters

Description

Returns a list of parameter settings for the specified parameter group family.

For more information about managing parameter groups, go to Amazon RedshiftParameter Groups in the Amazon Redshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeDefaultClusterParameters.html

Request

data DescribeDefaultClusterParameters

Request constructor

describeDefaultClusterParameters

Arguments

:: Text
ddcpParameterGroupFamily

-> DescribeDefaultClusterParameters

DescribeDefaultClusterParameters constructor.

The fields accessible through corresponding lenses are:

ddcpMarker :: Maybe Text

ddcpMaxRecords :: Maybe Int

ddcpParameterGroupFamily :: Text

Request lenses

ddcpMarker :: Lens' DescribeDefaultClusterParameters (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDefaultClusterParameters request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

ddcpMaxRecords :: Lens' DescribeDefaultClusterParameters (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: '100'

Constraints: minimum 20, maximum 100.

ddcpParameterGroupFamily :: Lens' DescribeDefaultClusterParameters Text

The name of the cluster parameter group family.

Response

data DescribeDefaultClusterParametersResponse

Response constructor

describeDefaultClusterParametersResponse

Arguments

:: DefaultClusterParameters
ddcprDefaultClusterParameters

-> DescribeDefaultClusterParametersResponse

DescribeDefaultClusterParametersResponse constructor.

The fields accessible through corresponding lenses are:

ddcprDefaultClusterParameters :: DefaultClusterParameters

Response lenses

ddcprDefaultClusterParameters :: Lens' DescribeDefaultClusterParametersResponse DefaultClusterParameters