Network.AWS.Redshift.DescribeEventSubscriptions

Description

Lists descriptions of all the Amazon Redshift event notifications subscription for a customer account. If you specify a subscription name, lists the description for that subscription.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeEventSubscriptions.html

Request

data DescribeEventSubscriptions

Request constructor

describeEventSubscriptions :: DescribeEventSubscriptions

DescribeEventSubscriptions constructor.

The fields accessible through corresponding lenses are:

des1Marker :: Maybe Text

des1MaxRecords :: Maybe Int

des1SubscriptionName :: Maybe Text

Request lenses

des1Marker :: Lens' DescribeEventSubscriptions (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeEventSubscriptions request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

des1MaxRecords :: Lens' DescribeEventSubscriptions (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: '100'

Constraints: minimum 20, maximum 100.

des1SubscriptionName :: Lens' DescribeEventSubscriptions (Maybe Text)

The name of the Amazon Redshift event notification subscription to be described.

Response

data DescribeEventSubscriptionsResponse

Response constructor

describeEventSubscriptionsResponse :: DescribeEventSubscriptionsResponse

DescribeEventSubscriptionsResponse constructor.

The fields accessible through corresponding lenses are:

desrEventSubscriptionsList :: [EventSubscription]

desrMarker :: Maybe Text

Response lenses

desrEventSubscriptionsList :: Lens' DescribeEventSubscriptionsResponse [EventSubscription]

A list of event subscriptions.

desrMarker :: Lens' DescribeEventSubscriptionsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.