Network.AWS.Redshift.DescribeEvents

Description

Returns events related to clusters, security groups, snapshots, and parameter groups for the past 14 days. Events specific to a particular cluster, security group, snapshot or parameter group can be obtained by providing the name as a parameter. By default, the past hour of events are returned.

This operation returns paginated results.

Creating a Request

describeEvents :: DescribeEvents

Creates a value of DescribeEvents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deStartTime - The beginning of the time interval to retrieve events for, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page. Example: 2009-07-08T18:00Z

deSourceType - The event source to retrieve events for. If no value is specified, all events are returned. Constraints: If SourceType is supplied, SourceIdentifier must also be provided. * Specify cluster when SourceIdentifier is a cluster identifier. * Specify cluster-security-group when SourceIdentifier is a cluster security group name. * Specify cluster-parameter-group when SourceIdentifier is a cluster parameter group name. * Specify cluster-snapshot when SourceIdentifier is a cluster snapshot identifier.

deSourceIdentifier - The identifier of the event source for which events will be returned. If this parameter is not specified, then all sources are included in the response. Constraints: If SourceIdentifier is supplied, SourceType must also be provided. * Specify a cluster identifier when SourceType is cluster . * Specify a cluster security group name when SourceType is cluster-security-group . * Specify a cluster parameter group name when SourceType is cluster-parameter-group . * Specify a cluster snapshot identifier when SourceType is cluster-snapshot .

deMarker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeEvents request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

deMaxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

deEndTime - The end of the time interval for which to retrieve events, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page. Example: 2009-07-08T18:00Z

deDuration - The number of minutes prior to the time of the request for which to retrieve events. For example, if the request is sent at 18:00 and you specify a duration of 60, then only events which have occurred after 17:00 will be returned. Default: 60

data DescribeEvents

See: describeEvents smart constructor.

Request Lenses

deStartTime :: Lens' DescribeEvents (Maybe UTCTime)

The beginning of the time interval to retrieve events for, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page. Example: 2009-07-08T18:00Z

deSourceType :: Lens' DescribeEvents (Maybe SourceType)

The event source to retrieve events for. If no value is specified, all events are returned. Constraints: If SourceType is supplied, SourceIdentifier must also be provided. * Specify cluster when SourceIdentifier is a cluster identifier. * Specify cluster-security-group when SourceIdentifier is a cluster security group name. * Specify cluster-parameter-group when SourceIdentifier is a cluster parameter group name. * Specify cluster-snapshot when SourceIdentifier is a cluster snapshot identifier.

deSourceIdentifier :: Lens' DescribeEvents (Maybe Text)

The identifier of the event source for which events will be returned. If this parameter is not specified, then all sources are included in the response. Constraints: If SourceIdentifier is supplied, SourceType must also be provided. * Specify a cluster identifier when SourceType is cluster . * Specify a cluster security group name when SourceType is cluster-security-group . * Specify a cluster parameter group name when SourceType is cluster-parameter-group . * Specify a cluster snapshot identifier when SourceType is cluster-snapshot .

deMarker :: Lens' DescribeEvents (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeEvents request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

deMaxRecords :: Lens' DescribeEvents (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

deEndTime :: Lens' DescribeEvents (Maybe UTCTime)

The end of the time interval for which to retrieve events, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page. Example: 2009-07-08T18:00Z

deDuration :: Lens' DescribeEvents (Maybe Int)

The number of minutes prior to the time of the request for which to retrieve events. For example, if the request is sent at 18:00 and you specify a duration of 60, then only events which have occurred after 17:00 will be returned. Default: 60

Destructuring the Response

describeEventsResponse

Arguments

:: Int
dersResponseStatus

-> DescribeEventsResponse

Creates a value of DescribeEventsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dersEvents - A list of Event instances.

dersMarker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dersResponseStatus - -- | The response status code.

data DescribeEventsResponse

See: describeEventsResponse smart constructor.

Response Lenses

dersEvents :: Lens' DescribeEventsResponse [Event]

A list of Event instances.

dersMarker :: Lens' DescribeEventsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dersResponseStatus :: Lens' DescribeEventsResponse Int

- | The response status code.