Network.AWS.Redshift.DescribeHsmClientCertificates

Description

Returns information about the specified HSM client certificate. If no certificate ID is specified, returns information about all the HSM certificates owned by your AWS customer account.

If you specify both tag keys and tag values in the same request, Amazon Redshift returns all HSM client certificates that match any combination of the specified keys and values. For example, if you have owner and environment for tag keys, and admin and test for tag values, all HSM client certificates that have any combination of those values are returned.

If both tag keys and values are omitted from the request, HSM client certificates are returned regardless of whether they have tag keys or values associated with them.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeHsmClientCertificates.html

Request

data DescribeHsmClientCertificates

Request constructor

describeHsmClientCertificates :: DescribeHsmClientCertificates

DescribeHsmClientCertificates constructor.

The fields accessible through corresponding lenses are:

dhccHsmClientCertificateIdentifier :: Maybe Text

dhccMarker :: Maybe Text

dhccMaxRecords :: Maybe Int

dhccTagKeys :: [Text]

dhccTagValues :: [Text]

Request lenses

dhccHsmClientCertificateIdentifier :: Lens' DescribeHsmClientCertificates (Maybe Text)

The identifier of a specific HSM client certificate for which you want information. If no identifier is specified, information is returned for all HSM client certificates owned by your AWS customer account.

dhccMarker :: Lens' DescribeHsmClientCertificates (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeHsmClientCertificates request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dhccMaxRecords :: Lens' DescribeHsmClientCertificates (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: '100'

Constraints: minimum 20, maximum 100.

dhccTagKeys :: Lens' DescribeHsmClientCertificates [Text]

A tag key or keys for which you want to return all matching HSM client certificates that are associated with the specified key or keys. For example, suppose that you have HSM client certificates that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the HSM client certificates that have either or both of these tag keys associated with them.

dhccTagValues :: Lens' DescribeHsmClientCertificates [Text]

A tag value or values for which you want to return all matching HSM client certificates that are associated with the specified tag value or values. For example, suppose that you have HSM client certificates that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the HSM client certificates that have either or both of these tag values associated with them.

Response

data DescribeHsmClientCertificatesResponse

Response constructor

describeHsmClientCertificatesResponse :: DescribeHsmClientCertificatesResponse

DescribeHsmClientCertificatesResponse constructor.

The fields accessible through corresponding lenses are:

dhccrHsmClientCertificates :: [HsmClientCertificate]

dhccrMarker :: Maybe Text

Response lenses

dhccrHsmClientCertificates :: Lens' DescribeHsmClientCertificatesResponse [HsmClientCertificate]

A list of the identifiers for one or more HSM client certificates used by Amazon Redshift clusters to store and retrieve database encryption keys in an HSM.

dhccrMarker :: Lens' DescribeHsmClientCertificatesResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.