Network.AWS.Redshift.DescribeOrderableClusterOptions

Description

Returns a list of orderable cluster options. Before you create a new cluster you can use this operation to find what options are available, such as the EC2 Availability Zones (AZ) in the specific AWS region that you can specify, and the node types you can request. The node types differ by available storage, memory, CPU and price. With the cost involved you might want to obtain a list of cluster options in the specific region and specify values when creating a cluster. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Cluster Management Guide .

This operation returns paginated results.

Creating a Request

describeOrderableClusterOptions :: DescribeOrderableClusterOptions

Creates a value of DescribeOrderableClusterOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

docoMarker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeOrderableClusterOptions request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

docoMaxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

docoClusterVersion - The version filter value. Specify this parameter to show only the available offerings matching the specified version. Default: All versions. Constraints: Must be one of the version returned from DescribeClusterVersions .

docoNodeType - The node type filter value. Specify this parameter to show only the available offerings matching the specified node type.

data DescribeOrderableClusterOptions

See: describeOrderableClusterOptions smart constructor.

Request Lenses

docoMarker :: Lens' DescribeOrderableClusterOptions (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeOrderableClusterOptions request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

docoMaxRecords :: Lens' DescribeOrderableClusterOptions (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

docoClusterVersion :: Lens' DescribeOrderableClusterOptions (Maybe Text)

The version filter value. Specify this parameter to show only the available offerings matching the specified version. Default: All versions. Constraints: Must be one of the version returned from DescribeClusterVersions .

docoNodeType :: Lens' DescribeOrderableClusterOptions (Maybe Text)

The node type filter value. Specify this parameter to show only the available offerings matching the specified node type.

Destructuring the Response

describeOrderableClusterOptionsResponse

Arguments

:: Int
docorsResponseStatus

-> DescribeOrderableClusterOptionsResponse

Creates a value of DescribeOrderableClusterOptionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

docorsMarker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

docorsOrderableClusterOptions - An OrderableClusterOption structure containing information about orderable options for the cluster.

docorsResponseStatus - -- | The response status code.

data DescribeOrderableClusterOptionsResponse

Contains the output from the DescribeOrderableClusterOptions action.

See: describeOrderableClusterOptionsResponse smart constructor.

Response Lenses

docorsMarker :: Lens' DescribeOrderableClusterOptionsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

docorsOrderableClusterOptions :: Lens' DescribeOrderableClusterOptionsResponse [OrderableClusterOption]

An OrderableClusterOption structure containing information about orderable options for the cluster.

docorsResponseStatus :: Lens' DescribeOrderableClusterOptionsResponse Int

- | The response status code.