Network.AWS.Redshift.DescribeReservedNodeOfferings

Description

Returns a list of the available reserved node offerings by Amazon Redshift with their descriptions including the node type, the fixed and recurring costs of reserving the node and duration the node will be reserved for you. These descriptions help you determine which reserve node offering you want to purchase. You then use the unique offering ID in you call to PurchaseReservedNodeOffering to reserve one or more nodes for your Amazon Redshift cluster.

For more information about managing parameter groups, go to PurchasingReserved Nodes in the Amazon Redshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_DescribeReservedNodeOfferings.html

Request

data DescribeReservedNodeOfferings

Request constructor

describeReservedNodeOfferings :: DescribeReservedNodeOfferings

DescribeReservedNodeOfferings constructor.

The fields accessible through corresponding lenses are:

drnoMarker :: Maybe Text

drnoMaxRecords :: Maybe Int

drnoReservedNodeOfferingId :: Maybe Text

Request lenses

drnoMarker :: Lens' DescribeReservedNodeOfferings (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeReservedNodeOfferings request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

drnoMaxRecords :: Lens' DescribeReservedNodeOfferings (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: '100'

Constraints: minimum 20, maximum 100.

drnoReservedNodeOfferingId :: Lens' DescribeReservedNodeOfferings (Maybe Text)

The unique identifier for the offering.

Response

data DescribeReservedNodeOfferingsResponse

Response constructor

describeReservedNodeOfferingsResponse :: DescribeReservedNodeOfferingsResponse

DescribeReservedNodeOfferingsResponse constructor.

The fields accessible through corresponding lenses are:

drnorMarker :: Maybe Text

drnorReservedNodeOfferings :: [ReservedNodeOffering]

Response lenses

drnorMarker :: Lens' DescribeReservedNodeOfferingsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

drnorReservedNodeOfferings :: Lens' DescribeReservedNodeOfferingsResponse [ReservedNodeOffering]

A list of reserved node offerings.