Network.AWS.Redshift.DescribeReservedNodes

Description

Returns the descriptions of the reserved nodes.

This operation returns paginated results.

Creating a Request

describeReservedNodes :: DescribeReservedNodes

Creates a value of DescribeReservedNodes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drnReservedNodeId - Identifier for the node reservation.

drnMarker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeReservedNodes request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

drnMaxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

data DescribeReservedNodes

See: describeReservedNodes smart constructor.

Request Lenses

drnReservedNodeId :: Lens' DescribeReservedNodes (Maybe Text)

Identifier for the node reservation.

drnMarker :: Lens' DescribeReservedNodes (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeReservedNodes request exceed the value specified in MaxRecords , AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

drnMaxRecords :: Lens' DescribeReservedNodes (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value. Default: 100 Constraints: minimum 20, maximum 100.

Destructuring the Response

describeReservedNodesResponse

Arguments

:: Int
drnrsResponseStatus

-> DescribeReservedNodesResponse

Creates a value of DescribeReservedNodesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drnrsReservedNodes - The list of ReservedNode objects.

drnrsMarker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

drnrsResponseStatus - -- | The response status code.

data DescribeReservedNodesResponse

See: describeReservedNodesResponse smart constructor.

Response Lenses

drnrsReservedNodes :: Lens' DescribeReservedNodesResponse [ReservedNode]

The list of ReservedNode objects.

drnrsMarker :: Lens' DescribeReservedNodesResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

drnrsResponseStatus :: Lens' DescribeReservedNodesResponse Int

- | The response status code.