Network.AWS.Redshift.DescribeResize

Description

Returns information about the last resize operation for the specified cluster. If no resize operation has ever been initiated for the specified cluster, a HTTP 404 error is returned. If a resize operation was initiated and completed, the status of the resize remains as SUCCEEDED until the next resize.

A resize operation can be requested using ModifyCluster and specifying a different number or type of nodes for the cluster.

Creating a Request

describeResize

Arguments

:: Text
drClusterIdentifier

-> DescribeResize

Creates a value of DescribeResize with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drClusterIdentifier - The unique identifier of a cluster whose resize progress you are requesting. This parameter is case-sensitive. By default, resize operations for all clusters defined for an AWS account are returned.

data DescribeResize

See: describeResize smart constructor.

Request Lenses

drClusterIdentifier :: Lens' DescribeResize Text

The unique identifier of a cluster whose resize progress you are requesting. This parameter is case-sensitive. By default, resize operations for all clusters defined for an AWS account are returned.

Destructuring the Response

describeResizeResponse

Arguments

:: Int
drrsResponseStatus

-> DescribeResizeResponse

Creates a value of DescribeResizeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drrsImportTablesNotStarted - The names of tables that have not been yet imported. Valid Values: List of table names

drrsStatus - The status of the resize operation. Valid Values: NONE | IN_PROGRESS | FAILED | SUCCEEDED

drrsEstimatedTimeToCompletionInSeconds - The estimated time remaining, in seconds, until the resize operation is complete. This value is calculated based on the average resize rate and the estimated amount of data remaining to be processed. Once the resize operation is complete, this value will be 0.

drrsAvgResizeRateInMegaBytesPerSecond - The average rate of the resize operation over the last few minutes, measured in megabytes per second. After the resize operation completes, this value shows the average rate of the entire resize operation.

drrsTargetNumberOfNodes - The number of nodes that the cluster will have after the resize operation is complete.

drrsTargetNodeType - The node type that the cluster will have after the resize operation is complete.

drrsImportTablesInProgress - The names of tables that are being currently imported. Valid Values: List of table names.

drrsImportTablesCompleted - The names of tables that have been completely imported . Valid Values: List of table names.

drrsProgressInMegaBytes - While the resize operation is in progress, this value shows the current amount of data, in megabytes, that has been processed so far. When the resize operation is complete, this value shows the total amount of data, in megabytes, on the cluster, which may be more or less than TotalResizeDataInMegaBytes (the estimated total amount of data before resize).

drrsTotalResizeDataInMegaBytes - The estimated total amount of data, in megabytes, on the cluster before the resize operation began.

drrsTargetClusterType - The cluster type after the resize operation is complete. Valid Values: multi-node | single-node

drrsElapsedTimeInSeconds - The amount of seconds that have elapsed since the resize operation began. After the resize operation completes, this value shows the total actual time, in seconds, for the resize operation.

drrsResponseStatus - -- | The response status code.

data DescribeResizeResponse

Describes the result of a cluster resize operation.

See: describeResizeResponse smart constructor.

Response Lenses

drrsImportTablesNotStarted :: Lens' DescribeResizeResponse [Text]

The names of tables that have not been yet imported. Valid Values: List of table names

drrsStatus :: Lens' DescribeResizeResponse (Maybe Text)

The status of the resize operation. Valid Values: NONE | IN_PROGRESS | FAILED | SUCCEEDED

drrsEstimatedTimeToCompletionInSeconds :: Lens' DescribeResizeResponse (Maybe Integer)

The estimated time remaining, in seconds, until the resize operation is complete. This value is calculated based on the average resize rate and the estimated amount of data remaining to be processed. Once the resize operation is complete, this value will be 0.

drrsAvgResizeRateInMegaBytesPerSecond :: Lens' DescribeResizeResponse (Maybe Double)

The average rate of the resize operation over the last few minutes, measured in megabytes per second. After the resize operation completes, this value shows the average rate of the entire resize operation.

drrsTargetNumberOfNodes :: Lens' DescribeResizeResponse (Maybe Int)

The number of nodes that the cluster will have after the resize operation is complete.

drrsTargetNodeType :: Lens' DescribeResizeResponse (Maybe Text)

The node type that the cluster will have after the resize operation is complete.

drrsImportTablesInProgress :: Lens' DescribeResizeResponse [Text]

The names of tables that are being currently imported. Valid Values: List of table names.

drrsImportTablesCompleted :: Lens' DescribeResizeResponse [Text]

The names of tables that have been completely imported . Valid Values: List of table names.

drrsProgressInMegaBytes :: Lens' DescribeResizeResponse (Maybe Integer)

While the resize operation is in progress, this value shows the current amount of data, in megabytes, that has been processed so far. When the resize operation is complete, this value shows the total amount of data, in megabytes, on the cluster, which may be more or less than TotalResizeDataInMegaBytes (the estimated total amount of data before resize).

drrsTotalResizeDataInMegaBytes :: Lens' DescribeResizeResponse (Maybe Integer)

The estimated total amount of data, in megabytes, on the cluster before the resize operation began.

drrsTargetClusterType :: Lens' DescribeResizeResponse (Maybe Text)

The cluster type after the resize operation is complete. Valid Values: multi-node | single-node

drrsElapsedTimeInSeconds :: Lens' DescribeResizeResponse (Maybe Integer)

The amount of seconds that have elapsed since the resize operation began. After the resize operation completes, this value shows the total actual time, in seconds, for the resize operation.

drrsResponseStatus :: Lens' DescribeResizeResponse Int

- | The response status code.