Network.AWS.Redshift.EnableLogging

Description

Starts logging information, such as queries and connection attempts, for the specified Amazon Redshift cluster.

Creating a Request

enableLogging

Arguments

:: Text
elClusterIdentifier

-> Text
elBucketName

-> EnableLogging

Creates a value of EnableLogging with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

elS3KeyPrefix - The prefix applied to the log file names. Constraints: * Cannot exceed 512 characters * Cannot contain spaces( ), double quotes ("), single quotes ('), a backslash (), or control characters. The hexadecimal codes for invalid characters are: * x00 to x20 * x22 * x27 * x5c * x7f or larger

elClusterIdentifier - The identifier of the cluster on which logging is to be started. Example: examplecluster

elBucketName - The name of an existing S3 bucket where the log files are to be stored. Constraints: * Must be in the same region as the cluster * The cluster must have read bucket and put object permissions

data EnableLogging

See: enableLogging smart constructor.

Request Lenses

elS3KeyPrefix :: Lens' EnableLogging (Maybe Text)

The prefix applied to the log file names. Constraints: * Cannot exceed 512 characters * Cannot contain spaces( ), double quotes ("), single quotes ('), a backslash (), or control characters. The hexadecimal codes for invalid characters are: * x00 to x20 * x22 * x27 * x5c * x7f or larger

elClusterIdentifier :: Lens' EnableLogging Text

The identifier of the cluster on which logging is to be started. Example: examplecluster

elBucketName :: Lens' EnableLogging Text

The name of an existing S3 bucket where the log files are to be stored. Constraints: * Must be in the same region as the cluster * The cluster must have read bucket and put object permissions

Destructuring the Response

loggingStatus :: LoggingStatus

Creates a value of LoggingStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsLastFailureTime - The last time when logs failed to be delivered.

lsLastSuccessfulDeliveryTime - The last time that logs were delivered.

lsS3KeyPrefix - The prefix applied to the log file names.

lsBucketName - The name of the S3 bucket where the log files are stored.

lsLoggingEnabled - true if logging is on, false if logging is off.

lsLastFailureMessage - The message indicating that logs failed to be delivered.

data LoggingStatus

Describes the status of logging for a cluster.

See: loggingStatus smart constructor.

Response Lenses

lsLastFailureTime :: Lens' LoggingStatus (Maybe UTCTime)

The last time when logs failed to be delivered.

lsLastSuccessfulDeliveryTime :: Lens' LoggingStatus (Maybe UTCTime)

The last time that logs were delivered.

lsS3KeyPrefix :: Lens' LoggingStatus (Maybe Text)

The prefix applied to the log file names.

lsBucketName :: Lens' LoggingStatus (Maybe Text)

The name of the S3 bucket where the log files are stored.

lsLoggingEnabled :: Lens' LoggingStatus (Maybe Bool)

true if logging is on, false if logging is off.

lsLastFailureMessage :: Lens' LoggingStatus (Maybe Text)

The message indicating that logs failed to be delivered.