Network.AWS.Redshift.EnableSnapshotCopy

Description

Enables the automatic copy of snapshots from one region to another region for a specified cluster.

Creating a Request

enableSnapshotCopy

Arguments

:: Text
escClusterIdentifier

-> Text
escDestinationRegion

-> EnableSnapshotCopy

Creates a value of EnableSnapshotCopy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

escRetentionPeriod - The number of days to retain automated snapshots in the destination region after they are copied from the source region. Default: 7. Constraints: Must be at least 1 and no more than 35.

escSnapshotCopyGrantName - The name of the snapshot copy grant to use when snapshots of an AWS KMS-encrypted cluster are copied to the destination region.

escClusterIdentifier - The unique identifier of the source cluster to copy snapshots from. Constraints: Must be the valid name of an existing cluster that does not already have cross-region snapshot copy enabled.

escDestinationRegion - The destination region that you want to copy snapshots to. Constraints: Must be the name of a valid region. For more information, see Regions and Endpoints in the Amazon Web Services General Reference.

data EnableSnapshotCopy

See: enableSnapshotCopy smart constructor.

Request Lenses

escRetentionPeriod :: Lens' EnableSnapshotCopy (Maybe Int)

The number of days to retain automated snapshots in the destination region after they are copied from the source region. Default: 7. Constraints: Must be at least 1 and no more than 35.

escSnapshotCopyGrantName :: Lens' EnableSnapshotCopy (Maybe Text)

The name of the snapshot copy grant to use when snapshots of an AWS KMS-encrypted cluster are copied to the destination region.

escClusterIdentifier :: Lens' EnableSnapshotCopy Text

The unique identifier of the source cluster to copy snapshots from. Constraints: Must be the valid name of an existing cluster that does not already have cross-region snapshot copy enabled.

escDestinationRegion :: Lens' EnableSnapshotCopy Text

The destination region that you want to copy snapshots to. Constraints: Must be the name of a valid region. For more information, see Regions and Endpoints in the Amazon Web Services General Reference.

Destructuring the Response

enableSnapshotCopyResponse

Arguments

:: Int
escrsResponseStatus

-> EnableSnapshotCopyResponse

Creates a value of EnableSnapshotCopyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

escrsCluster - Undocumented member.

escrsResponseStatus - -- | The response status code.

data EnableSnapshotCopyResponse

See: enableSnapshotCopyResponse smart constructor.

Response Lenses

escrsCluster :: Lens' EnableSnapshotCopyResponse (Maybe Cluster)

Undocumented member.

escrsResponseStatus :: Lens' EnableSnapshotCopyResponse Int

- | The response status code.