Network.AWS.Redshift.ModifyClusterParameterGroup

Description

Modifies the parameters of a parameter group.

For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide .

Creating a Request

modifyClusterParameterGroup

Arguments

:: Text
mcpgParameterGroupName

-> ModifyClusterParameterGroup

Creates a value of ModifyClusterParameterGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mcpgParameterGroupName - The name of the parameter group to be modified.

mcpgParameters - An array of parameters to be modified. A maximum of 20 parameters can be modified in a single request. For each parameter to be modified, you must supply at least the parameter name and parameter value; other name-value pairs of the parameter are optional. For the workload management (WLM) configuration, you must supply all the name-value pairs in the wlm_json_configuration parameter.

data ModifyClusterParameterGroup

See: modifyClusterParameterGroup smart constructor.

Request Lenses

mcpgParameterGroupName :: Lens' ModifyClusterParameterGroup Text

The name of the parameter group to be modified.

mcpgParameters :: Lens' ModifyClusterParameterGroup [Parameter]

An array of parameters to be modified. A maximum of 20 parameters can be modified in a single request. For each parameter to be modified, you must supply at least the parameter name and parameter value; other name-value pairs of the parameter are optional. For the workload management (WLM) configuration, you must supply all the name-value pairs in the wlm_json_configuration parameter.

Destructuring the Response

clusterParameterGroupNameMessage :: ClusterParameterGroupNameMessage

Creates a value of ClusterParameterGroupNameMessage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpgnmParameterGroupStatus - The status of the parameter group. For example, if you made a change to a parameter group name-value pair, then the change could be pending a reboot of an associated cluster.

cpgnmParameterGroupName - The name of the cluster parameter group.

data ClusterParameterGroupNameMessage

See: clusterParameterGroupNameMessage smart constructor.

Response Lenses

cpgnmParameterGroupStatus :: Lens' ClusterParameterGroupNameMessage (Maybe Text)

The status of the parameter group. For example, if you made a change to a parameter group name-value pair, then the change could be pending a reboot of an associated cluster.

cpgnmParameterGroupName :: Lens' ClusterParameterGroupNameMessage (Maybe Text)

The name of the cluster parameter group.