Network.AWS.Redshift.ModifyClusterParameterGroup

Description

Modifies the parameters of a parameter group.

For more information about managing parameter groups, go to Amazon RedshiftParameter Groups in the Amazon Redshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_ModifyClusterParameterGroup.html

Request

data ModifyClusterParameterGroup

Request constructor

modifyClusterParameterGroup

Arguments

:: Text
mcpgParameterGroupName

-> ModifyClusterParameterGroup

ModifyClusterParameterGroup constructor.

The fields accessible through corresponding lenses are:

mcpgParameterGroupName :: Text

mcpgParameters :: [Parameter]

Request lenses

mcpgParameterGroupName :: Lens' ModifyClusterParameterGroup Text

The name of the parameter group to be modified.

mcpgParameters :: Lens' ModifyClusterParameterGroup [Parameter]

An array of parameters to be modified. A maximum of 20 parameters can be modified in a single request.

For each parameter to be modified, you must supply at least the parameter name and parameter value; other name-value pairs of the parameter are optional.

For the workload management (WLM) configuration, you must supply all the name-value pairs in the wlm_json_configuration parameter.

Response

data ModifyClusterParameterGroupResponse

Response constructor

modifyClusterParameterGroupResponse :: ModifyClusterParameterGroupResponse

ModifyClusterParameterGroupResponse constructor.

The fields accessible through corresponding lenses are:

mcpgrParameterGroupName :: Maybe Text

mcpgrParameterGroupStatus :: Maybe Text

Response lenses

mcpgrParameterGroupName :: Lens' ModifyClusterParameterGroupResponse (Maybe Text)

The name of the cluster parameter group.

mcpgrParameterGroupStatus :: Lens' ModifyClusterParameterGroupResponse (Maybe Text)

The status of the parameter group. For example, if you made a change to a parameter group name-value pair, then the change could be pending a reboot of an associated cluster.