Network.AWS.Redshift.ModifyClusterSubnetGroup

Description

Modifies a cluster subnet group to include the specified list of VPC subnets. The operation replaces the existing list of subnets with the new list of subnets.

Creating a Request

modifyClusterSubnetGroup

Arguments

:: Text
mcsgClusterSubnetGroupName

-> ModifyClusterSubnetGroup

Creates a value of ModifyClusterSubnetGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mcsgDescription - A text description of the subnet group to be modified.

mcsgClusterSubnetGroupName - The name of the subnet group to be modified.

mcsgSubnetIds - An array of VPC subnet IDs. A maximum of 20 subnets can be modified in a single request.

data ModifyClusterSubnetGroup

See: modifyClusterSubnetGroup smart constructor.

Request Lenses

mcsgDescription :: Lens' ModifyClusterSubnetGroup (Maybe Text)

A text description of the subnet group to be modified.

mcsgClusterSubnetGroupName :: Lens' ModifyClusterSubnetGroup Text

The name of the subnet group to be modified.

mcsgSubnetIds :: Lens' ModifyClusterSubnetGroup [Text]

An array of VPC subnet IDs. A maximum of 20 subnets can be modified in a single request.

Destructuring the Response

modifyClusterSubnetGroupResponse

Arguments

:: Int
mcsgrsResponseStatus

-> ModifyClusterSubnetGroupResponse

Creates a value of ModifyClusterSubnetGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mcsgrsClusterSubnetGroup - Undocumented member.

mcsgrsResponseStatus - -- | The response status code.

data ModifyClusterSubnetGroupResponse

See: modifyClusterSubnetGroupResponse smart constructor.

Response Lenses

mcsgrsClusterSubnetGroup :: Lens' ModifyClusterSubnetGroupResponse (Maybe ClusterSubnetGroup)

Undocumented member.

mcsgrsResponseStatus :: Lens' ModifyClusterSubnetGroupResponse Int

- | The response status code.