Network.AWS.Redshift.ModifySnapshotCopyRetentionPeriod

Description

Modifies the number of days to retain automated snapshots in the destination region after they are copied from the source region.

Creating a Request

modifySnapshotCopyRetentionPeriod

Arguments

:: Text
mscrpClusterIdentifier

-> Int
mscrpRetentionPeriod

-> ModifySnapshotCopyRetentionPeriod

Creates a value of ModifySnapshotCopyRetentionPeriod with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mscrpClusterIdentifier - The unique identifier of the cluster for which you want to change the retention period for automated snapshots that are copied to a destination region. Constraints: Must be the valid name of an existing cluster that has cross-region snapshot copy enabled.

mscrpRetentionPeriod - The number of days to retain automated snapshots in the destination region after they are copied from the source region. If you decrease the retention period for automated snapshots that are copied to a destination region, Amazon Redshift will delete any existing automated snapshots that were copied to the destination region and that fall outside of the new retention period. Constraints: Must be at least 1 and no more than 35.

data ModifySnapshotCopyRetentionPeriod

See: modifySnapshotCopyRetentionPeriod smart constructor.

Request Lenses

mscrpClusterIdentifier :: Lens' ModifySnapshotCopyRetentionPeriod Text

The unique identifier of the cluster for which you want to change the retention period for automated snapshots that are copied to a destination region. Constraints: Must be the valid name of an existing cluster that has cross-region snapshot copy enabled.

mscrpRetentionPeriod :: Lens' ModifySnapshotCopyRetentionPeriod Int

The number of days to retain automated snapshots in the destination region after they are copied from the source region. If you decrease the retention period for automated snapshots that are copied to a destination region, Amazon Redshift will delete any existing automated snapshots that were copied to the destination region and that fall outside of the new retention period. Constraints: Must be at least 1 and no more than 35.

Destructuring the Response

modifySnapshotCopyRetentionPeriodResponse

Arguments

:: Int
mscrprsResponseStatus

-> ModifySnapshotCopyRetentionPeriodResponse

Creates a value of ModifySnapshotCopyRetentionPeriodResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mscrprsCluster - Undocumented member.

mscrprsResponseStatus - -- | The response status code.

data ModifySnapshotCopyRetentionPeriodResponse

See: modifySnapshotCopyRetentionPeriodResponse smart constructor.

Response Lenses

mscrprsCluster :: Lens' ModifySnapshotCopyRetentionPeriodResponse (Maybe Cluster)

Undocumented member.

mscrprsResponseStatus :: Lens' ModifySnapshotCopyRetentionPeriodResponse Int

- | The response status code.