Network.AWS.Redshift.PurchaseReservedNodeOffering

Description

Allows you to purchase reserved nodes. Amazon Redshift offers a predefined set of reserved node offerings. You can purchase one or more of the offerings. You can call the DescribeReservedNodeOfferings API to obtain the available reserved node offerings. You can call this API by providing a specific reserved node offering and the number of nodes you want to reserve.

For more information about reserved node offerings, go to Purchasing Reserved Nodes in the Amazon Redshift Cluster Management Guide .

Creating a Request

purchaseReservedNodeOffering

Arguments

:: Text
prnoReservedNodeOfferingId

-> PurchaseReservedNodeOffering

Creates a value of PurchaseReservedNodeOffering with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prnoNodeCount - The number of reserved nodes that you want to purchase. Default: 1

prnoReservedNodeOfferingId - The unique identifier of the reserved node offering you want to purchase.

data PurchaseReservedNodeOffering

See: purchaseReservedNodeOffering smart constructor.

Request Lenses

prnoNodeCount :: Lens' PurchaseReservedNodeOffering (Maybe Int)

The number of reserved nodes that you want to purchase. Default: 1

prnoReservedNodeOfferingId :: Lens' PurchaseReservedNodeOffering Text

The unique identifier of the reserved node offering you want to purchase.

Destructuring the Response

purchaseReservedNodeOfferingResponse

Arguments

:: Int
prnorsResponseStatus

-> PurchaseReservedNodeOfferingResponse

Creates a value of PurchaseReservedNodeOfferingResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prnorsReservedNode - Undocumented member.

prnorsResponseStatus - -- | The response status code.

data PurchaseReservedNodeOfferingResponse

See: purchaseReservedNodeOfferingResponse smart constructor.

Response Lenses

prnorsReservedNode :: Lens' PurchaseReservedNodeOfferingResponse (Maybe ReservedNode)

Undocumented member.

prnorsResponseStatus :: Lens' PurchaseReservedNodeOfferingResponse Int

- | The response status code.