Network.AWS.Redshift.RebootCluster

Description

Reboots a cluster. This action is taken as soon as possible. It results in a momentary outage to the cluster, during which the cluster status is set to rebooting . A cluster event is created when the reboot is completed. Any pending cluster modifications (see ModifyCluster ) are applied at this reboot. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Cluster Management Guide .

Creating a Request

rebootCluster

Arguments

:: Text
rcClusterIdentifier

-> RebootCluster

Creates a value of RebootCluster with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcClusterIdentifier - The cluster identifier.

data RebootCluster

See: rebootCluster smart constructor.

Request Lenses

rcClusterIdentifier :: Lens' RebootCluster Text

The cluster identifier.

Destructuring the Response

rebootClusterResponse

Arguments

:: Int
rcrsResponseStatus

-> RebootClusterResponse

Creates a value of RebootClusterResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcrsCluster - Undocumented member.

rcrsResponseStatus - -- | The response status code.

data RebootClusterResponse

See: rebootClusterResponse smart constructor.

Response Lenses

rcrsCluster :: Lens' RebootClusterResponse (Maybe Cluster)

Undocumented member.

rcrsResponseStatus :: Lens' RebootClusterResponse Int

- | The response status code.