Network.AWS.Redshift.RestoreFromClusterSnapshot

Description

Creates a new cluster from a snapshot. By default, Amazon Redshift creates the resulting cluster with the same configuration as the original cluster from which the snapshot was created, except that the new cluster is created with the default cluster security and parameter groups. After Amazon Redshift creates the cluster, you can use the ModifyCluster API to associate a different security group and different parameter group with the restored cluster. If you are using a DS node type, you can also choose to change to another DS node type of the same size during restore.

If you restore a cluster into a VPC, you must provide a cluster subnet group where you want the cluster restored.

For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Cluster Management Guide .

Creating a Request

restoreFromClusterSnapshot

Arguments

:: Text
rfcsClusterIdentifier

-> Text
rfcsSnapshotIdentifier

-> RestoreFromClusterSnapshot

Creates a value of RestoreFromClusterSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rfcsEnhancedVPCRouting - An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide. If this option is true , enhanced VPC routing is enabled. Default: false

rfcsAdditionalInfo - Reserved.

rfcsPubliclyAccessible - If true , the cluster can be accessed from a public network.

rfcsSnapshotClusterIdentifier - The name of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

rfcsHSMConfigurationIdentifier - Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

rfcsClusterSecurityGroups - A list of security groups to be associated with this cluster. Default: The default cluster security group for Amazon Redshift. Cluster security groups only apply to clusters outside of VPCs.

rfcsAutomatedSnapshotRetentionPeriod - The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot . Default: The value selected for the cluster from which the snapshot was taken. Constraints: Must be a value from 0 to 35.

rfcsClusterSubnetGroupName - The name of the subnet group where you want to cluster restored. A snapshot of cluster in VPC can be restored only in VPC. Therefore, you must provide subnet group name where you want the cluster restored.

rfcsHSMClientCertificateIdentifier - Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

rfcsElasticIP - The elastic IP (EIP) address for the cluster.

rfcsPreferredMaintenanceWindow - The weekly time range (in UTC) during which automated cluster maintenance can occur. Format: ddd:hh24:mi-ddd:hh24:mi Default: The value selected for the cluster from which the snapshot was taken. For more information about the time blocks for each region, see Maintenance Windows in Amazon Redshift Cluster Management Guide. Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun Constraints: Minimum 30-minute window.

rfcsKMSKeyId - The AWS Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster that you restore from a shared snapshot.

rfcsAvailabilityZone - The Amazon EC2 Availability Zone in which to restore the cluster. Default: A random, system-chosen Availability Zone. Example: us-east-1a

rfcsVPCSecurityGroupIds - A list of Virtual Private Cloud (VPC) security groups to be associated with the cluster. Default: The default VPC security group is associated with the cluster. VPC security groups only apply to clusters in VPCs.

rfcsIAMRoles - A list of AWS Identity and Access Management (IAM) roles that can be used by the cluster to access other AWS services. You must supply the IAM roles in their Amazon Resource Name (ARN) format. You can supply up to 10 IAM roles in a single request. A cluster can have up to 10 IAM roles associated at any time.

rfcsOwnerAccount - The AWS customer account used to create or copy the snapshot. Required if you are restoring a snapshot you do not own, optional if you own the snapshot.

rfcsNodeType - The node type that the restored cluster will be provisioned with. Default: The node type of the cluster from which the snapshot was taken. You can modify this if you are using any DS node type. In that case, you can choose to restore into another DS node type of the same size. For example, you can restore ds1.8xlarge into ds2.8xlarge, or ds2.xlarge into ds1.xlarge. If you have a DC instance type, you must restore into that same instance type and size. In other words, you can only restore a dc1.large instance type into another dc1.large instance type. For more information about node types, see About Clusters and Nodes in the Amazon Redshift Cluster Management Guide

rfcsAllowVersionUpgrade - If true , major version upgrades can be applied during the maintenance window to the Amazon Redshift engine that is running on the cluster. Default: true

rfcsClusterParameterGroupName - The name of the parameter group to be associated with this cluster. Default: The default Amazon Redshift cluster parameter group. For information about the default parameter group, go to Working with Amazon Redshift Parameter Groups . Constraints: * Must be 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

rfcsPort - The port number on which the cluster accepts connections. Default: The same port as the original cluster. Constraints: Must be between 1115 and 65535 .

rfcsClusterIdentifier - The identifier of the cluster that will be created from restoring the snapshot. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * Alphabetic characters must be lowercase. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. * Must be unique for all clusters within an AWS account.

rfcsSnapshotIdentifier - The name of the snapshot from which to create the new cluster. This parameter isn't case sensitive. Example: my-snapshot-id

data RestoreFromClusterSnapshot

See: restoreFromClusterSnapshot smart constructor.

Request Lenses

rfcsEnhancedVPCRouting :: Lens' RestoreFromClusterSnapshot (Maybe Bool)

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide. If this option is true , enhanced VPC routing is enabled. Default: false

rfcsAdditionalInfo :: Lens' RestoreFromClusterSnapshot (Maybe Text)

Reserved.

rfcsPubliclyAccessible :: Lens' RestoreFromClusterSnapshot (Maybe Bool)

If true , the cluster can be accessed from a public network.

rfcsSnapshotClusterIdentifier :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The name of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

rfcsHSMConfigurationIdentifier :: Lens' RestoreFromClusterSnapshot (Maybe Text)

Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

rfcsClusterSecurityGroups :: Lens' RestoreFromClusterSnapshot [Text]

A list of security groups to be associated with this cluster. Default: The default cluster security group for Amazon Redshift. Cluster security groups only apply to clusters outside of VPCs.

rfcsAutomatedSnapshotRetentionPeriod :: Lens' RestoreFromClusterSnapshot (Maybe Int)

The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot . Default: The value selected for the cluster from which the snapshot was taken. Constraints: Must be a value from 0 to 35.

rfcsClusterSubnetGroupName :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The name of the subnet group where you want to cluster restored. A snapshot of cluster in VPC can be restored only in VPC. Therefore, you must provide subnet group name where you want the cluster restored.

rfcsHSMClientCertificateIdentifier :: Lens' RestoreFromClusterSnapshot (Maybe Text)

Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

rfcsElasticIP :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The elastic IP (EIP) address for the cluster.

rfcsPreferredMaintenanceWindow :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The weekly time range (in UTC) during which automated cluster maintenance can occur. Format: ddd:hh24:mi-ddd:hh24:mi Default: The value selected for the cluster from which the snapshot was taken. For more information about the time blocks for each region, see Maintenance Windows in Amazon Redshift Cluster Management Guide. Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun Constraints: Minimum 30-minute window.

rfcsKMSKeyId :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The AWS Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster that you restore from a shared snapshot.

rfcsAvailabilityZone :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The Amazon EC2 Availability Zone in which to restore the cluster. Default: A random, system-chosen Availability Zone. Example: us-east-1a

rfcsVPCSecurityGroupIds :: Lens' RestoreFromClusterSnapshot [Text]

A list of Virtual Private Cloud (VPC) security groups to be associated with the cluster. Default: The default VPC security group is associated with the cluster. VPC security groups only apply to clusters in VPCs.

rfcsIAMRoles :: Lens' RestoreFromClusterSnapshot [Text]

A list of AWS Identity and Access Management (IAM) roles that can be used by the cluster to access other AWS services. You must supply the IAM roles in their Amazon Resource Name (ARN) format. You can supply up to 10 IAM roles in a single request. A cluster can have up to 10 IAM roles associated at any time.

rfcsOwnerAccount :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The AWS customer account used to create or copy the snapshot. Required if you are restoring a snapshot you do not own, optional if you own the snapshot.

rfcsNodeType :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The node type that the restored cluster will be provisioned with. Default: The node type of the cluster from which the snapshot was taken. You can modify this if you are using any DS node type. In that case, you can choose to restore into another DS node type of the same size. For example, you can restore ds1.8xlarge into ds2.8xlarge, or ds2.xlarge into ds1.xlarge. If you have a DC instance type, you must restore into that same instance type and size. In other words, you can only restore a dc1.large instance type into another dc1.large instance type. For more information about node types, see About Clusters and Nodes in the Amazon Redshift Cluster Management Guide

rfcsAllowVersionUpgrade :: Lens' RestoreFromClusterSnapshot (Maybe Bool)

If true , major version upgrades can be applied during the maintenance window to the Amazon Redshift engine that is running on the cluster. Default: true

rfcsClusterParameterGroupName :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The name of the parameter group to be associated with this cluster. Default: The default Amazon Redshift cluster parameter group. For information about the default parameter group, go to Working with Amazon Redshift Parameter Groups . Constraints: * Must be 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

rfcsPort :: Lens' RestoreFromClusterSnapshot (Maybe Int)

The port number on which the cluster accepts connections. Default: The same port as the original cluster. Constraints: Must be between 1115 and 65535 .

rfcsClusterIdentifier :: Lens' RestoreFromClusterSnapshot Text

The identifier of the cluster that will be created from restoring the snapshot. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * Alphabetic characters must be lowercase. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. * Must be unique for all clusters within an AWS account.

rfcsSnapshotIdentifier :: Lens' RestoreFromClusterSnapshot Text

The name of the snapshot from which to create the new cluster. This parameter isn't case sensitive. Example: my-snapshot-id

Destructuring the Response

restoreFromClusterSnapshotResponse

Arguments

:: Int
rfcsrsResponseStatus

-> RestoreFromClusterSnapshotResponse

Creates a value of RestoreFromClusterSnapshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rfcsrsCluster - Undocumented member.

rfcsrsResponseStatus - -- | The response status code.

data RestoreFromClusterSnapshotResponse

See: restoreFromClusterSnapshotResponse smart constructor.

Response Lenses

rfcsrsCluster :: Lens' RestoreFromClusterSnapshotResponse (Maybe Cluster)

Undocumented member.

rfcsrsResponseStatus :: Lens' RestoreFromClusterSnapshotResponse Int

- | The response status code.