Network.AWS.Redshift.RestoreFromClusterSnapshot

Description

Creates a new cluster from a snapshot. Amazon Redshift creates the resulting cluster with the same configuration as the original cluster from which the snapshot was created, except that the new cluster is created with the default cluster security and parameter group. After Amazon Redshift creates the cluster you can use the ModifyCluster API to associate a different security group and different parameter group with the restored cluster.

If you restore a cluster into a VPC, you must provide a cluster subnet group where you want the cluster restored.

For more information about working with snapshots, go to Amazon RedshiftSnapshots in the Amazon Redshift Cluster Management Guide.

http://docs.aws.amazon.com/redshift/latest/APIReference/API_RestoreFromClusterSnapshot.html

Request

data RestoreFromClusterSnapshot

Request constructor

restoreFromClusterSnapshot

Arguments

:: Text
rfcsClusterIdentifier

-> Text
rfcsSnapshotIdentifier

-> RestoreFromClusterSnapshot

RestoreFromClusterSnapshot constructor.

The fields accessible through corresponding lenses are:

rfcsAllowVersionUpgrade :: Maybe Bool

rfcsAutomatedSnapshotRetentionPeriod :: Maybe Int

rfcsAvailabilityZone :: Maybe Text

rfcsClusterIdentifier :: Text

rfcsClusterParameterGroupName :: Maybe Text

rfcsClusterSecurityGroups :: [Text]

rfcsClusterSubnetGroupName :: Maybe Text

rfcsElasticIp :: Maybe Text

rfcsHsmClientCertificateIdentifier :: Maybe Text

rfcsHsmConfigurationIdentifier :: Maybe Text

rfcsKmsKeyId :: Maybe Text

rfcsOwnerAccount :: Maybe Text

rfcsPort :: Maybe Int

rfcsPreferredMaintenanceWindow :: Maybe Text

rfcsPubliclyAccessible :: Maybe Bool

rfcsSnapshotClusterIdentifier :: Maybe Text

rfcsSnapshotIdentifier :: Text

rfcsVpcSecurityGroupIds :: [Text]

Request lenses

rfcsAllowVersionUpgrade :: Lens' RestoreFromClusterSnapshot (Maybe Bool)

If true, major version upgrades can be applied during the maintenance window to the Amazon Redshift engine that is running on the cluster.

Default: true

rfcsAutomatedSnapshotRetentionPeriod :: Lens' RestoreFromClusterSnapshot (Maybe Int)

The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot .

Default: The value selected for the cluster from which the snapshot was taken.

Constraints: Must be a value from 0 to 35.

rfcsAvailabilityZone :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The Amazon EC2 Availability Zone in which to restore the cluster.

Default: A random, system-chosen Availability Zone.

Example: 'us-east-1a'

rfcsClusterIdentifier :: Lens' RestoreFromClusterSnapshot Text

The identifier of the cluster that will be created from restoring the snapshot.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens. Alphabetic characters must be lowercase. First character must be a letter. Cannot end with a hyphen or contain two consecutive hyphens. Must be unique for all clusters within an AWS account.

rfcsClusterParameterGroupName :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The name of the parameter group to be associated with this cluster.

Default: The default Amazon Redshift cluster parameter group. For information about the default parameter group, go to Working with AmazonRedshift Parameter Groups.

Constraints:

Must be 1 to 255 alphanumeric characters or hyphens. First character must be a letter. Cannot end with a hyphen or contain two consecutive hyphens.

rfcsClusterSecurityGroups :: Lens' RestoreFromClusterSnapshot [Text]

A list of security groups to be associated with this cluster.

Default: The default cluster security group for Amazon Redshift.

Cluster security groups only apply to clusters outside of VPCs.

rfcsClusterSubnetGroupName :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The name of the subnet group where you want to cluster restored.

A snapshot of cluster in VPC can be restored only in VPC. Therefore, you must provide subnet group name where you want the cluster restored.

rfcsElasticIp :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The elastic IP (EIP) address for the cluster.

rfcsHsmClientCertificateIdentifier :: Lens' RestoreFromClusterSnapshot (Maybe Text)

Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

rfcsHsmConfigurationIdentifier :: Lens' RestoreFromClusterSnapshot (Maybe Text)

Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

rfcsKmsKeyId :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The AWS Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster that you restore from a shared snapshot.

rfcsOwnerAccount :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The AWS customer account used to create or copy the snapshot. Required if you are restoring a snapshot you do not own, optional if you own the snapshot.

rfcsPort :: Lens' RestoreFromClusterSnapshot (Maybe Int)

The port number on which the cluster accepts connections.

Default: The same port as the original cluster.

Constraints: Must be between '1115' and '65535'.

rfcsPreferredMaintenanceWindow :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The weekly time range (in UTC) during which automated cluster maintenance can occur.

Format: 'ddd:hh24:mi-ddd:hh24:mi'

Default: The value selected for the cluster from which the snapshot was taken. For more information about the time blocks for each region, see Maintenance Windows in Amazon Redshift Cluster Management Guide.

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Minimum 30-minute window.

rfcsPubliclyAccessible :: Lens' RestoreFromClusterSnapshot (Maybe Bool)

If true, the cluster can be accessed from a public network.

rfcsSnapshotClusterIdentifier :: Lens' RestoreFromClusterSnapshot (Maybe Text)

The name of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

rfcsSnapshotIdentifier :: Lens' RestoreFromClusterSnapshot Text

The name of the snapshot from which to create the new cluster. This parameter isn't case sensitive.

Example: 'my-snapshot-id'

rfcsVpcSecurityGroupIds :: Lens' RestoreFromClusterSnapshot [Text]

A list of Virtual Private Cloud (VPC) security groups to be associated with the cluster.

Default: The default VPC security group is associated with the cluster.

VPC security groups only apply to clusters in VPCs.

Response

data RestoreFromClusterSnapshotResponse

Response constructor

restoreFromClusterSnapshotResponse :: RestoreFromClusterSnapshotResponse

RestoreFromClusterSnapshotResponse constructor.

The fields accessible through corresponding lenses are:

rfcsrCluster :: Maybe Cluster

Response lenses

rfcsrCluster :: Lens' RestoreFromClusterSnapshotResponse (Maybe Cluster)