Network.AWS.Redshift.RevokeClusterSecurityGroupIngress

Description

Revokes an ingress rule in an Amazon Redshift security group for a previously authorized IP range or Amazon EC2 security group. To add an ingress rule, see AuthorizeClusterSecurityGroupIngress . For information about managing security groups, go to Amazon Redshift Cluster Security Groups in the Amazon Redshift Cluster Management Guide .

Creating a Request

revokeClusterSecurityGroupIngress

Arguments

:: Text
rcsgiClusterSecurityGroupName

-> RevokeClusterSecurityGroupIngress

Creates a value of RevokeClusterSecurityGroupIngress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcsgiEC2SecurityGroupOwnerId - The AWS account number of the owner of the security group specified in the EC2SecurityGroupName parameter. The AWS access key ID is not an acceptable value. If EC2SecurityGroupOwnerId is specified, EC2SecurityGroupName must also be provided. and CIDRIP cannot be provided. Example: 111122223333

rcsgiEC2SecurityGroupName - The name of the EC2 Security Group whose access is to be revoked. If EC2SecurityGroupName is specified, EC2SecurityGroupOwnerId must also be provided and CIDRIP cannot be provided.

rcsgiCIdRIP - The IP range for which to revoke access. This range must be a valid Classless Inter-Domain Routing (CIDR) block of IP addresses. If CIDRIP is specified, EC2SecurityGroupName and EC2SecurityGroupOwnerId cannot be provided.

rcsgiClusterSecurityGroupName - The name of the security Group from which to revoke the ingress rule.

data RevokeClusterSecurityGroupIngress

See: revokeClusterSecurityGroupIngress smart constructor.

Request Lenses

rcsgiEC2SecurityGroupOwnerId :: Lens' RevokeClusterSecurityGroupIngress (Maybe Text)

The AWS account number of the owner of the security group specified in the EC2SecurityGroupName parameter. The AWS access key ID is not an acceptable value. If EC2SecurityGroupOwnerId is specified, EC2SecurityGroupName must also be provided. and CIDRIP cannot be provided. Example: 111122223333

rcsgiEC2SecurityGroupName :: Lens' RevokeClusterSecurityGroupIngress (Maybe Text)

The name of the EC2 Security Group whose access is to be revoked. If EC2SecurityGroupName is specified, EC2SecurityGroupOwnerId must also be provided and CIDRIP cannot be provided.

rcsgiCIdRIP :: Lens' RevokeClusterSecurityGroupIngress (Maybe Text)

The IP range for which to revoke access. This range must be a valid Classless Inter-Domain Routing (CIDR) block of IP addresses. If CIDRIP is specified, EC2SecurityGroupName and EC2SecurityGroupOwnerId cannot be provided.

rcsgiClusterSecurityGroupName :: Lens' RevokeClusterSecurityGroupIngress Text

The name of the security Group from which to revoke the ingress rule.

Destructuring the Response

revokeClusterSecurityGroupIngressResponse

Arguments

:: Int
rcsgirsResponseStatus

-> RevokeClusterSecurityGroupIngressResponse

Creates a value of RevokeClusterSecurityGroupIngressResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcsgirsClusterSecurityGroup - Undocumented member.

rcsgirsResponseStatus - -- | The response status code.

data RevokeClusterSecurityGroupIngressResponse

See: revokeClusterSecurityGroupIngressResponse smart constructor.

Response Lenses

rcsgirsClusterSecurityGroup :: Lens' RevokeClusterSecurityGroupIngressResponse (Maybe ClusterSecurityGroup)

Undocumented member.

rcsgirsResponseStatus :: Lens' RevokeClusterSecurityGroupIngressResponse Int

- | The response status code.