Network.AWS.Redshift.V2012_12_01.DeleteClusterParameterGroup

Description

Deletes a specified Amazon Redshift parameter group. You cannot delete a parameter group if it is associated with a cluster. https:/redshift.us-east-1.amazonaws.com ?Action=DeleteClusterParameterGroup &ParameterGroupName=parametergroup1 &Version=2012-12-01 &x-amz-algorithm=AWS4-HMAC-SHA256 &x-amz-credential=AKIAIOSFODNN7EXAMPLE20121208us-east-1redshiftaws4_request &x-amz-date=20121208T015410Z &x-amz-signedheaders=content-type;host;x-amz-date 29674ca0-40da-11e2-b679-dba6cf515770.

Request

data DeleteClusterParameterGroup

Request constructor

mkDeleteClusterParameterGroup

Arguments

:: Text
dcpgParameterGroupName

-> DeleteClusterParameterGroup

Smart constructor for the minimum required parameters to construct a valid DeleteClusterParameterGroup request.

Request lenses

dcpgParameterGroupName :: Lens' DeleteClusterParameterGroup Text

The name of the parameter group to be deleted. Constraints: Must be the name of an existing cluster parameter group. Cannot delete a default cluster parameter group.

Response

data DeleteClusterParameterGroupResponse