Network.AWS.Redshift.V2012_12_01.DescribeLoggingStatus

Description

Describes whether information, such as queries and connection attempts, is being logged for the specified Amazon Redshift cluster.

Request

data DescribeLoggingStatus

Request constructor

mkDescribeLoggingStatus

Arguments

:: Text
dlsClusterIdentifier

-> DescribeLoggingStatus

Smart constructor for the minimum required parameters to construct a valid DescribeLoggingStatus request.

Request lenses

dlsClusterIdentifier :: Lens' DescribeLoggingStatus Text

The identifier of the cluster to get the logging status from. Example: examplecluster.

Response

data DescribeLoggingStatusResponse

Describes the status of logging for a cluster.

Response lenses

dlsrsLoggingEnabled :: Lens' DescribeLoggingStatusResponse (Maybe Bool)

true if logging is on, false if logging is off.

dlsrsBucketName :: Lens' DescribeLoggingStatusResponse (Maybe Text)

The name of the S3 bucket where the log files are stored.

dlsrsS3KeyPrefix :: Lens' DescribeLoggingStatusResponse (Maybe Text)

The prefix applied to the log file names.

dlsrsLastSuccessfulDeliveryTime :: Lens' DescribeLoggingStatusResponse (Maybe ISO8601)

The last time when logs were delivered.

dlsrsLastFailureTime :: Lens' DescribeLoggingStatusResponse (Maybe ISO8601)

The last time when logs failed to be delivered.

dlsrsLastFailureMessage :: Lens' DescribeLoggingStatusResponse (Maybe Text)

The message indicating that logs failed to be delivered.