Network.AWS.Redshift.V2012_12_01.DisableLogging

Description

Stops logging information, such as queries and connection attempts, for the specified Amazon Redshift cluster.

Request

data DisableLogging

Request constructor

mkDisableLogging

Arguments

:: Text
dlClusterIdentifier

-> DisableLogging

Smart constructor for the minimum required parameters to construct a valid DisableLogging request.

Request lenses

dlClusterIdentifier :: Lens' DisableLogging Text

The identifier of the cluster on which logging is to be stopped. Example: examplecluster.

Response

data DisableLoggingResponse

Describes the status of logging for a cluster.

Response lenses

dlrLoggingEnabled :: Lens' DisableLoggingResponse (Maybe Bool)

true if logging is on, false if logging is off.

dlrBucketName :: Lens' DisableLoggingResponse (Maybe Text)

The name of the S3 bucket where the log files are stored.

dlrS3KeyPrefix :: Lens' DisableLoggingResponse (Maybe Text)

The prefix applied to the log file names.

dlrLastSuccessfulDeliveryTime :: Lens' DisableLoggingResponse (Maybe ISO8601)

The last time when logs were delivered.

dlrLastFailureTime :: Lens' DisableLoggingResponse (Maybe ISO8601)

The last time when logs failed to be delivered.

dlrLastFailureMessage :: Lens' DisableLoggingResponse (Maybe Text)

The message indicating that logs failed to be delivered.