Network.AWS.Redshift.V2012_12_01.EnableSnapshotCopy

Description

Enables the automatic copy of snapshots from one region to another region for a specified cluster.

Request

data EnableSnapshotCopy

Request constructor

mkEnableSnapshotCopy

Arguments

:: Text
escClusterIdentifier

-> Text
escDestinationRegion

-> EnableSnapshotCopy

Smart constructor for the minimum required parameters to construct a valid EnableSnapshotCopy request.

Request lenses

escClusterIdentifier :: Lens' EnableSnapshotCopy Text

The unique identifier of the source cluster to copy snapshots from. Constraints: Must be the valid name of an existing cluster that does not already have cross-region snapshot copy enabled.

escDestinationRegion :: Lens' EnableSnapshotCopy Text

The destination region that you want to copy snapshots to. Constraints: Must be the name of a valid region. For more information, see Regions and Endpoints in the Amazon Web Services General Reference.

escRetentionPeriod :: Lens' EnableSnapshotCopy (Maybe Integer)

The number of days to retain automated snapshots in the destination region after they are copied from the source region. Default: 7. Constraints: Must be at least 1 and no more than 35.

Response

data EnableSnapshotCopyResponse

Response lenses

escrsCluster :: Lens' EnableSnapshotCopyResponse (Maybe Cluster)

Describes a cluster.