Network.AWS.Redshift.V2012_12_01.Types

Description

Amazon Redshift is a fast, fully managed, petabyte-scale data warehouse service that makes it simple and cost-effective to efficiently analyze all your data using your existing business intelligence tools. You can start small for just $0.25 per hour with no commitments or upfront costs and scale to a petabyte or more for $1,000 per terabyte per year, less than a tenth of most other data warehousing solutions.

Service

data Redshift

Supported version (2012-12-01) of the Amazon Redshift service.

XML

xmlOptions :: Tagged a XMLOptions

SourceType

data SourceType

Constructors

SourceTypeCluster
cluster

SourceTypeClusterParameterGroup
cluster-parameter-group

SourceTypeClusterSecurityGroup
cluster-security-group

SourceTypeClusterSnapshot
cluster-snapshot

AccountWithRestoreAccess

data AccountWithRestoreAccess

Describes an AWS customer account authorized to restore a snapshot.

mkAccountWithRestoreAccess :: AccountWithRestoreAccess

Smart constructor for the minimum required fields to construct a valid AccountWithRestoreAccess data type to populate a request.

awraAccountId :: Lens' AccountWithRestoreAccess (Maybe Text)

The identifier of an AWS customer account authorized to restore a snapshot.

AvailabilityZone

data AvailabilityZone

Describes an availability zone.

mkAvailabilityZone :: AvailabilityZone

Smart constructor for the minimum required fields to construct a valid AvailabilityZone data type to populate a request.

azName :: Lens' AvailabilityZone (Maybe Text)

The name of the availability zone.

Cluster

data Cluster

Describes a cluster.

mkCluster :: Cluster

Smart constructor for the minimum required fields to construct a valid Cluster data type to populate a request.

cClusterIdentifier :: Lens' Cluster (Maybe Text)

The unique identifier of the cluster.

cNodeType :: Lens' Cluster (Maybe Text)

The node type for the nodes in the cluster.

cClusterStatus :: Lens' Cluster (Maybe Text)

The current state of this cluster. Possible values include available, creating, deleting, rebooting, renaming, and resizing.

cModifyStatus :: Lens' Cluster (Maybe Text)

The status of a modify operation, if any, initiated for the cluster.

cMasterUsername :: Lens' Cluster (Maybe Text)

The master user name for the cluster. This name is used to connect to the database that is specified in DBName.

cDBName :: Lens' Cluster (Maybe Text)

The name of the initial database that was created when the cluster was created. This same name is returned for the life of the cluster. If an initial database was not specified, a database named "dev" was created by default.

cEndpoint :: Lens' Cluster (Maybe Endpoint)

The connection endpoint.

cClusterCreateTime :: Lens' Cluster (Maybe ISO8601)

The date and time that the cluster was created.

cAutomatedSnapshotRetentionPeriod :: Lens' Cluster (Maybe Integer)

The number of days that automatic cluster snapshots are retained.

cClusterSecurityGroups :: Lens' Cluster [ClusterSecurityGroupMembership]

A list of cluster security group that are associated with the cluster. Each security group is represented by an element that contains ClusterSecurityGroup.Name and ClusterSecurityGroup.Status subelements. Cluster security groups are used when the cluster is not created in a VPC. Clusters that are created in a VPC use VPC security groups, which are listed by the VpcSecurityGroups parameter.

cVpcSecurityGroups :: Lens' Cluster [VpcSecurityGroupMembership]

A list of Virtual Private Cloud (VPC) security groups that are associated with the cluster. This parameter is returned only if the cluster is in a VPC.

cClusterParameterGroups :: Lens' Cluster [ClusterParameterGroupStatus]

The list of cluster parameter groups that are associated with this cluster.

cClusterSubnetGroupName :: Lens' Cluster (Maybe Text)

The name of the subnet group that is associated with the cluster. This parameter is valid only when the cluster is in a VPC.

cVpcId :: Lens' Cluster (Maybe Text)

The identifier of the VPC the cluster is in, if the cluster is in a VPC.

cAvailabilityZone :: Lens' Cluster (Maybe Text)

The name of the Availability Zone in which the cluster is located.

cPreferredMaintenanceWindow :: Lens' Cluster (Maybe Text)

The weekly time range (in UTC) during which system maintenance can occur.

cPendingModifiedValues :: Lens' Cluster (Maybe PendingModifiedValues)

If present, changes to the cluster are pending. Specific pending changes are identified by subelements.

cClusterVersion :: Lens' Cluster (Maybe Text)

The version ID of the Amazon Redshift engine that is running on the cluster.

cAllowVersionUpgrade :: Lens' Cluster (Maybe Bool)

If true, version upgrades will be applied automatically to the cluster during the maintenance window.

cNumberOfNodes :: Lens' Cluster (Maybe Integer)

The number of compute nodes in the cluster.

cPubliclyAccessible :: Lens' Cluster (Maybe Bool)

If true, the cluster can be accessed from a public network.

cEncrypted :: Lens' Cluster (Maybe Bool)

If true, data in the cluster is encrypted at rest.

cRestoreStatus :: Lens' Cluster (Maybe RestoreStatus)

Describes the status of a cluster restore action. Returns null if the cluster was not created by restoring a snapshot.

cHsmStatus :: Lens' Cluster (Maybe HsmStatus)

Reports whether the Amazon Redshift cluster has finished applying any HSM settings changes specified in a modify cluster command. Values: active, applying.

cClusterSnapshotCopyStatus :: Lens' Cluster (Maybe ClusterSnapshotCopyStatus)

Returns the destination region and retention period that are configured for cross-region snapshot copy.

cClusterPublicKey :: Lens' Cluster (Maybe Text)

The public key for the cluster.

cClusterNodes :: Lens' Cluster [ClusterNode]

The nodes in a cluster.

cElasticIpStatus :: Lens' Cluster (Maybe ElasticIpStatus)

Describes the status of the elastic IP (EIP) address.

cClusterRevisionNumber :: Lens' Cluster (Maybe Text)

The specific revision number of the database in the cluster.

ClusterNode

data ClusterNode

The identifier of a node in a cluster. --&gt;.

mkClusterNode :: ClusterNode

Smart constructor for the minimum required fields to construct a valid ClusterNode data type to populate a request.

cnNodeRole :: Lens' ClusterNode (Maybe Text)

Whether the node is a leader node or a compute node.

cnPrivateIPAddress :: Lens' ClusterNode (Maybe Text)

The private IP address of a node within a cluster.

cnPublicIPAddress :: Lens' ClusterNode (Maybe Text)

The public IP address of a node within a cluster.

ClusterParameterGroup

data ClusterParameterGroup

Describes a parameter group.

mkClusterParameterGroup :: ClusterParameterGroup

Smart constructor for the minimum required fields to construct a valid ClusterParameterGroup data type to populate a request.

cpgParameterGroupName :: Lens' ClusterParameterGroup (Maybe Text)

The name of the cluster parameter group.

cpgParameterGroupFamily :: Lens' ClusterParameterGroup (Maybe Text)

The name of the cluster parameter group family that this cluster parameter group is compatible with.

cpgDescription :: Lens' ClusterParameterGroup (Maybe Text)

The description of the parameter group.

ClusterParameterGroupStatus

data ClusterParameterGroupStatus

Describes the status of a parameter group.

mkClusterParameterGroupStatus :: ClusterParameterGroupStatus

Smart constructor for the minimum required fields to construct a valid ClusterParameterGroupStatus data type to populate a request.

cpgsParameterGroupName :: Lens' ClusterParameterGroupStatus (Maybe Text)

The name of the cluster parameter group.

cpgsParameterApplyStatus :: Lens' ClusterParameterGroupStatus (Maybe Text)

The status of parameter updates.

ClusterSecurityGroup

data ClusterSecurityGroup

Describes a security group.

mkClusterSecurityGroup :: ClusterSecurityGroup

Smart constructor for the minimum required fields to construct a valid ClusterSecurityGroup data type to populate a request.

csgClusterSecurityGroupName :: Lens' ClusterSecurityGroup (Maybe Text)

The name of the cluster security group to which the operation was applied.

csgDescription :: Lens' ClusterSecurityGroup (Maybe Text)

A description of the security group.

csgEC2SecurityGroups :: Lens' ClusterSecurityGroup [EC2SecurityGroup]

A list of EC2 security groups that are permitted to access clusters associated with this cluster security group.

csgIPRanges :: Lens' ClusterSecurityGroup [IPRange]

A list of IP ranges (CIDR blocks) that are permitted to access clusters associated with this cluster security group.

ClusterSecurityGroupMembership

data ClusterSecurityGroupMembership

Describes a security group.

mkClusterSecurityGroupMembership :: ClusterSecurityGroupMembership

Smart constructor for the minimum required fields to construct a valid ClusterSecurityGroupMembership data type to populate a request.

csgmClusterSecurityGroupName :: Lens' ClusterSecurityGroupMembership (Maybe Text)

The name of the cluster security group.

csgmStatus :: Lens' ClusterSecurityGroupMembership (Maybe Text)

The status of the cluster security group.

ClusterSnapshotCopyStatus

data ClusterSnapshotCopyStatus

Returns the destination region and retention period that are configured for cross-region snapshot copy.

mkClusterSnapshotCopyStatus :: ClusterSnapshotCopyStatus

Smart constructor for the minimum required fields to construct a valid ClusterSnapshotCopyStatus data type to populate a request.

cscsDestinationRegion :: Lens' ClusterSnapshotCopyStatus (Maybe Text)

The destination region that snapshots are automatically copied to when cross-region snapshot copy is enabled.

cscsRetentionPeriod :: Lens' ClusterSnapshotCopyStatus (Maybe Integer)

The number of days that automated snapshots are retained in the destination region after they are copied from a source region.

ClusterSubnetGroup

data ClusterSubnetGroup

Describes a subnet group.

mkClusterSubnetGroup :: ClusterSubnetGroup

Smart constructor for the minimum required fields to construct a valid ClusterSubnetGroup data type to populate a request.

csgrClusterSubnetGroupName :: Lens' ClusterSubnetGroup (Maybe Text)

The name of the cluster subnet group.

csgrDescription :: Lens' ClusterSubnetGroup (Maybe Text)

The description of the cluster subnet group.

csgrVpcId :: Lens' ClusterSubnetGroup (Maybe Text)

The VPC ID of the cluster subnet group.

csgrSubnetGroupStatus :: Lens' ClusterSubnetGroup (Maybe Text)

The status of the cluster subnet group. Possible values are Complete, Incomplete and Invalid.

csgrSubnets :: Lens' ClusterSubnetGroup [Subnet]

A list of the VPC Subnet elements.

ClusterVersion

data ClusterVersion

Describes a cluster version, including the parameter group family and description of the version.

mkClusterVersion :: ClusterVersion

Smart constructor for the minimum required fields to construct a valid ClusterVersion data type to populate a request.

cvClusterVersion :: Lens' ClusterVersion (Maybe Text)

The version number used by the cluster.

cvClusterParameterGroupFamily :: Lens' ClusterVersion (Maybe Text)

The name of the cluster parameter group family for the cluster.

cvDescription :: Lens' ClusterVersion (Maybe Text)

The description of the cluster version.

DefaultClusterParameters

data DefaultClusterParameters

Describes the default cluster parameters for a parameter group family.

mkDefaultClusterParameters :: DefaultClusterParameters

Smart constructor for the minimum required fields to construct a valid DefaultClusterParameters data type to populate a request.

dcp1ParameterGroupFamily :: Lens' DefaultClusterParameters (Maybe Text)

The name of the cluster parameter group family to which the engine default parameters apply.

dcp1Marker :: Lens' DefaultClusterParameters (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

dcp1Parameters :: Lens' DefaultClusterParameters [Parameter]

The list of cluster default parameters.

EC2SecurityGroup

data EC2SecurityGroup

Describes an Amazon EC2 security group.

mkEC2SecurityGroup :: EC2SecurityGroup

Smart constructor for the minimum required fields to construct a valid EC2SecurityGroup data type to populate a request.

ecsgStatus :: Lens' EC2SecurityGroup (Maybe Text)

The status of the EC2 security group.

ecsgEC2SecurityGroupName :: Lens' EC2SecurityGroup (Maybe Text)

The name of the EC2 Security Group.

ecsgEC2SecurityGroupOwnerId :: Lens' EC2SecurityGroup (Maybe Text)

The AWS ID of the owner of the EC2 security group specified in the EC2SecurityGroupName field.

ElasticIpStatus

data ElasticIpStatus

Describes the status of the elastic IP (EIP) address.

mkElasticIpStatus :: ElasticIpStatus

Smart constructor for the minimum required fields to construct a valid ElasticIpStatus data type to populate a request.

eisElasticIp :: Lens' ElasticIpStatus (Maybe Text)

The elastic IP (EIP) address for the cluster.

eisStatus :: Lens' ElasticIpStatus (Maybe Text)

Describes the status of the elastic IP (EIP) address.

Endpoint

data Endpoint

The connection endpoint.

mkEndpoint :: Endpoint

Smart constructor for the minimum required fields to construct a valid Endpoint data type to populate a request.

eAddress :: Lens' Endpoint (Maybe Text)

The DNS address of the Cluster.

ePort :: Lens' Endpoint (Maybe Integer)

The port that the database engine is listening on.

Event

data Event

Describes an event.

mkEvent :: Event

Smart constructor for the minimum required fields to construct a valid Event data type to populate a request.

erSourceIdentifier :: Lens' Event (Maybe Text)

The identifier for the source of the event.

erSourceType :: Lens' Event (Maybe SourceType)

The source type for this event.

erMessage :: Lens' Event (Maybe Text)

The text of this event.

erEventCategories :: Lens' Event [Text]

A list of the event categories.

erSeverity :: Lens' Event (Maybe Text)

The severity of the event. Values: ERROR, INFO.

erDate :: Lens' Event (Maybe ISO8601)

The date and time of the event.

erEventId :: Lens' Event (Maybe Text)

The identifier of the event.

EventCategoriesMap

data EventCategoriesMap

mkEventCategoriesMap :: EventCategoriesMap

Smart constructor for the minimum required fields to construct a valid EventCategoriesMap data type to populate a request.

ecmSourceType :: Lens' EventCategoriesMap (Maybe Text)

The Amazon Redshift source type, such as cluster or cluster-snapshot, that the returned categories belong to.

ecmEvents :: Lens' EventCategoriesMap [EventInfoMap]

The events in the event category.

EventInfoMap

data EventInfoMap

mkEventInfoMap :: EventInfoMap

Smart constructor for the minimum required fields to construct a valid EventInfoMap data type to populate a request.

eimEventId :: Lens' EventInfoMap (Maybe Text)

The identifier of an Amazon Redshift event.

eimEventCategories :: Lens' EventInfoMap [Text]

The category of an Amazon Redshift event.

eimEventDescription :: Lens' EventInfoMap (Maybe Text)

The description of an Amazon Redshift event.

eimSeverity :: Lens' EventInfoMap (Maybe Text)

The severity of the event. Values: ERROR, INFO.

EventSubscription

data EventSubscription

mkEventSubscription :: EventSubscription

Smart constructor for the minimum required fields to construct a valid EventSubscription data type to populate a request.

esCustomerAwsId :: Lens' EventSubscription (Maybe Text)

The AWS customer account associated with the Amazon Redshift event notification subscription.

esCustSubscriptionId :: Lens' EventSubscription (Maybe Text)

The name of the Amazon Redshift event notification subscription.

esSnsTopicArn :: Lens' EventSubscription (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic used by the event notification subscription.

esStatus :: Lens' EventSubscription (Maybe Text)

The status of the Amazon Redshift event notification subscription. Constraints: Can be one of the following: active | no-permission | topic-not-exist The status "no-permission" indicates that Amazon Redshift no longer has permission to post to the Amazon SNS topic. The status "topic-not-exist" indicates that the topic was deleted after the subscription was created.

esSubscriptionCreationTime :: Lens' EventSubscription (Maybe ISO8601)

The date and time the Amazon Redshift event notification subscription was created.

esSourceType :: Lens' EventSubscription (Maybe Text)

The source type of the events returned the Amazon Redshift event notification, such as cluster, or cluster-snapshot.

esSourceIdsList :: Lens' EventSubscription [Text]

A list of the sources that publish events to the Amazon Redshift event notification subscription.

esEventCategoriesList :: Lens' EventSubscription [Text]

The list of Amazon Redshift event categories specified in the event notification subscription. Values: Configuration, Management, Monitoring, Security.

esSeverity :: Lens' EventSubscription (Maybe Text)

The event severity specified in the Amazon Redshift event notification subscription. Values: ERROR, INFO.

esEnabled :: Lens' EventSubscription (Maybe Bool)

A Boolean value indicating whether the subscription is enabled. true indicates the subscription is enabled.

HsmClientCertificate

data HsmClientCertificate

Returns information about an HSM client certificate. The certificate is stored in a secure Hardware Storage Module (HSM), and used by the Amazon Redshift cluster to encrypt data files.

mkHsmClientCertificate :: HsmClientCertificate

Smart constructor for the minimum required fields to construct a valid HsmClientCertificate data type to populate a request.

hccHsmClientCertificateIdentifier :: Lens' HsmClientCertificate (Maybe Text)

The identifier of the HSM client certificate.

hccHsmClientCertificatePublicKey :: Lens' HsmClientCertificate (Maybe Text)

The public key that the Amazon Redshift cluster will use to connect to the HSM. You must register the public key in the HSM.

HsmConfiguration

data HsmConfiguration

Returns information about an HSM configuration, which is an object that describes to Amazon Redshift clusters the information they require to connect to an HSM where they can store database encryption keys.

mkHsmConfiguration :: HsmConfiguration

Smart constructor for the minimum required fields to construct a valid HsmConfiguration data type to populate a request.

hcHsmConfigurationIdentifier :: Lens' HsmConfiguration (Maybe Text)

The name of the Amazon Redshift HSM configuration.

hcDescription :: Lens' HsmConfiguration (Maybe Text)

A text description of the HSM configuration.

hcHsmIpAddress :: Lens' HsmConfiguration (Maybe Text)

The IP address that the Amazon Redshift cluster must use to access the HSM.

hcHsmPartitionName :: Lens' HsmConfiguration (Maybe Text)

The name of the partition in the HSM where the Amazon Redshift clusters will store their database encryption keys.

HsmStatus

data HsmStatus

Reports whether the Amazon Redshift cluster has finished applying any HSM settings changes specified in a modify cluster command. Values: active, applying.

mkHsmStatus :: HsmStatus

Smart constructor for the minimum required fields to construct a valid HsmStatus data type to populate a request.

hsHsmClientCertificateIdentifier :: Lens' HsmStatus (Maybe Text)

Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

hsHsmConfigurationIdentifier :: Lens' HsmStatus (Maybe Text)

Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

hsStatus :: Lens' HsmStatus (Maybe Text)

Reports whether the Amazon Redshift cluster has finished applying any HSM settings changes specified in a modify cluster command. Values: active, applying.

IPRange

data IPRange

Describes an IP range used in a security group.

mkIPRange :: IPRange

Smart constructor for the minimum required fields to construct a valid IPRange data type to populate a request.

iprStatus :: Lens' IPRange (Maybe Text)

The status of the IP range, for example, "authorized".

iprCIDRIP :: Lens' IPRange (Maybe Text)

The IP range in Classless Inter-Domain Routing (CIDR) notation.

OrderableClusterOption

data OrderableClusterOption

Describes an orderable cluster option.

mkOrderableClusterOption :: OrderableClusterOption

Smart constructor for the minimum required fields to construct a valid OrderableClusterOption data type to populate a request.

ocoClusterVersion :: Lens' OrderableClusterOption (Maybe Text)

The version of the orderable cluster.

ocoClusterType :: Lens' OrderableClusterOption (Maybe Text)

The cluster type, for example multi-node.

ocoNodeType :: Lens' OrderableClusterOption (Maybe Text)

The node type for the orderable cluster.

ocoAvailabilityZones :: Lens' OrderableClusterOption [AvailabilityZone]

A list of availability zones for the orderable cluster.

Parameter

data Parameter

Describes a parameter in a cluster parameter group.

mkParameter :: Parameter

Smart constructor for the minimum required fields to construct a valid Parameter data type to populate a request.

pParameterName :: Lens' Parameter (Maybe Text)

The name of the parameter.

pParameterValue :: Lens' Parameter (Maybe Text)

The value of the parameter.

pDescription :: Lens' Parameter (Maybe Text)

A description of the parameter.

pSource :: Lens' Parameter (Maybe Text)

The source of the parameter value, such as "engine-default" or "user".

pDataType :: Lens' Parameter (Maybe Text)

The data type of the parameter.

pAllowedValues :: Lens' Parameter (Maybe Text)

The valid range of values for the parameter.

pIsModifiable :: Lens' Parameter (Maybe Bool)

If true, the parameter can be modified. Some parameters have security or operational implications that prevent them from being changed.

pMinimumEngineVersion :: Lens' Parameter (Maybe Text)

The earliest engine version to which the parameter can apply.

PendingModifiedValues

data PendingModifiedValues

If present, changes to the cluster are pending. Specific pending changes are identified by subelements.

mkPendingModifiedValues :: PendingModifiedValues

Smart constructor for the minimum required fields to construct a valid PendingModifiedValues data type to populate a request.

pmvMasterUserPassword :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the master user password for the cluster.

pmvNodeType :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the cluster's node type.

pmvNumberOfNodes :: Lens' PendingModifiedValues (Maybe Integer)

The pending or in-progress change of the number of nodes in the cluster.

pmvClusterType :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the cluster type.

pmvClusterVersion :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the service version.

pmvAutomatedSnapshotRetentionPeriod :: Lens' PendingModifiedValues (Maybe Integer)

The pending or in-progress change of the automated snapshot retention period.

pmvClusterIdentifier :: Lens' PendingModifiedValues (Maybe Text)

The pending or in-progress change of the new identifier for the cluster.

RecurringCharge

data RecurringCharge

Describes a recurring charge.

mkRecurringCharge :: RecurringCharge

Smart constructor for the minimum required fields to construct a valid RecurringCharge data type to populate a request.

rcRecurringChargeAmount :: Lens' RecurringCharge (Maybe Double)

The amount charged per the period of time specified by the recurring charge frequency.

rcRecurringChargeFrequency :: Lens' RecurringCharge (Maybe Text)

The frequency at which the recurring charge amount is applied.

ReservedNode

data ReservedNode

Describes a reserved node.

mkReservedNode :: ReservedNode

Smart constructor for the minimum required fields to construct a valid ReservedNode data type to populate a request.

rnReservedNodeId :: Lens' ReservedNode (Maybe Text)

The unique identifier for the reservation.

rnReservedNodeOfferingId :: Lens' ReservedNode (Maybe Text)

The identifier for the reserved node offering.

rnNodeType :: Lens' ReservedNode (Maybe Text)

The node type of the reserved node.

rnStartTime :: Lens' ReservedNode (Maybe ISO8601)

The time the reservation started. You purchase a reserved node offering for a duration. This is the start time of that duration.

rnDuration :: Lens' ReservedNode (Maybe Integer)

The duration of the node reservation in seconds.

rnFixedPrice :: Lens' ReservedNode (Maybe Double)

The fixed cost Amazon Redshift charged you for this reserved node.

rnUsagePrice :: Lens' ReservedNode (Maybe Double)

The hourly rate Amazon Redshift charge you for this reserved node.

rnCurrencyCode :: Lens' ReservedNode (Maybe Text)

The currency code for the reserved cluster.

rnNodeCount :: Lens' ReservedNode (Maybe Integer)

The number of reserved compute nodes.

rnState :: Lens' ReservedNode (Maybe Text)

The state of the reserved Compute Node. Possible Values: pending-payment-This reserved node has recently been purchased, and the sale has been approved, but payment has not yet been confirmed. active-This reserved node is owned by the caller and is available for use. payment-failed-Payment failed for the purchase attempt.

rnOfferingType :: Lens' ReservedNode (Maybe Text)

The anticipated utilization of the reserved node, as defined in the reserved node offering.

rnRecurringCharges :: Lens' ReservedNode [RecurringCharge]

The recurring charges for the reserved node.

ReservedNodeOffering

data ReservedNodeOffering

Describes a reserved node offering.

mkReservedNodeOffering :: ReservedNodeOffering

Smart constructor for the minimum required fields to construct a valid ReservedNodeOffering data type to populate a request.

rnoReservedNodeOfferingId :: Lens' ReservedNodeOffering (Maybe Text)

The offering identifier.

rnoNodeType :: Lens' ReservedNodeOffering (Maybe Text)

The node type offered by the reserved node offering.

rnoDuration :: Lens' ReservedNodeOffering (Maybe Integer)

The duration, in seconds, for which the offering will reserve the node.

rnoFixedPrice :: Lens' ReservedNodeOffering (Maybe Double)

The upfront fixed charge you will pay to purchase the specific reserved node offering.

rnoUsagePrice :: Lens' ReservedNodeOffering (Maybe Double)

The rate you are charged for each hour the cluster that is using the offering is running.

rnoCurrencyCode :: Lens' ReservedNodeOffering (Maybe Text)

The currency code for the compute nodes offering.

rnoOfferingType :: Lens' ReservedNodeOffering (Maybe Text)

The anticipated utilization of the reserved node, as defined in the reserved node offering.

rnoRecurringCharges :: Lens' ReservedNodeOffering [RecurringCharge]

The charge to your account regardless of whether you are creating any clusters using the node offering. Recurring charges are only in effect for heavy-utilization reserved nodes.

RestoreStatus

data RestoreStatus

Describes the status of a cluster restore action. Returns null if the cluster was not created by restoring a snapshot.

mkRestoreStatus :: RestoreStatus

Smart constructor for the minimum required fields to construct a valid RestoreStatus data type to populate a request.

rsStatus :: Lens' RestoreStatus (Maybe Text)

The status of the restore action. Returns starting, restoring, completed, or failed.

rsCurrentRestoreRateInMegaBytesPerSecond :: Lens' RestoreStatus (Maybe Double)

The number of megabytes per second being transferred from the backup storage. Returns the average rate for a completed backup.

rsSnapshotSizeInMegaBytes :: Lens' RestoreStatus (Maybe Integer)

The size of the set of snapshot data used to restore the cluster.

rsProgressInMegaBytes :: Lens' RestoreStatus (Maybe Integer)

The number of megabytes that have been transferred from snapshot storage.

rsElapsedTimeInSeconds :: Lens' RestoreStatus (Maybe Integer)

The amount of time an in-progress restore has been running, or the amount of time it took a completed restore to finish.

rsEstimatedTimeToCompletionInSeconds :: Lens' RestoreStatus (Maybe Integer)

The estimate of the time remaining before the restore will complete. Returns 0 for a completed restore.

Snapshot

data Snapshot

Describes a snapshot.

mkSnapshot :: Snapshot

Smart constructor for the minimum required fields to construct a valid Snapshot data type to populate a request.

sSnapshotIdentifier :: Lens' Snapshot (Maybe Text)

The snapshot identifier that is provided in the request.

sClusterIdentifier :: Lens' Snapshot (Maybe Text)

The identifier of the cluster for which the snapshot was taken.

sSnapshotCreateTime :: Lens' Snapshot (Maybe ISO8601)

The time (UTC) when Amazon Redshift began the snapshot. A snapshot contains a copy of the cluster data as of this exact time.

sStatus :: Lens' Snapshot (Maybe Text)

The snapshot status. The value of the status depends on the API operation used. CreateClusterSnapshot and CopyClusterSnapshot returns status as "creating". DescribeClusterSnapshots returns status as "creating", "available", "final snapshot", or "failed". DeleteClusterSnapshot returns status as "deleted".

sPort :: Lens' Snapshot (Maybe Integer)

The port that the cluster is listening on.

sAvailabilityZone :: Lens' Snapshot (Maybe Text)

The Availability Zone in which the cluster was created.

sClusterCreateTime :: Lens' Snapshot (Maybe ISO8601)

The time (UTC) when the cluster was originally created.

sMasterUsername :: Lens' Snapshot (Maybe Text)

The master user name for the cluster.

sClusterVersion :: Lens' Snapshot (Maybe Text)

The version ID of the Amazon Redshift engine that is running on the cluster.

sSnapshotType :: Lens' Snapshot (Maybe Text)

The snapshot type. Snapshots created using CreateClusterSnapshot and CopyClusterSnapshot will be of type "manual".

sNodeType :: Lens' Snapshot (Maybe Text)

The node type of the nodes in the cluster.

sNumberOfNodes :: Lens' Snapshot (Maybe Integer)

The number of nodes in the cluster.

sDBName :: Lens' Snapshot (Maybe Text)

The name of the database that was created when the cluster was created.

sVpcId :: Lens' Snapshot (Maybe Text)

The VPC identifier of the cluster if the snapshot is from a cluster in a VPC. Otherwise, this field is not in the output.

sEncrypted :: Lens' Snapshot (Maybe Bool)

If true, the data in the snapshot is encrypted at rest.

sEncryptedWithHSM :: Lens' Snapshot (Maybe Bool)

A boolean that indicates whether the snapshot data is encrypted using the HSM keys of the source cluster. true indicates that the data is encrypted using HSM keys.

sAccountsWithRestoreAccess :: Lens' Snapshot [AccountWithRestoreAccess]

A list of the AWS customer accounts authorized to restore the snapshot. Returns null if no accounts are authorized. Visible only to the snapshot owner.

sOwnerAccount :: Lens' Snapshot (Maybe Text)

For manual snapshots, the AWS customer account used to create or copy the snapshot. For automatic snapshots, the owner of the cluster. The owner can perform all snapshot actions, such as sharing a manual snapshot.

sTotalBackupSizeInMegaBytes :: Lens' Snapshot (Maybe Double)

The size of the complete set of backup data that would be used to restore the cluster.

sActualIncrementalBackupSizeInMegaBytes :: Lens' Snapshot (Maybe Double)

The size of the incremental backup.

sBackupProgressInMegaBytes :: Lens' Snapshot (Maybe Double)

The number of megabytes that have been transferred to the snapshot backup.

sCurrentBackupRateInMegaBytesPerSecond :: Lens' Snapshot (Maybe Double)

The number of megabytes per second being transferred to the snapshot backup. Returns 0 for a completed backup.

sEstimatedSecondsToCompletion :: Lens' Snapshot (Maybe Integer)

The estimate of the time remaining before the snapshot backup will complete. Returns 0 for a completed backup.

sElapsedTimeInSeconds :: Lens' Snapshot (Maybe Integer)

The amount of time an in-progress snapshot backup has been running, or the amount of time it took a completed backup to finish.

sSourceRegion :: Lens' Snapshot (Maybe Text)

The source region from which the snapshot was copied.

Subnet

data Subnet

Describes a subnet.

mkSubnet :: Subnet

Smart constructor for the minimum required fields to construct a valid Subnet data type to populate a request.

srSubnetIdentifier :: Lens' Subnet (Maybe Text)

The identifier of the subnet.

srSubnetAvailabilityZone :: Lens' Subnet (Maybe AvailabilityZone)

Describes an availability zone.

srSubnetStatus :: Lens' Subnet (Maybe Text)

The status of the subnet.

VpcSecurityGroupMembership

data VpcSecurityGroupMembership

Describes the members of a VPC security group.

mkVpcSecurityGroupMembership :: VpcSecurityGroupMembership

Smart constructor for the minimum required fields to construct a valid VpcSecurityGroupMembership data type to populate a request.

vsgmVpcSecurityGroupId :: Lens' VpcSecurityGroupMembership (Maybe Text)

vsgmStatus :: Lens' VpcSecurityGroupMembership (Maybe Text)